% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Not again a title!}
\usage{
add(x, y)
}
\arguments{
\item{x}{a numeric value, again.}

\item{y}{also a numeric value. Already told you so!
#'}
}
\value{
guess what - a numeric value!
}
\description{
And I also don't want a subtitle.
}
\examples{
add(1,2)
>3

}
