% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numbers.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{The first number}

\item{y}{The second number}
}
\value{
The sum of x and y
}
\description{
This function adds two numbers
}
\examples{
add_numbers(2,3)
}
