% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addValue.R
\name{addValue}
\alias{addValue}
\title{Add values}
\usage{
addValue(x, y)
}
\arguments{
\item{x}{One value}

\item{y}{another value}
}
\value{
numeric value
}
\description{
A simple function that adds values
}
\examples{
val <- addValue(1,3)
}
