% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_plot.R
\name{admix_plot}
\alias{admix_plot}
\title{Plot STRUCTURE/ADMIXTURE results}
\usage{
admix_plot(q.mat, pop.col = 1, prob.col = 2, sort.probs = TRUE,
  label.pops = TRUE, col = NULL, horiz = TRUE,
  legend.position = c("top", "left", "right", "bottom", "none"))
}
\arguments{
\item{q}{matrices}
}
\value{
ggplot for different K
}
\description{
The function plots STRUCTURE/ADMIXTURE plots given Q-matrices.
}
