% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajoute_deux.R
\name{ajoute_deux}
\alias{ajoute_deux}
\title{Ajoute deux}
\usage{
ajoute_deux(x)
}
\arguments{
\item{x}{Un numérique.}
}
\value{
y Ce numérique plus deux.
}
\description{
Prends un nombre quelconque en entrée, et retourne ce nombre plus deux.
}
\examples{

ajoute_deux(7)

}
