% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeList.R
\name{aList}
\alias{aList}
\title{Listing the data for individual patient}
\usage{
aList(inputData)
}
\arguments{
\item{inputData}{An imported dataset}
}
\value{
A table containing the Glucose Value and Internal Time for a specific individual
}
\description{
Listing the data for individual patient
}
