% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatever.R
\name{amean}
\alias{amean}
\title{My Function}
\usage{
amean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical value}
}
\value{
Computed mean
}
\description{
Something...
}
\examples{
stopifnot(amean(c(1, 2, 3)) == 2)
stopifnot(is.na(amean(c(1, NA, 3))))
stopifnot(amean(c(1, NA, 3), TRUE) == 2)
stopifnot(is.nan(amean(numeric(0))))

}
