% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Return animal's sound}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{An animal}

\item{sound}{A sound}
}
\value{
A statement about the animal's sound
}
\description{
Return animal's sound
}
\examples{
animal_sounds("dog", "woof")
animal_sounds("cat", "meow")
}
