% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{anotherfun}
\alias{anotherfun}
\title{second function
there are no arguments to this function}
\usage{
anotherfun(x)
}
\value{
just prints stuff
}
\description{
second function
there are no arguments to this function
}
\details{
more destil about this func
}
