% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APC_C_N.R
\name{APC_C_N}
\alias{APC_C_N}
\title{Sample grouping with APC gene mutation status}
\usage{
APC_C_N(SNV, is.TCGA = TRUE, sample_id_data)
}
\arguments{
\item{SNV}{A data frame containing SNV data.即MAF格式的数据}

\item{is.TCGA}{is TCGA data or not}

\item{sample_id_data}{data.frame格式入组患者编号，表头名称需为Tumor_Sample_Barcode}
}
\value{
A data frame of sample grouping with APC gene
}
\description{
This function groups samples based on mutations at both ends of the APC gene according to MAF format data.
}
\examples{
data("SNV_data")
data("gene_group_data")
colnames(gene_group_data)[colnames(gene_group_data) == "SAMPLE_ID"] <- "Tumor_Sample_Barcode"
result <- APC_C_N(SNV = SNV_data,
                  is.TCGA=TRUE,
                  sample_id_data=gene_group_data)
print(result)
}
\references{
1. Mondaca S, Walch H, Nandakumar S, Chatila WK, Schultz N, Yaeger R. \bold{Specific Mutations in APC,
   but Not Alterations in DNA Damage Response, Associate With Outcomes of Patients With Metastatic Colorectal Cancer.}
   \emph{Gastroenterology}. 2020;159(5):1975-1978.e4. doi:10.1053/j.gastro.2020.07.041\cr
2. Peng H, Ying J, Zang J, et al. \bold{Specific Mutations in APC,
   with Prognostic Implications in Metastatic Colorectal Cancer.}
   \emph{Cancer Res Treat}. 2023;55(4):1270-1280. doi:10.4143/crt.2023.415
}
\author{
Pengmin Yang
}
