% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageGapSize.R
\name{averageGapSize}
\alias{averageGapSize}
\title{Calculate the average gap size}
\usage{
averageGapSize(individualDataset)
}
\arguments{
\item{inputId}{A string of he unique identifier}
}
\value{
The average gap size
}
\description{
Calculate the average gap size
}
