% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveTimeGap.R
\name{aveTimeGap}
\alias{aveTimeGap}
\title{Calculate the most commonn time gap.}
\usage{
aveTimeGap(individualDataset)
}
\arguments{
\item{inputId}{A string of he unique identifier.}
}
\value{
The most common time gap.
}
\description{
Calculate the most commonn time gap.
}
