% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{B.loss.draws}
\alias{B.loss.draws}
\title{BINDER LOSS Function}
\usage{
B.loss.draws(PYclust, parti = PYclust, useless_param = TRUE)
}
\description{
Computes Binder loss for each partition and the minimizer
}
