% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_Newton.R
\name{backtrack}
\alias{backtrack}
\title{Backtracking}
\usage{
backtrack(fx, x, t, df, alpha = 0.5, beta = 0.9)
}
\arguments{
\item{fx}{handle to function that returns objective function values}

\item{x}{current parameter estimate}

\item{t}{current step-size}

\item{df}{the value of the gradient of objective function evaluated at the current x}

\item{alpha}{the backtracking parameter}

\item{beta}{the decrementing multiplier}
}
\description{
Backtracking
}
