% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{BarcodescoreR}
\alias{BarcodescoreR}
\title{BarcodescoreR}
\usage{
BarcodescoreR(barcodes, barlength)
}
\arguments{
\item{barcodes}{a character vector containing barcodes either 6 or 8 nucleotides in length}

\item{barlength}{Either 6 or 8 indicating the length of the barcodes to compute the AC percentage at each position for.}
}
\value{
a vector containing the percentages at each position as well as a message indicating whether a barcode has been duplicated or not.
}
\description{
Barcodes are used for multiplexing deep sequencing reactions.  However, in order to ensure the barcodes can be demultiplexed
efficiently, illumina requirea as close to a AC:TG ratio of 1:1 as possible for each position of the barcode.  Therefore, the
AC percentage of the barcode at a given position should be close to .5 across all barcodes.
}
\examples{
barcodes1 <- c("AACGCC","AAGGTA","AATTCC","ACACAG","ACACTC","ACACTG","ACAGGA","ACCTGT","ACGAAG","ACGACT","ACGTCA","CTGTTC",
               "GCCTAA","GATCTG","GTAGCC","GGAACT","GGACGG","GCGGAC","GGCCAC","GCTACC","GCTCAT","GTATAG","GTCGTC","GAATGA")
BarcodescoreR(barcodes1, barlength = 6)

barcodes2 <- c("AACGCCAT","AAGGTACG","AATTCCGG","ACACAGAG","ACACTCAG","ACACTGTG","ACAGGACA","ACCTGTAG","ACGAAGGT","ACGACTTG",
               "ACGTCAAC","CTGTTCAC","GCCTAAGT","GATCTGGT","GTAGCCGT","GGAACTGT","GGACGGGT","GCGGACGT","GGCCACGT","GCTACCGT",
               "GCTCATGT","GTATAGGT","GTCGTCGT","GAATGAGT")
BarcodescoreR(barcodes2, barlength = 8)

barcodes3 <- c("AACGCC","AAGGTA","AATTCC","ACACAG","ACACTC","ACACTG","ACAGGA","ACCTGT","ACGAAG","ACGACT","ACGTCA","CTGTTC",
               "GCCTAA","GATCTG","GTAGCC","GGAACT","GGACGG","GCGGAC","GGCCAC","GCTACC","GCTCAT","GTATAG","GTCGTC","AACGCC")
BarcodescoreR(barcodes3, barlength = 6)
}
\author{
Brendan Gongol
}
