% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{BarcodeselectR}
\alias{BarcodeselectR}
\title{BarcodeselectR}
\usage{
BarcodeselectR(barcodes, number, barcode_length)
}
\arguments{
\item{barcodes}{A character vector containing barcodes either 6 or 8 nucleotides in length}

\item{number}{A numerical value indicating the number of barcodes the user wishes the function to return.}

\item{bar_length}{A numerical value, either 6 or 8, indicating the length of the barcodes to compute.}
}
\value{
A data table.  Each row contains the combination of barcodes analyzed, the AC percentage at each position, and a numerical score of the overal barcode AC:TG diversity.
}
\description{
Barcodes are used for multiplexing deep sequencing reactions.  However, in order to ensure the barcodes can be demultiplexed
efficiently, illumina requirea as close to a AC:TG ratio of 1:1 as possible for each position of the barcode.  Therefore, the
AC percentage of the barcode at a given position should be close to .5 across all barcodes.  This function returns the AC percentage
for all combinations of specified barcodes.
}
\examples{
practice8 <- c("GTGTTCTA","AAACATCG","CATCAAGT","GTCTGTCA","AACAACCA","CATACCAA",
               "TGGAACAA","ACCTCCAA","ACACGACC","CTGAGCCA","ACACAGAA","TATCAGCA",
               "CTGTAGCC","TGGCTTCA","CCAGTTCA","AGATGTAC","ACAGCAGA","ACAAGCTA",
               "ACGCTCGA","CACTTCGA","AAGACGGA","AACGCTTA","GAATCTGA", "ATCCTGTA")
DT <- BarcodeselectR(barcodes = practice8, number = 23, barcode_length = 8)
DT <- DT[order(DT$score),]
DT

practice6 <- c("GTGTTC","AAACAT","CATCAA","GTCTGT","AACAAC","CATACC",
               "TGGAAC","ACCTCC","ACACGA","CTGAGC","ACACAG","TATCAG",
               "CTGTAG","TGGCTT","CCAGTT","AGATGT","ACAGCA","ACAAGC",
               "ACGCTC","CACTTC","AAGACG","AACGCT","GAATCT", "ATCCTG")
DT <- BarcodeselectR(barcodes = practice6, number = 23, barcode_length = 6)
DT <- DT[order(DT$score),]
DT
}
\author{
Brendan Gongol
}
