\name{bigen}
\alias{bigen}
\title{Binomial Distribution}
\usage{
bigen(p)
}
\description{
This function generates a random number from Binomial Distribution which is number of success in n Bernoulli trial.
}
\arguments{
  \item{p}{probability of seccuss on each trial}
}
\references{
  dr.gharib and mr.heidari
}
\examples{
brgen(0.2)
}
