% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindStudyEvalutaion.R
\name{bindStudyEvalutaion}
\alias{bindStudyEvalutaion}
\title{学習用/評価用データの結合}
\usage{
bindStudyEvalutaion(STUDY, EVALUATION)
}
\arguments{
\item{STUDY}{data.frame 学習用データ}

\item{EVALUATION}{data.frame 評価用データ}
}
\value{
data.frame 結合されたデータ
}
\description{
学習用/評価用データの結合
}
