% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainbody}
\alias{brainbody}
\title{Cranes Data Set}
\format{
A data frame with 2 variables: \code{species}, \code{brain},
\code{body}, \code{gestation}, and \code{litter}.
}
\usage{
brainbody
}
\description{
This data set contains information on the brain weight, body weight,
gestation length, and litter size for various animals.
}
\keyword{datasets}
