\name{brgen}
\alias{brgen}
\title{Bernoulli Numbers}
\usage{
brgen(p)
}
\description{
This function generates a random number from Bernoulli distribution.
meaning that returns 1 with probability p and 0 with probability (1-p).
}
\arguments{
  \item{p}{probability of seccuss}
}
\references{
  dr.gharib and mr.heidari
}
\examples{
brgen(0.2)
}
