% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_mean.R
\name{calculate_group_mean}
\alias{calculate_group_mean}
\title{Calculate Group Mean}
\usage{
calculate_group_mean(df, numeric_var, factor_var)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{numeric_var}{A string representing the name of the numeric variable.}

\item{factor_var}{A string representing the name of the factor (binary) variable.}
}
\value{
A numeric value representing the mean of the specified numeric variable for the filtered rows.
}
\description{
This function calculates the mean of a numeric variable for rows where another variable has the value TRUE.
}
\examples{
calculate_group_mean(example_data, "Age", "Likes_Coffee")
}
