% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_2.R
\name{calculate_idf}
\alias{calculate_idf}
\title{Calculate Intensity-Duration-Frequency (IDF) quantiles and generate plot}
\usage{
calculate_idf(
  data_list,
  return_periods = c(2, 5, 10, 25, 50, 100),
  durations = c(1, 2, 5, 10, 24, 48),
  aggregation_function = "mean",
  distribution = "gev",
  ...
)
}
\arguments{
\item{data_list}{a list of data tables containing time series data with columns "time", "value", and "location_id"}

\item{return_periods}{vector of return periods for which IDF quantiles are calculated}

\item{durations}{vector of durations for which IDF quantiles are calculated (in hours)}

\item{aggregation_function}{aggregation function for calculating rolling statistics (default is "mean")}

\item{distribution}{distribution to fit for IDF modeling (default is "gev")}
}
\value{
a list containing two elements:
  - 'data': a data table with columns for location id, duration, return period, and corresponding IDF values
  - 'plot': a ggplot object visualizing the IDF curve
}
\description{
This function calculates IDF quantiles based on input data and parameters, and
generates a plot for visual representation of the IDF curve
}
\examples{
idf_result <- calculate_idf(data_list = my_data, return_periods = c(2, 5, 10, 25, 50, 100), durations = c(1, 2, 5, 10, 24, 48))
print(idf_result$data)
print(idf_result$plot)
}
