% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{calculateRFM}
\alias{calculateRFM}
\title{calculate RFM}
\usage{
calculateRFM(transactions, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\value{
Returns a data table containing...
}
\description{
calculate RFM
}
\details{
\code{data} contains the data

Return value
}
