% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDays.R
\name{calDays}
\alias{calDays}
\title{Calculate the number of days}
\usage{
calDays(individualDataset)
}
\arguments{
\item{inputId}{A string of he unique identifier.}
}
\value{
The number of days.
}
\description{
Calculate the number of days
}
