% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\alias{initialize,Candidate-method}
\alias{show,Candidate-method}
\title{Candidate object}
\description{
Objects of class \code{Candidate} are created using the \code{createCandidate} function
}
\details{
#' An object of the class `Candidate' has the following slots:
\itemize{
\item \code{name} Name of the candidate
\item \code{party} Candidate's party, either "Democratic" or "Republican"
\item \code{delegatesWon} Number of delegates the candidate has won so far
\item \code{delegatesNeeded} Number of additional delegates needed to secure nomination
}
}
\author{
Jacob H. Hample: \email{jacob.hample@wustl.edu}
}

