% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_var.R
\name{cap_var}
\alias{cap_var}
\title{Cap}
\usage{
cap_var(x, cap, floor)
}
\arguments{
\item{x}{Numerisk vektor}

\item{cap}{Maksverdi}

\item{floor}{Minverdi}
}
\value{
Returnerer en cappet versjon av x.
}
\description{
Denne funksjonen tar inn en vektor, et tak (maksverdi), et gulv (minverdi) og returnerer vektoren med maksverdi lik takverdi
og gulv lik minimumsverdi.
}
\examples{

data <- c(1:10)
cap_var(data, cap = 8, floor = 2)

}
