% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchphrase.R
\name{catchphrase}
\alias{catchphrase}
\title{my_function}
\usage{
catchphrase(character = "")
}
\arguments{
\item{character}{Character string with options for lab member names}
}
\value{
Catchphrase of the chosen character
}
\description{
This function will add two parameters together and return the sum
}
