% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_format_date.R
\name{cd_format_date}
\alias{cd_format_date}
\title{Corporate Design: Format Date}
\usage{
cd_format_date(date, lang = default(rmarkdown::metadata$lang, "en-US"))
}
\arguments{
\item{date}{Current date.}

\item{lang}{Language, e.g. \code{"de-DE"}, \code{"de-CH"} or \code{"en-US"}. Defaults to
the document's \code{lang} metadata field, or \code{"en-US"}.}
}
\value{
Character string, date in local format.
}
\description{
Format date in local language format.
}
\examples{
cd_format_date("2012-01-01", "de-DE")
cd_format_date("2012-01-01", "en-US")
}
