% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_page_title.R
\name{cd_page_title}
\alias{cd_page_title}
\title{Corporate Design: Title Page}
\usage{
cd_page_title(
  title = default(rmarkdown::metadata$title, "Title"),
  subtitle = default(rmarkdown::metadata$subtitle, "Subtitle"),
  date = default(rmarkdown::metadata$date, cd_format_date(Sys.Date()))
)
}
\arguments{
\item{title}{Document title}

\item{subtitle}{Document subtitle}

\item{date}{Document creation date}
}
\value{
Object of class \code{"knit_asis"} (so that knitr will treat it as is). LaTeX code for title page.
}
\description{
Example function to generate a title page
}
\examples{
cd_page_title(
  title = "My Title",
  subtitle = "My Subtitle"
)
}
