% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeFactor.R
\name{changeFactor}
\alias{changeFactor}
\title{文字列データをfactorに変換}
\usage{
changeFactor(DATA)
}
\arguments{
\item{data.frame}{データフレーム}
}
\value{
data.frame 文字列の列が factor 化されたデータ
}
\description{
文字列データをfactorに変換
}
