% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRegu.R
\name{checkRegu}
\alias{checkRegu}
\title{Check if the time series is regular or irregular}
\usage{
checkRegu(individualDataset)
}
\arguments{
\item{individualDataset}{An individual dataset}
}
\value{
A summary table of checking regularity
}
\description{
Check if the time series is regular or irregular
}
