% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_fit_DDM_only.R
\name{chi_square_optim_DDM}
\alias{chi_square_optim_DDM}
\title{chi_square_optim_DDM}
\usage{
chi_square_optim_DDM(
  params,
  observations,
  returnFit,
  model = 1,
  hm_up = NULL,
  hm_low = NULL
)
}
\arguments{
\item{params}{Parameters of the model to be fitted + fixed parameters}

\item{observations}{Data to be fitted}

\item{returnFit}{if 0, then returns predicted data. Else, fit the parameters}

\item{denominator}{Divide by predicted values in the goal function if True}
}
\value{
predicted data if returnFit == 0, else returns the prediction error
}
\description{
chi_square_optim_DDM
}
\details{
This is the function used to simultaneously fit reaction time, accuracy and confidence judgments via quantile optimization.
}
