% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_fit_DDM_only.R
\name{chi_square_optim_DDM_fullconfRT}
\alias{chi_square_optim_DDM_fullconfRT}
\title{chi_square_optim_DDM_fullconfRT}
\usage{
chi_square_optim_DDM_fullconfRT(
  params,
  observations,
  returnFit,
  confRT_name = "RTconf"
)
}
\arguments{
\item{params}{Parameters of the model to be fitted + fixed parameters}

\item{observations}{Data to be fitted}

\item{returnFit}{if 0, then returns predicted data. Else, fit the parameters}

\item{confRT_name}{Name of the column in observations containing the confidence reaction time}
}
\value{
predicted data if returnFit == 0, else returns the prediction error
}
\description{
chi_square_optim_DDM_fullconfRT
}
\details{
This is the function used to simultaneously fit reaction time, accuracy and confidence judgments via quantile optimization.
}
