% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ChiPseqPeakPlotter}
\alias{ChiPseqPeakPlotter}
\title{ChiPseqPeakPlotter}
\usage{
ChiPseqPeakPlotter(treatment_bw, control_bw, chr, start, end, average_dist,
  fill_dist, type = "all", peaks_tx, peaks_ctrl)
}
\arguments{
\item{treatment_bw}{a character vector indicating the location of the treatment bigwig file.}

\item{control_bw}{a character vector indicating the location of the treatment bigwig file.}

\item{chr}{a character vector indicating the chromosome number.}

\item{start}{a number indicating the start site.}

\item{end}{a number indicating the end site.}

\item{average_dist}{a number indicating the number of bases to average the peak height.}

\item{fill_dist}{a number indicating the number of average replicates to fill.}

\item{type}{a character either "all" or "called" indicating whether to include all peaks for a given area or only those called by peak caller.}

\item{peaks_tx}{a character vector indicating the location of the treatment MACS2 peak file.}

\item{peaks_ctrl}{a character vector indicating the location of the control MACS2 peak file.}
}
\value{
a plot indicating the peak score for a specified chromosome position.
}
\description{
This function plotts the peak height of a CHiPseq experiment for a given chromosome location.
}
\examples{
ChiPseqPeakPlotter(treatment_bw <- "./example_datasets/8hr-PS1_S16_L008_R1_001.fastq.gz.hisat_PS_8hr.bw",
                   control_bw <- "./example_datasets/bigwig_files/8hr-OS1_S7_L008_R1_001.fastq.gz.hisat_OS_8hr.bw",
                   chr <- "10",
                   start <- 69266000,
                   end <- 69405881,
                   average_dist <- 50,
                   fill_dist <- 50,
                   type = "all")

ChiPseqPeakPlotter(treatment_bw <- "./example_datasets/8hr-PS1_S16_L008_R1_001.fastq.gz.hisat_PS_8hr.bw",
                   control_bw <- "./example_datasets/8hr-OS1_S7_L008_R1_001.fastq.gz.hisat_OS_8hr.bw",
                   chr <- "10",
                   start <- 69266000,
                   end <- 69405881,
                   average_dist <- 50,
                   fill_dist <- 50,
                   type = "called",
                   peaks_tx <- "./example_datasets/8hr-PS1_S16_L008_R1_001.fastq.gz.hisat_PS_8hr.bam_macs2_peaks.annotatedpeakAnno.xls",
                   peaks_ctrl <- "./example_datasets/8hr-OS1_S7_L008_R1_001.fastq.gz.hisat_OS_8hr.bam_macs2_peaks.annotatedpeakAnno.xls")
}
\author{
Brendan Gongol
}
