% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ChiPseqregionscoreR}
\alias{ChiPseqregionscoreR}
\title{ChiPseqregionscoreR}
\usage{
ChiPseqregionscoreR(treatment_bw, control_bw, chr, start, end,
  average_dist)
}
\arguments{
\item{treatment_bw}{a character vector indicating the location of the treatment bigwig file.}

\item{control_bw}{a character vector indicating the location of the treatment bigwig file.}

\item{chr}{a character vector indicating the chromosome number.}

\item{start}{a number indicating the start site.}

\item{end}{a number indicating the end site.}

\item{average_dist}{a number indicating the number of bases to average the peak height.}
}
\value{
a data.table of the average peak score for a specified chromosome region.
}
\description{
This function returns the average peak height of a CHiPseq experiment for a given chromosome location.
}
\examples{
dt <- ChiPseqregionscoreR(treatment_bw <- "./results/bigwig_files/2hr-PS1_S13_L008_R1_001.fastq.gz.hisat_PS_2hr.bw",
                          control_bw <- "./results/bigwig_files/2hr-OS1_S4_L008_R1_001.fastq.gz.hisat_OS_2hr.bw",
                          chr <- "12",
                          start <- 48105253 - 5000, # 48101253
                          end <- 48105253 + 5000, # 48150404
                          average_dist <- 50)
hr2 <- data.table(dt[dt$fill == "tx",]$mea/dt[dt$fill == "ctrl",]$mea)
hr2$time = "hr2"

dt <- ChiPseqregionscoreR(treatment_bw <- "./results/bigwig_files/8hr-PS1_S16_L008_R1_001.fastq.gz.hisat_PS_8hr.bw",
                   control_bw <- "./results/bigwig_files/8hr-OS1_S7_L008_R1_001.fastq.gz.hisat_OS_8hr.bw",
                   chr <- "12",
                   start <- 48105253 - 5000, # 48101253
                   end <- 48105253 + 5000, # 48150404
                   average_dist <- 50)
hr8 <- data.table(dt[dt$fill == "tx",]$mea/dt[dt$fill == "ctrl",]$mea)
hr8$time = "hr8"

dt <- ChiPseqregionscoreR(treatment_bw <- "./results/bigwig_files/16hr-OS1_S10_L008_R1_001.fastq.gz.hisat_OS_16hr.bw",
                          control_bw <- "./results/bigwig_files/16hr-PS1_S19_L008_R1_001.fastq.gz.hisat_PS_16hr.bw",
                          chr <- "12",
                          start <- 48105253 - 5000, # 48101253
                          end <- 48105253 + 5000, # 48150404
                          average_dist <- 50)
hr16 <- data.table(dt[dt$fill == "tx",]$mea/dt[dt$fill == "ctrl",]$mea)
hr16$time = "hr16"

dt2 <- rbind(hr2, hr8)
dt2 <- rbind(dt2, hr16); setnames(dt2, colnames(dt2), c("mea", "fill"))
dt2 <- dt2[dt2$mea > 0,]
dt2$mea <- log2(dt2$mea)
dt2 <- dt2[!dt2$mea == "-Inf",]
dt2 <- dt2[!dt2$mea == "Inf",]
dt2
}
\author{
Brendan Gongol
}
