% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Romove the data that not starting around 0:00 and ending around 23;59}
\usage{
clean(data)
}
\arguments{
\item{data}{Imported daclastaset}
}
\value{
A clean dataset
}
\description{
Romove the data that not starting around 0:00 and ending around 23;59
}
