% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_unemp}
\alias{clean_unemp}
\title{Easily filter unemployment data for Luxembourg}
\usage{
clean_unemp(
  unemp_data,
  year_of_interest = NULL,
  place_name_of_interest = NULL,
  level_of_interest = NULL,
  col_of_interest
)
}
\arguments{
\item{year_of_interest}{Optional: The year that should be kept. Leave empty to select every year.}

\item{place_name_of_interest}{Optional: The name of the place of interest: leave empty to select every place in `level_of_interest`.}

\item{level_of_interest}{Optional: The level of interest: one of `Country`, `Canton`, `Commune`. Leave empty to select every level with the same place name.}

\item{col_of_interest}{A column of the `unemp` data frame that you wish to select.}

\item{unemp}{A data frame containing unemployment data for Luxembourg.}
}
\value{
A data frame
}
\description{
Easily filter unemployment data for Luxembourg
}
\details{
Users can filter data on two variables: the name of the place of interest, and the level of interest.
By leaving the argument `place_name_of_interest` empty
}
\examples{
# Filter on cantons
clean_unemp(unemp,
            level_of_interest = "Canton",
            active_population)
# Filter on a specific commune
clean_unemp(unemp_2013,
            place_name_of_interest = "Luxembourg",
            level_of_interest = "Commune",
            active_population)
# Filter on every level called Luxembourg
clean_unemp(unemp_2013,
            place_name_of_interest = "Luxembourg",
            active_population)
}
