% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster.reorder}
\alias{cluster.reorder}
\title{Reordering a dataframe of partitions}
\usage{
cluster.reorder(clust.matrix)
}
\description{
Reorders output clusters in such a way that the cluster number is increasing by 1 starting from 1 in order of appearance
}
