% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMean.R
\name{cMean}
\alias{cMean}
\title{Compute the Mean of a Vector}
\usage{
cMean(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The mean of `x`.
}
\description{
This function uses the .C interface to call a pre-compiled `C` function to compute the mean of a vector.
}
