% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_SV_get.R
\name{CNV_SV_get}
\alias{CNV_SV_get}
\title{Retrieve CNV and SV data from the NGS registration form}
\usage{
CNV_SV_get(data, Type = c("重排/融合", "扩增"))
}
\arguments{
\item{data}{is the data of NGS registration form}

\item{Type}{A list.eg.Type=c("重排/融合","扩增")}
}
\value{
A data frame with gene amplification and fusion information with each sample
}
\description{
This function Retrieve CNV and SV data from the NGS registration form.
}
\examples{
data("data")
result<-CNV_SV_get(data=data,Type=c("重排/融合","扩增"))
print(result)
}
\author{
Pengmin Yang
}
