% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_diffs.R
\name{codon_diff}
\alias{codon_diff}
\title{Find codon differences between sequences}
\usage{
codon_diff(
  baseline_seq,
  compare_seqs,
  diff_type = c("synonymous", "non-synonymous", "all"),
  output = c("num", "position"),
  seq_type = c("DNA", "RNA"),
  ignore_case = TRUE
)
}
\arguments{
\item{baseline_seq}{Character vector of length one that all other sequences
are being compared to.}

\item{compare_seqs}{Character vector of any length that will be compared to
the baseline sequence.}

\item{diff_type}{Either "synonymous", "non-synonymous", or "all".
\itemize{
\item "all" : find all the codon differences
\item "non-synonymous" : find all of the non-synonymous codon differences
\item "synonymous" : find all of the synonymous codon differences
}}

\item{output}{Either "num" (return the number of differences) or "position"
(return the positions of the differences).}

\item{seq_type}{Character value of "DNA" or "RNA" describing whether the sequences
use "T" or "U".}

\item{ignore_case}{Logical value (TRUE / FALSE) designating whether the
function should ignore case. Defaults to TRUE.}
}
\value{
Either a list of numeric vectors describing the positions of the
differences or a numeric vector of the number of differences.
}
\description{
Function takes a baseline sequence and any number of sequences to compare to
and determines the number or the positions of the codon (3 characters)
differences.
}
\examples{
codon_diff('AGTGGG', 'CGTGAG', output = 'num', diff_type = 'all', seq_type = 'DNA')

seq1 = 'ACTGACTGG'
seq2 = c('ACGGGCCGG', 'ACTGACTGG', 'GCTCAGTGG')
codon_diff(seq1, seq2, output = 'num', diff_type = 'synonymous',
    seq_type = 'DNA', ignore_case = TRUE)
codon_diff(seq1, seq2, output = 'position', diff_type = 'non-synonymous',
    seq_type = 'DNA', ignore_case = FALSE)

seq1 = 'ACUGACUGG'
seq2 = c('ACGGGCCGG', 'ACUGACUGG', 'GCUCAGUGG')
codon_diff(seq1, seq2, output = 'num', diff_type = 'synonymous',
    seq_type = 'RNA', ignore_case = TRUE)
codon_diff(seq1, seq2, output = 'position', diff_type = 'non-synonymous',
    seq_type = 'RNA', ignore_case = FALSE)

\dontrun{
codon_diff(seq1, seq2)
codon_diff('AGTCGA', 'TTA')
}
}
