% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_aa.R
\name{codon_to_aa}
\alias{codon_to_aa}
\title{Convert codons to amino acids}
\usage{
codon_to_aa(x, seq_type = c("DNA", "RNA"))
}
\arguments{
\item{x}{Character vector of three nucleotide codons}

\item{seq_type}{Character value of either "DNA" or "RNA" depending on whether
"T" or "U" is used in the sequence.}
}
\value{
Character vector of amino acids
}
\description{
Takes a character vector of codons and converts them to their amino acids.
Useful for creating an amino acid column from a column of codons, in order to
facet a plot for example.
}
\examples{
codon_to_aa(c('CAT', 'GAC', 'TTT', 'ATG'), seq_type = 'DNA')
codon_to_aa('GAT', seq_type = 'DNA')
codon_to_aa(c('CAU', 'GAC', 'UUU', 'AUG'), seq_type = 'RNA')
codon_to_aa('GAU', seq_type = 'RNA')
}
