% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMode.R
\name{colMode}
\alias{colMode}
\title{Return the mode of a column of a data set.}
\usage{
colMode(x)
}
\arguments{
\item{x}{A data frame column.}
}
\value{
The mode of column \code{y}.
}
\description{
Return the mode of a column of a data set.
}
\examples{
colMode(mtcars$cyl)
}
