% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maximal_partial_clique3.R
\name{compute_maximal_partial_clique3}
\alias{compute_maximal_partial_clique3}
\title{Compute maximal partial clique using a heuristic approach}
\usage{
compute_maximal_partial_clique3(adj_mat, alpha, max_iteration = 1000)
}
\arguments{
\item{adj_mat}{A symmetrical matrix with 0s and 1s. Its diagonal must be 1s and it must have 5 to 50 rows.}

\item{alpha}{The prespecified density we can accept for a partially connected clique.}

\item{max_iteration}{Maximum number of iterations.}
}
\value{
A list containing the number of nodes that are included in the maximal partial clique and the actual density of this clique.
}
\description{
Compute maximal partial clique using a heuristic approach
}
\examples{
data <- matrix(sample(0:1, 100, replace = TRUE), 10, 10)
data[lower.tri(data)] <- t(data)[lower.tri(data)]
diag(data) <- 1
result <- compute_maximal_partial_clique3(data, alpha=0.7)
}
