% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{ConfidenceInterval}
\alias{ConfidenceInterval}
\title{Calculate confidence interval for mean}
\usage{
ConfidenceInterval(df)
}
\arguments{
\item{df}{Numeric 2 dimension vector}
}
\value{
Good way: one number - confidence interval for mean of whole
input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the confidence interval for mean
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
ConfidenceInterval(data.frame)
ConfidenceInterval(matrix)
ConfidenceInterval(vector)
ConfidenceInterval(List)
}
