% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_SNPS.R
\name{count_SNPS}
\alias{count_SNPS}
\title{Count number of synonymous and non synonymous SNP.}
\usage{
count_SNPS(dnabin)
}
\arguments{
\item{DNAbin}{object.}
}
\value{
list of SNPs count
}
\description{
The function counts the number of SNP changes.
}
\details{
Count the total number of SNPs and further anayses number of syn vs non-syn SNPs.
}
