% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numDays.R
\name{countDays}
\alias{countDays}
\title{Plot the Glucose Value on y-axis and Internal Time on x-axis basing on ID}
\usage{
countDays(inputId)
}
\arguments{
\item{inputId}{A string of he unique identifier}
}
\value{
Number of days recorded for Glucose Value
}
\description{
Plot the Glucose Value on y-axis and Internal Time on x-axis basing on ID
}
