% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cranes}
\alias{cranes}
\title{Cranes Data Set}
\format{
A data frame with 2 variables: \code{cranes} and \code{year}.
}
\usage{
cranes
}
\description{
This data set contains information on the number of cranes at Aransas
National Wildlife Refuge in Austwell, Texas by year from 1938 to 2016.
}
\keyword{datasets}
