% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_object.R
\name{create}
\alias{create}
\title{Pour créer un objet de type Random_walk}
\usage{
create(sigma)
}
\arguments{
\item{sigma}{un double qui contient l'écart type de la marche aléatoire.}
}
\value{
un objet de class Random_walk
}
\description{
Pour créer un objet de type Random_walk
}
\examples{
create(2)

}
