% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelHM_plot.R
\name{create_modelHM}
\alias{create_modelHM}
\title{Plot a confidence heat map using the alpha/beta model}
\usage{
create_modelHM(
  alpha,
  beta,
  type = "upper",
  plot = T,
  cexax = 1.5,
  cexlab = 3,
  cexmain = 2.5,
  dt = 0.0025,
  ev_bound = 0.5,
  ev_window = 0.0125,
  upperRT = 5
)
}
\arguments{
\item{alpha}{}

\item{beta}{}

\item{type}{One of 3 possible types of heat map (upper, lower and symmetrical).
Upper (resp. lower) shows the confidence linked to an upper (resp. lower) boundary hit.
Symmetrical shows the symmetrical heat map. Defaults to "upper"}

\item{plot}{If TRUE, the heat map is directly plotted. Defaults to TRUE}

\item{cexax}{Axis font size}

\item{cexlab}{Labels font size}

\item{cexmain}{Title font size}

\item{dt}{Time resolution}

\item{ev_bound}{Maximum/minimum evidence}

\item{ev_window}{Evidence resolution. Should be sufficiently higher than dt
to avoid NULL values within the heat map}

\item{upperRT}{Maximum time}
}
\value{

}
\description{
Plot a confidence heat map using the alpha/beta model
}
