% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Creates an object of class `Candidate'}
\usage{
createCandidate(name, party, delegatesWon)
}
\arguments{
\item{name}{A string with the name of the presidential candidate}

\item{party}{A string with the candidate's party, either "Democratic" or "Republican"}

\item{delegatesWon}{A numeric with the number of delegates the candidate has won thus far}
}
\value{
An object of class `Candidate' that contains
\itemize{
\item \code{name} Name of the candidate
\item \code{party} Candidate's party, either "Democratic" or "Republican"
\item \code{delegatesWon} Number of delegates the candidate has won so far
\item \code{delegatesNeeded} Number of additional delegates needed to secure nomination
}
}
\description{
A function to create a new object of class Candidate, that automatically calculates
how many delegates the candidate still needs to secure her party's nomination.
}
\examples{

createCandidate("Hillary Clinton", "Democratic", 1614)

}
\author{
Jacob H. Hample: \email{jacob.hample@wustl.edu}
}

