% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelHM_plot.R
\name{custom_sign}
\alias{custom_sign}
\title{Sign of each value within a vector}
\usage{
custom_sign(x)
}
\arguments{
\item{x}{vector of numerical values}
}
\description{
Returns the sign of each value within a vector. If 0, returns 1 or
-1 with equal probability
}
