\docType{class}
\name{DataM-class}
\alias{DataM-class}
\title{Data-like object to be used by the dataMining() function}
\description{
Objects of class \code{DataM} contain 2 slots, the first of
which contains the values for the dependent variable and
the second of which contains a matrix of values for the
predictor variables.
}
\details{
An object of the class `DataM' has the following slots:
\itemize{ \item \code{depvar} The added or subtracted
squared values \item \code{covariates} The added or
subtracted squared values }
}
\author{
Thomas Carroll: \email{thomasscarroll89
}

