% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPartition.R
\name{dataPartition}
\alias{dataPartition}
\title{学習用データを訓練データと検証データに分割}
\usage{
dataPartition(DATA, PERCENTAGE = 0.8)
}
\arguments{
\item{DATA}{data.frame 学習用データ}

\item{PERCENTAGE}{numeric 訓練用データの割合 ( default 0.8 )}
}
\value{
list(訓練データ, 検証データ)
}
\description{
学習用データを訓練データと検証データに分割
}
