% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derviatives.R
\name{dd_exact}
\alias{dd_exact}
\title{Directional Derivative Exact}
\usage{
dd_exact(gradf, a, b)
}
\arguments{
\item{gradf}{handle to function that returns the gradient of the function f}

\item{a}{point at which the directional derivative is evaluated}

\item{b}{the direction vector}
}
\description{
\code{dd_exact} computes the exact directional derivative of the multivariate
function f, at the point a, in the direction b.
}
