% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DDM_with_confidence_slow}
\alias{DDM_with_confidence_slow}
\title{DDM trials generation with fixed post-decision accumulation time}
\usage{
DDM_with_confidence_slow(v, a, ter, z, ntrials, s, dt, t2time, postdriftmod)
}
\arguments{
\item{v}{Drift rate}

\item{a}{Decision boundary, evidence at t = 0 is equal to a*z. Decision boundaries are a and -a}

\item{ter}{Non-decision time}

\item{z}{Starting point bias in evidence. For unbiased DDM, use z = 0}

\item{ntrials}{Number of DDM trials to generate}

\item{s}{Scaling parameter}

\item{dt}{Time step size}

\item{t2time}{Post-decision accumulation time}

\item{postdriftmod}{V-ratio. If equal to 1, post-decision drift rate is equal to decision drift rate}
}
\description{
This function generates trials using the Drift Diffusion Model. Additional
evidence is accumulated for a fixed time after the decision boundary has been reached.
Note: here, decision boundaries are a and -a and starting point is equal to a*z.
}
