% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DDM_with_confidence_slow_fullconfRT}
\alias{DDM_with_confidence_slow_fullconfRT}
\title{DDM trials generation with multiple post-decision accumulation time}
\usage{
DDM_with_confidence_slow_fullconfRT(
  v,
  a,
  ter,
  z,
  ntrials,
  s,
  dt,
  t2distribution,
  postdriftmod
)
}
\arguments{
\item{v}{Drift rate}

\item{a}{Decision boundary, evidence at t = 0 is equal to a*z. Decision boundaries are a and -a}

\item{ter}{Non-decision time}

\item{z}{Starting point bias in evidence. For unbiased DDM, use z = 0}

\item{ntrials}{Number of DDM trials to generate. Must be a multiple of length(t2time)
so that all confidence RT trials are repeated the same number of times}

\item{s}{Scaling parameter}

\item{dt}{Time step size}

\item{postdriftmod}{V-ratio. If equal to 1, post-decision drift rate is equal to decision drift rate}

\item{t2time}{Vector containing the whole confidence RT distribution}
}
\description{
This function generates trials using the Drift Diffusion Model. Additional
evidence is accumulated for a fixed time after the decision boundary has been reached.
Note: here, decision boundaries are a and -a and starting point is equal to a*z.
}
