% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{river where fish is located}
  \item{MILE}{distance upstream in miles}
  \item{SPECIES}{type of fish species}
  \item{LENGTH}{length of fish found in centimeters}
  \item{WEIGHT}{weight of fish found in grams}
  \item{DDT}{level of DDT found in fish}
  ...
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
ddt
}
\description{
A data set on a study into the levels of DDT in parts per million found in various
species of fish in the Tennessee River
}
\keyword{datasets}
