% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown.R
\name{default}
\alias{default}
\title{Set Default Value}
\usage{
default(x, default = "")
}
\arguments{
\item{x}{Character string, or \code{NULL}}

\item{default}{If \code{x} is \code{NULL}, \code{default} is used.}
}
\value{
Character string
}
\description{
Set default value of an argument. If \code{x} is \code{NULL}, \code{default} is used.
}
\examples{
default(NULL)
default(NULL, "my default")
default("a string", "my default")
}
