% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{demultiplex}
\alias{demultiplex}
\title{demultiplex}
\usage{
demultiplex(x, barcode, nreads, cutoff)
}
\arguments{
\item{x}{fastq file to demultiplex}

\item{barcode}{vector containing barcodes to demultiplex}

\item{nreads}{number of records in fastq file to stream through in a given block}

\item{cutoff}{minimum read score cutoff}
}
\value{
a new vector containing the fold changes of the original values.
}
\description{
This function demultiplexes a fastq file into individual fastq files
}
\examples{
demultiplex(x=fastq[1], barcode=c("TT", "AA", "GG"), nreads=50, cutoff=20)
}
\author{
Brendan Gongol
}
