% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Density plot}
\usage{
density_plot(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
ggplot object
}
\description{
Plots a density plot of the data
}
\examples{
density_plot(rnorm(50))

}
