% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_mean.R
\name{df_mean}
\alias{df_mean}
\title{The first line is the title}
\usage{
df_mean(data)
}
\arguments{
\item{data}{data frame}
}
\description{
The second section is a longer
description of the function.
This can go on for multiple lines.
}
\examples{
my_mean(1:5)
}
\seealso{
\code{\link[base]{my_mean}}
}
\author{
Daniel D. Sjoberg
}
