% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derviatives.R
\name{dg_approx}
\alias{dg_approx}
\title{Directional Derivative Approximate}
\usage{
dg_approx(f, a, b, h = 1e-13)
}
\arguments{
\item{f}{handle to function that returns the function f}

\item{a}{point at which the directional derivative is evaluated}

\item{b}{the direction vector}

\item{h}{small displacement}
}
\description{
\code{dg_approx} computes an approximate directional derivative of the multivariate
function f, at the point a, in the direction b.
}
