% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Rolling a dice}
\usage{
dice(n = 100)
}
\arguments{
\item{n}{the number of times rolling the dice.}
}
\value{
the frequency of each result and a frequency bar chart with a red line at the p=1/6 average level.
}
\description{
This function provides utilities to calculate the probabilities of each dice-rolling result and plot the frequency bar chart simultaneously.
}
\examples{
dice(100)
dice(500)
}
