% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{This is my division function}
\usage{
divide(x, y)
}
\arguments{
\item{x}{this is the first value to be divided by}

\item{y}{this is the second value you will divide by}
}
\value{
This function returns the quotient of x and y
}
\description{
This is my division function
}
\examples{
## Start with something simple
divide(1,1)

## Now something more difficult
divide(49,60)

}
