% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{divisi}
\alias{divisi}
\title{As quatro operacoes desagregadas}
\usage{
divisi(a, b)
}
\arguments{
\item{a}{primeiro numero}

\item{b}{segundo numero}
}
\value{
Retorna a divisao de a e b
}
\description{
A divisao de a e b
}
\details{
Calculo aritmetico simples
}
\keyword{a,}
\keyword{b}
\keyword{divisao,}
