\name{dugen}
\alias{dugen}
\title{Uniform Numbers}
\usage{
dugen(begin,end)
}
\description{
This function generates a random number from uniform distribution between begin and end.
}
\arguments{
  \item{begin}{begining of interval}
  \item{end}{end of interval}
}
\references{
  dr.gharib and mr.heidari
}
\examples{
dugen(2,6)
}
