% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy_function.R
\name{Dummy_function}
\alias{Dummy_function}
\title{Dummy Function
Takes in an int and squares it}
\usage{
Dummy_function(x = 2)
}
\arguments{
\item{integer}{}
}
\value{
integer
}
\description{
Dummy Function
Takes in an int and squares it
}

