% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{Perform EM algorithm for ABO blood type example}
\usage{
EM(n, para)
}
\arguments{
\item{n}{data (nA,nAB,nB,nO)}

\item{parameters}{allele frequency probabilities (pA, pB, and pO)}
}
\value{
updated parameters
}
\description{
Perform EM algorithm for ABO blood type example
}
