% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:average.R
\name{ensemble_coding_Avg}
\alias{ensemble_coding_Avg}
\title{Ensemble Coding Averages}
\usage{
ensemble_coding_Avg(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the condition's average.
}
\description{
Ensemble Coding Averages
}
\examples{
ensemble_coding_Avg(c(1,2,3,4,5,6,6,8,9,9))
}
