% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucl_dist.R
\name{eucl_dist}
\alias{eucl_dist}
\title{Euclidean Distance Between Two Points}
\usage{
eucl_dist(P1, P2)
}
\arguments{
\item{P1}{Numeric vector of length 2 representing the coordinates (x, y) of the first point}

\item{P2}{Numeric vector of length 2 representing the coordinates (x, y) of the second point}
}
\value{
d A numeric value representing the Euclidean distance between P1 and P2
}
\description{
Euclidean Distance Between Two Points
}
\examples{
eucl_dist(c(0, 0), c(3, 4))
}
\author{
Your name
}
