% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactci2.R
\name{exactci2}
\alias{exactci2}
\title{Clopper-Pearson Exact Confidence Interval}
\usage{
exactci2(r, n, conflev)
}
\arguments{
\item{r}{Number of reponses}

\item{n}{Total number of subjects
you can continue the description in the next line}

\item{conflev}{Confidence level. Always two-sided.}
}
\value{
Clopper-Pearson CI
}
\description{
This function calculates the Clopper-Pearson exact confidence interval
}
\examples{

exactci2 (r=4, n=20, conflev=0.95)

}
