% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
A data frame with 5 rows and 3 variables:
\describe{
  \item{ID}{A character vector representing the ID of each observation.}
  \item{Age}{A numeric vector representing the age of each observation.}
  \item{Likes_Coffee}{A factor vector indicating whether the observation likes coffee (TRUE) or not (FALSE).}
}
}
\usage{
example_data
}
\description{
A dataset containing example data.
}
\examples{
example_data
}
\keyword{datasets}
