% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_variables.R
\name{exclude_variables}
\alias{exclude_variables}
\title{ダミー化された data.frame の特定の文字列を含む不要な列名を排除}
\usage{
exclude_variables(DATA, VECTOR)
}
\arguments{
\item{DATA}{data.frame 学習用データ}

\item{VECTOR}{vector 削除したい文字列}
}
\value{
vector 抽出された文字列
}
\description{
ダミー化された data.frame の特定の文字列を含む不要な列名を排除
}
