% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_geneIDtoSYMBOL.R
\name{exp_geneIDtoSYMBOL}
\alias{exp_geneIDtoSYMBOL}
\title{Epression data gene ID to SYMBOL}
\usage{
exp_geneIDtoSYMBOL(exp, genecoltype = "ENTREZID")
}
\arguments{
\item{exp}{is the data of gene expression for samples}

\item{genecoltype}{exp gene的类型。 c("ENTREZID","ENSEMBL","MAP");SYMBOL:"CEBPG",ENTREZID:"1054",ENSEMBL:"ENSG00000153879",MAP:"19q13.11"}
}
\description{
This function convert the gene name of expression data to SYMBOL.
}
\examples{
data("exp_raw")
result<-exp_geneIDtoSYMBOL(exp=exp_raw,genecoltype="ENTREZID")
print(result[["gene"]][1:10,1:2])
print(result[["data"]][1:10,1:5])
}
\author{
Pengmin Yang
}
