\name{expgen}
\alias{expgen}
\title{Exponential Distribution}
\usage{
expgen(lambda)
}
\description{
This function generates a random number from Exponential distribution with parameter lambda.
}
\arguments{
  \item{lambda}{parameter of exponential distribution}
}
\references{
  dr.gharib and mr.heidari
}
\examples{
expgen(10)
}
