% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_files_by_date.R
\name{extract_files_by_date}
\alias{extract_files_by_date}
\title{File Identification and Extraction.}
\usage{
extract_files_by_date(partial_file_name, directory, date)
}
\arguments{
\item{partial_file_name}{a character string which is a partial file name}

\item{directory}{a character string which is the directory path for the function to identify files from}

\item{date}{date}
}
\value{
a vector of file names as strings
}
\description{
extract_file_by_date identifies files matching a partial file string and returns those with a date string in their name which is later than a given date. This function is designed to identify files which are regularly produced (data extracts) and to identify potential files which could contain data for processing a within a given time period.

The function is primarily designed to filter out historic data extracts for a given report and identify potential ones which contain the relevant data. Can be utilised to ensure full Periods worth of data is obtained when loaded in and filtered.
}
