% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.R
\name{extract_values_from_nc}
\alias{extract_values_from_nc}
\title{Process NetCDF data and extract values for specified locations}
\usage{
extract_values_from_nc(
  data_directory = "./data",
  file_pattern = ".nc",
  location_ids
)
}
\arguments{
\item{data_directory}{path to the directory containing the files}

\item{file_pattern}{file pattern to match NetCDF files}

\item{location_ids}{vector of location IDs to extract values}
}
\value{
a list of data tables split by location_id
}
\description{
This function reads NetCDF files from a specified directory, extracts relevant
information for specified locations, and returns a list of data tables with columns for time,
location_id, and corresponding values
}
\examples{
extract_values_from_nc(data_directory = "./data", file_pattern = ".nc", location_ids = c(327, 328, 329))

}
