% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Dataset of farm animal sounds}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{letter}{The animal as a character string.}
\item{index}{The corresponding sound.}
}
}
\usage{
farm_animals
}
\description{
A dataset of farm animals and the sounds they make.
}
\keyword{datasets}
