% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state

This is a function that takes the state number and year as an input and displays a plot
of the state with the number of fatalaties mapped as dots at the appropriate location of the
incident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the index of the state}

\item{year}{An interger representing the specified year}
}
\value{
a map, it plots the location of the record on the map for the specified state
and year
}
\examples{
\dontrun{
fars_map_state(state.num = 6, year = 2013)
fars_map_state(5, 2013)
}

}
