% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years

The function fars_read_years is used to take the year as an input and create a filename
using make_filename function. The file is then read using the fars_read function.
If the year is invalid it returns an error message. The function returns the month and year
from a specified data set}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integer input of years}
}
\value{
If an error is generated a warning message "invalid year: " with
the invalid year information is displayed. If no error is generated it displays "NULL"
}
\examples{
\dontrun{
fars_read_years(list(2013, 2014))
fars_read_years(c(2013, 2014, 2015))
fars_read_years(2014)
}

}
