% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize years

The fars_summarize_years function takes a year as an input and summarizes the number of
accidents by month within the input year specified}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Is an integer that takes the years as input and displays a monthwise summary
of the number of fatal injuries}
}
\value{
a data frame displaying a month number and the consolidated number of fatal
injuries foreach month of the specified year
}
\examples{
\dontrun{
fars_summarize_years(list(2013, 2014))
fars_summarize_years(c(2013, 2014))
fars_summarize_years(2013)
}

}
