% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmerge.R
\name{fastmerge}
\alias{fastmerge}
\title{Merges two data frames}
\usage{
fastmerge(d1, d2)
}
\arguments{
\item{d1}{first data frame to merge}

\item{d2}{second data frame to merge}
}
\value{
Merged d1 and d2
}
\description{
Custom function to merge two data frames together, even if they do
not have the exact same column names
}
