% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{FCcalculator}
\alias{FCcalculator}
\title{FCcalculator}
\usage{
FCcalculator(vectorTX, vectorCTRL)
}
\arguments{
\item{vector1}{a vector containing the first set of values}

\item{vector2}{a vector containing the second set of values}
}
\value{
a new vector containing the fold changes of the original values.
}
\description{
This function takes two vectors and calculates the fold changes between them.
}
\examples{

}
\author{
Brendan Gongol
}
