% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tests.R
\name{fetch_tests}
\alias{fetch_tests}
\title{Fetch tests from the database}
\usage{
fetch_tests(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
A data frame containing the fetched tests.
}
\description{
This function executes a query to fetch the tests from the database.
}
