% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{file_sorteR}
\alias{file_sorteR}
\title{file_sortR a file sorting function}
\usage{
file_sorteR(directory, newdirectory, files, Name_start, Name_end,
  remove_old = FALSE)
}
\arguments{
\item{directory}{path to the directory storing the files of interest. Must not have any spaces in the file path.}

\item{newdirectory}{a path containing the name of the new directory to make. Must not have any spaces in the file path.}

\item{files}{a numerical string indicating what files in the directory to move}

\item{Name_start}{a number indicating the start position of the file name to create a directory for.}

\item{Name_end}{a numer indicating the end position of the file name to create a directory for.}

\item{remove_old}{a TRUE/FALSE logical.  If TRUE, will delete the old files.}
}
\value{
a new directory containing sub directories housing original files.
}
\description{
This function takes a directory of your choosing that contains everal files with similar names and
creates new directory containing subdirectories containing files named according to the subsetted file name specified,
and copies the files into the new directories.
}
\examples{
file_sorteR(directory = "C:/Users/Brendan/Desktop/AMPKKO2_1", newdirectory = "C:/Users/Brendan/Desktop/AMPK_practice",
files = 2:343, Name_start = 14, Name_end = 19, remove_old = TRUE)
}
\author{
Brendan Gongol
}
