% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_annotation.R
\name{filtered_annotation}
\alias{filtered_annotation}
\title{Test annotation filtered}
\usage{
filtered_annotation(annotations, x_min, x_max)
}
\arguments{
\item{annotations}{result of get_annotation}

\item{x_min}{ggpubr::geom_signif 参数xmin}

\item{x_max}{ggpubr::geom_signif 参数xmax}
}
\value{
A list with filtered annotation,xmin and xmax
}
\description{
This function performs filtered annotation
}
\author{
Pengmin Yang
}
