% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from the fire station (miles)}
  \item{DAMAGE}{Damage cost in 1000 dollars}
}
}
\source{
Example dataset
}
\usage{
fire
}
\description{
A dataset containing fire damage statistics.
}
\keyword{datasets}
