% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{firedam}
\alias{firedam}
\title{A dataset containing distances and damages}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{distance}{Distince}
\item{Damage}{Damage}
...
}
}
\description{
A dataset containing distances and damages
}
