% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireData.R
\docType{data}
\name{fireData}
\alias{fireData}
\title{Fire damage data set}
\format{
A data frame with 15 rows and 5 variables:
\describe{
  \item{DISTANCE}{distance from firestation in miles}
  \item{DAMAGE}{fire damage in thousands of dollars}
  ...
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fireData
}
\description{
A data set on a study into the amount of fire damage in major residential fires to the distance
between the residence and the nearest fire station
}
\keyword{datasets}
