% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_model}
\alias{fit_model}
\title{Easily filter unemployment data for Luxembourg}
\usage{
fit_model(data)
}
\arguments{
\item{dataA}{predicate condition indicating the regional level of interest. See details for more.}
}
\value{
A data frame
}
\description{
Easily filter unemployment data for Luxembourg
}
\details{
This function allows the user to create a data frame for several regional levels. The first level
is the complete country. The second level are cantons, and the third level are neither cantons
nor the whole country, so the communes. Individual communes can be selected as well.
`level` must be predicate condition passed down to dplyr::filter. See the examples below
for its usage.
}
\examples{
# basic function
get_data <- function(file) {
  aq <- data.frame(diamonds) \%>\%
  filter(!is.na(price))
}

fit_model <- function(data) {
lm(price ~ carat, data) \%>\%
coefficients()
}
}
