% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_area_circle.R
\name{fn_area_circle}
\alias{fn_area_circle}
\title{Calculate the Area of the Circle}
\usage{
fn_area_circle(r)
}
\arguments{
\item{r}{radius}
}
\value{
Area of the Circle
}
\description{
Calculate the Area of the Circle
}
