% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnAreaCircle.R
\name{fnAreaCircle}
\alias{fnAreaCircle}
\title{Calculate area of circle}
\usage{
fnAreaCircle(r)
}
\arguments{
\item{r}{radius of the circle}
}
\description{
Given a colour make n lighter or darker shades
}
