% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{Ftest}
\alias{Ftest}
\title{Conduct a Ftest for SLR under H0:beta1=0}
\usage{
Ftest(X, Y)
}
\arguments{
\item{X}{a vector}

\item{Y}{a vector}
}
\value{
Ftest value
}
\description{
Conduct a Ftest for SLR under H0:beta1=0
}
\examples{
data(mtcars)
attach(mtcars)
Ftest(mpg,wt)
detach(mtcars)
}
