% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{fun_function}
\alias{fun_function}
\title{A Fun Function}
\usage{
fun_function(x)
}
\arguments{
\item{x}{Input}
}
\description{
This function allows you to find square of every input.
}
\examples{
fun_function(5)
}
\keyword{x}
