% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_Newton.R
\name{fx_hp}
\alias{fx_hp}
\title{Objective Function for HP-filtering}
\usage{
fx_hp(y, theta, Dkn, lambda = 0)
}
\arguments{
\item{y}{response}

\item{theta}{regression coefficient vector}

\item{Dkn}{sparse differencing matrix}

\item{lambda}{regularization parameter}
}
\description{
Objective Function for HP-filtering
}
