% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssx-internal.R
\name{gamlssx-internal}
\alias{gamlssx-internal}
\title{Internal gamlssx functions}
\description{
Internal gamlssx functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
