% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gamlssx_methods}
\alias{gamlssx_methods}
\alias{print.gamlssx}
\alias{summary.gamlssx}
\title{Methods for objects of class \code{"gamlssx"}}
\usage{
\method{print}{gamlssx}(x, ...)

\method{summary}{gamlssx}(object, ...)
}
\arguments{
\item{x, object}{An object inheriting from class \code{"gamlssx"}, a result of a
call to \code{\link{fitThresholdGP}}.}

\item{...}{Additional arguments.}
}
\value{
See \code{\link[gamlss:print.gamlss]{gamlss::print.gamlss}} and \code{\link[gamlss:summary.gamlss]{gamlss::summary.gamlss}}.
}
\description{
Methods for objects of class \code{"gamlssx"} returned from \code{\link{fitThresholdGP}}.
}
\details{
These \code{print} and \code{summary} methods have been created to avoid
printing to the console a huge statement of the call to the function
\code{gamlss::gamlss}. Instead, the call to \code{fitThresholdGP} is printed.
}
