% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapNum.R
\name{gapNum}
\alias{gapNum}
\title{Detect the time gap more than 5 min}
\usage{
gapNum(individualSubset)
}
\arguments{
\item{individualSubset}{A subset of data for each individual}
}
\value{
The gaps in time series
}
\description{
Detect the time gap more than 5 min
}
