% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowGap.R
\name{gapRow}
\alias{gapRow}
\title{Detect the time gap more than 5 min}
\usage{
gapRow(individualDataset)
}
\arguments{
\item{inputId}{A string of he unique identifier}
}
\value{
The gaps in time series
}
\description{
Detect the time gap more than 5 min
}
