% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.R, R/gapSize.R
\name{gapSize}
\alias{gapSize}
\title{Detect the time gap more than 5 min}
\usage{
gapSize(individualDataset)

gapSize(individualDataset)
}
\arguments{
\item{individualDataset}{An individual dataset}

\item{inputId}{A string of he unique identifier}
}
\value{
The gaps in time series

The gaps in time series
}
\description{
Detect the time gap more than 5 min

Detect the time gap more than 5 min
}
