\name{gegen}
\alias{gegen}
\title{Geometric Distribution}
\usage{
gengen(p)
}
\description{
This function generates a random number from Geometric Distribution i.e. number of Bernoulli trials before reaching first seccuss.
}
\arguments{
  \item{p}{probability of seccuss on each trial}
}
\references{
  dr.gharib and mr.heidari
}
\examples{
brgen(0.2)
}
