% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.mut_exp.R
\name{gene.mut_exp}
\alias{gene.mut_exp}
\title{Gene Mutation and Expression Data Visualization}
\usage{
gene.mut_exp(
  mutation_data,
  exp,
  colnum = 3,
  exp_cleaning = FALSE,
  gene,
  top = 10,
  visual = TRUE,
  test_type = "parametric",
  title = "CRC TCGA",
  test = "wilcox.test",
  only_red = TRUE,
  gene_vaf = F,
  color_0.05 = "#FF34B3",
  color_0.01 = "#9400D3",
  color_0.001 = "#CD3700",
  color_0.0001 = "red",
  bar = F
)
}
\arguments{
\item{mutation_data}{A data frame containing SNV data.即MAF格式的数据}

\item{exp}{is the data of gene expression for samples}

\item{colnum}{exp 列名为 sample开始的列数}

\item{exp_cleaning}{是否对数据进行清洗，默认FALSE}

\item{gene}{gene name,gene=c("PREX2","TP53","APC")}

\item{top}{突变率前x的基因, top=10}

\item{visual}{可视化否，默认TURE}

\item{test_type}{ggstatsplot::ggbetweenstats,test_type= "parametric"}

\item{title}{ggstatsplot::ggbetweenstats,title = NULL}

\item{test}{参考ggpubr::compare_means，根据需要选择}

\item{only_red}{axis.text.x颜色，根据显著性显示}

\item{gene_vaf}{gene_vaf defalut is FALSE}

\item{bar}{defalut is FALSE}

\item{color_0.05/color_0.01/color_0.001/color_0.0001}{axis.text.x颜色，根据不同显著性显示}
}
\value{
A list of Gene Expression Difference with gene mutated status and visualizing
}
\description{
This function visualizes the  Gene Expression Difference with gene mutated status.
}
\examples{
data("exp_raw")
data("mutation_CRC")
exp_CRC<-exp_geneIDtoSYMBOL(exp=exp_raw,genecoltype="ENTREZID")
exp_CRC<-exp_CRC$data
result<-gene.mut_exp(mutation_data=mutation_CRC,exp=exp_CRC[,-1],colnum=2,gene=NULL,top=10,visual=TRUE,
            test_type= "parametric",title= "CRC TCGA",test="wilcox.test",only_red=TRUE,gene_vaf=F,
             color_0.05= "#FF34B3",color_0.01="#9400D3",color_0.001="#CD3700",color_0.0001="red",bar=T)
}
\author{
Pengmin Yang
}
