% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFilter.R
\name{gene_filter}
\alias{gene_filter}
\title{Gene Filter}
\usage{
gene_filter(data, cutoff = 0.1, percent = 0.8)
}
\arguments{
\item{data}{Data matrix}

\item{cutoff}{The cutoff defined the expressed gene Defaults to 0.1}

\item{percent}{The percentage of total sample used to set cutoff Defaults to 0.8}
}
\value{
Filtered data matrix
}
\description{
This function perform gene filter based on expression matrix.
}
\examples{
gene_filter(data, cutoff=0.1, percent=0.8)
}
\author{
WKL
}
\keyword{filter}
