% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfamily.R
\name{GenPareto}
\alias{GenPareto}
\alias{GP}
\alias{GPfisher}
\alias{GPquasi}
\alias{dGenPareto}
\alias{pGenPareto}
\alias{qGenPareto}
\alias{rGenPareto}
\title{GP family distribution for fitting a GAMLSS}
\usage{
GPfisher(mu.link = "log", sigma.link = "identity")

GPquasi(mu.link = "log", sigma.link = "identity")

dGenPareto(x, mu = 1, sigma = 0, log = FALSE)

pGenPareto(q, mu = 1, sigma = 0, lower.tail = TRUE, log.p = FALSE)

qGenPareto(p, mu = 1, sigma = 0, lower.tail = TRUE, log.p = FALSE)

rGenPareto(n, mu = 1, sigma = 0)
}
\arguments{
\item{mu.link}{Defines the \code{mu.link}, with \code{"log"} link as the default
for the \code{mu} parameter.}

\item{sigma.link}{Defines the \code{sigma.link}, with \code{"identity"} link as the
default for the \code{sigma} parameter.}

\item{x, q}{Vector of quantiles.}

\item{mu, sigma}{Vectors of scale and shape parameter values.}

\item{log, log.p}{Logical. If \code{TRUE}, probabilities \verb{eqn\{p\}} are given as
\eqn{\log(p)}.}

\item{lower.tail}{Logical. If \code{TRUE} (the default), probabilities are
\eqn{P[X \leq x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken to
be the number required.}
}
\value{
\code{GPfisher()} and \code{GPquasi()} each return a
\code{\link[gamlss.dist:gamlss.family]{gamlss.dist::gamlss.family()}} object which can be used to fit a
regression model with a GP response distribution using the
\code{\link[gamlss:gamlss]{gamlss::gamlss()}}function. \code{dGenPareto()} gives the density,
\code{pGenPareto()} gives the distribution function, \code{qGenPareto()} gives the
quantile function, and \code{rGenPareto()} generates random deviates.
}
\description{
The functions \code{GPfisher()} and \code{GPquasi()} each define the generalized
Pareto (GP) family distribution, a two parameter distribution, for
a \code{\link[gamlss.dist:gamlss.family]{gamlss.dist::gamlss.family()}} object to be used in GAMLSS fitting using
the function \code{\link[gamlss:gamlss]{gamlss::gamlss}}. The only difference between \code{GPfisher()}
and \code{GPquasi()} is the form of scoring method used to define the weights
used in the fitting algorithm. Fisher's scoring, based on the expected
Fisher information is used in \code{GPfisher()}, whereas a quasi-Newton scoring,
based on the cross products of the first derivatives of the log-likelihood,
is used in \code{GPquasi()}. The functions \code{dGenPareto} \code{pGenPareto},
\code{qGenPareto} and \code{rGenPareto} define the density, distribution function,
quantile function and random generation for the specific parameterization of
the generalized Pareto distribution given in \strong{Details} below.
}
\details{
The distribution function of a GP distribution with parameters
\code{scale} = \eqn{\mu (> 0)} and \code{shape} = \eqn{\xi} (\eqn{= \nu}) is
\deqn{F(x) = P(X \leq x) = 1 - \left( 1 + \frac{\xi x}{\mu} \right)^{-1/\xi}.}
If \eqn{\xi = 0} the Generalized Pareto distribution is equivalent to an
Exponential distribution with parameter \eqn{1/\mu}.

The support of the distribution depends on \eqn{\xi}: it is
\eqn{x \geq 0}{x >= 0} for \eqn{\xi \geq 0};
\eqn{0 \leq x \leq - \mu / \xi}{0 <= x <= - \mu / \xi} for \eqn{\xi < 0}{\xi < 0}.
See
\url{https://en.wikipedia.org/wiki/Generalized_Pareto_distribution}
and/or Chapter 4 of Coles (2001) for further information.

For each observation in the data, the restriction that \eqn{\xi > -1/2} is
imposed, which is necessary for the usual asymptotic likelihood theory to be
applicable.
}
\section{Examples}{

See the examples in \code{\link{fitGP}}.
}

\references{
Coles, S. G. (2001) \emph{An Introduction to Statistical
Modeling of Extreme Values}, Springer-Verlag, London.
Chapter 4: \doi{10.1007/978-1-4471-3675-0_4}
}
\seealso{
\code{\link{fitGP}},
\code{\link[gamlss.dist:gamlss.family]{gamlss.dist::gamlss.family()}},
\code{\link[gamlss:gamlss]{gamlss::gamlss()}}
}
