% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation.R
\name{get_annotation}
\alias{get_annotation}
\title{Test annotation get}
\usage{
get_annotation(data, group, types, test = "wilcox.test")
}
\arguments{
\item{data}{长数据格式，最后一列为值}

\item{group}{分组的名称}

\item{types}{A character vector indicating the type.例如c("Wild", "Mut")}

\item{test}{参考ggpubr::compare_means，根据需要选择}
}
\value{
A list with test annotation
}
\description{
This function performs statistical difference testing between every two groups
}
\author{
Pengmin Yang
}
