% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rscu.R
\name{get_rscu}
\alias{get_rscu}
\title{Calculate the RSCU of sequence(s)}
\usage{
get_rscu(seq, seq_type = c("DNA", "RNA"), as.data.frame = TRUE)
}
\arguments{
\item{seq}{Character vector, DNAString, RNAString, DNAStringSet, or RNAStringSet of sequences}

\item{seq_type}{Character value of either "DNA" or "RNA" depending on whether
"T" or "U" is used in the sequence.}

\item{as.data.frame}{Logical value determining if the returned result should be a data.frame or
a tibble. Defaults to TRUE.}
}
\value{
A data.frame or a tibble with codons as column names, sequences as rows, and the RSCUs
as the values.
}
\description{
Calculate the RSCU of sequence(s)
}
\examples{
seq = paste0("ATGGCGTCCCCGTCTCGGAGACTGCAGACTAAACCAGTCATTACTTGTTTC",
"AAGAGCGTTCTGCTAATCTACACTTTTATTTTCTGGATCACTGGCGTTATCCTTCTTG",
"CAGTTGGCATTTGGGGCAAGGTGAGCCTGGAGAATTACTTTTCTCTTTTAAATGAGAA",
"GGCCACCAATGTCCCCTTCGTGCTCATTGCTACTGGTACCGTCATTATTCTTTTGGGC",
"ACCTTTGGTTGTTTTGCTACCTGCCGAGCTTCTGCATGGATGCTAAAACTGTATGCAA",
"TGTTTCTGACTCTCGTTTTTTTGGTCGAACTGGTCGCTGCCATCGTAGGATTTGTTTT",
"CAGACATGAGATTAAGAACAGCTTTAAGAATAATTATGAGAAGGCTTTGAAGCAGTAT",
"AACTCTACAGGAGATTATAGAAGCCATGCAGTAGACAAGATCCAAAATACGTTGCATT",
"GTTGTGGTGTCACCGATTATAGAGATTGGACAGATACTAATTATTACTCAGAAAAAGG",
"ATTTCCTAAGAGTTGCTGTAAACTTGAAGATTGTACTCCACAGAGAGATGCAGACAAA",
"GTAAACAATGAAGGTTGTTTTATAAAGGTGATGACCATTATAGAGTCAGAAATGGGAG",
"TCGTTGCAGGAATTTCCTTTGGAGTTGCTTGCTTCCAACTGATTGGAATCTTTCTCGC",
"CTACTGCCTCTCTCGTGCCATAACAAATAACCAGTATGAGATAGTGTAA")

get_rscu(seq, seq_type = 'DNA')
get_rscu(Biostrings::DNAString(seq), seq_type = 'DNA')
get_rscu(seq, as.data.frame = FALSE, seq_type = 'DNA')

seqs = c(paste0("ATGGCGTCCCCGTCTCGGAGACTGCAGACTAAACCAGTCATTACTTGTT",
"TCAAGAGCGTTCTGCTAATCTACACTTTTATTTTCTGGATCACTGGCGTTATCCTTCTT",
"GCAGTTGGCATTTGGGGCAAGGTGAGCCTGGAGAATTACTTTTCTCTTTTAAATGAGAA",
"GGCCACCAATGTCCCCTTCGTGCTCATTGCTACTGGTACCGTCATTATTCTTTTGGGCA",
"CCTTTGGTTGTTTTGCTACCTGCCGAGCTTCTGCATGGATGCTAAAACTGTATGCAATG",
"TTTCTGACTCTCGTTTTTTTGGTCGAACTGGTCGCTGCCATCGTAGGATTTGTTTTCAG",
"ACATGAGATTAAGAACAGCTTTAAGAATAATTATGAGAAGGCTTTGAAGCAGTATAACT",
"CTACAGGAGATTATAGAAGCCATGCAGTAGACAAGATCCAAAATACGTTGCATTGTTGT",
"GGTGTCACCGATTATAGAGATTGGACAGATACTAATTATTACTCAGAAAAAGGATTTCC",
"TAAGAGTTGCTGTAAACTTGAAGATTGTACTCCACAGAGAGATGCAGACAAAGTAAACA",
"ATGAAGGTTGTTTTATAAAGGTGATGACCATTATAGAGTCAGAAATGGGAGTCGTTGCA",
"GGAATTTCCTTTGGAGTTGCTTGCTTCCAACTGATTGGAATCTTTCTCGCCTACTGCCT",
"CTCTCGTGCCATAACAAATAACCAGTATGAGATAGTGTAA"),
paste0("ATGGCCTCCTTGGAAGTCAGTCGTAGTCCTCGCAGGTCTCGGCGGGAGCTGGAAGTGC",
"GCAGTCCACGACAGAACAAATATTCGGTGCTTTTACCTACCTACAACGAGCGCGAGAAC",
"CTGCCGCTCATCGTGTGGCTGCTGGTGAAAAGCTTCTCCGAGAGTGGAATCAACTATGA",
"AATTATAATCATAGATGATGGAAGCCCAGATGGAACAAGGGATGTTGCTGAACAGTTGG",
"AGAAGATCTATGGGTCAGACAGAATTCTTCTAAGACCACGAGAGAAAAAGTTGGGACTA",
"GGAACTGCATATATTCATGGAATGAAACATGCCACAGGAAACTACATCATTATTATGGA",
"TGCTGATCTCTCACACCATCCAAAATTTATTCCTGAATTTATTAGGAAGCAAAAGGAGG",
"GTAATTTTGATATTGTCTCTGGAACTCGCTACAAAGGAAATGGAGGTGTATATGGCTGG",
"GATTTGAAAAGAAAAATAATCAGTGATGGAGTTTTGCCATGTTGCCCAGGCTGGTCACT",
"CCTGGGCTCAAGTGATCCAGCCATCTTGGCCTCCTGGGATTACAGCCGTGGGGCCAATT",
"TTTTAACTCAGATCTTGCTGAGACCAGGAGCATCTGATTTAACAGGAAGTTTCAGATTA",
"TACCGAAAAGAAGTTCTAGAGAAATTAATAGAAAAATGTGTTTCTAAAGGCTACGTCTT",
"CCAGATGGAGATGATTGTTCGGGCAAGACAGTTGAATTATACTATTGGCGAGGTTCCAA",
"TATCATTTGTGGATCGTGTTTATGGTGAATCCAAGTTGGGAGGAAATGAAATAGTATCT",
"TTCTTGAAAGGATTATTGACTCTTTTTGCTACTACATAA"))

get_rscu(seqs, seq_type = 'DNA')
}
