% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHello.R
\name{getHello}
\alias{getHello}
\title{getHello Function}
\usage{
getHello()
}
\description{
print "Hello World!" to the std
}
\examples{
getHello()
}
\references{
\url{https://lin-chen-va.github.io/}
}
\seealso{
\code{\link{getRand}}
}
\author{
Lin Chen, \email{lin.chen@ieee.org}
}
\keyword{print}
