% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHightlyCorDescr.R
\name{getHightlyCorDescr}
\alias{getHightlyCorDescr}
\title{相関が強い列名を取得}
\usage{
getHightlyCorDescr(DATA, CUTOFF = 0.75)
}
\arguments{
\item{DATA}{data.frame 学習用データ}

\item{CUTOFF}{numeric ( default 0.75 )}
}
\value{
vector 相関が強い列名一覧
}
\description{
相関が強い列名を取得
}
