% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearZeroVar.R
\name{getNearZeroVar}
\alias{getNearZeroVar}
\title{情報量が少ない ( 分散がほぼ 0 ) 列名を取得}
\usage{
getNearZeroVar(DATA)
}
\arguments{
\item{DATA}{data.frame 学習用データ}
}
\value{
vector 情報量が少ない ( 分散がほぼ 0 ) 列名一覧
}
\description{
情報量が少ない ( 分散がほぼ 0 ) 列名を取得
}
