% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRand.R
\name{getRand}
\alias{getRand}
\title{getRand Function}
\usage{
getRand(n)
}
\arguments{
\item{n}{the number of random numbers}
}
\description{
get a sequence of random numbers between 0 and 1
}
\examples{
v = getRand(10)
}
\references{
\url{https://lin-chen-va.github.io/}
}
\author{
Lin Chen, \email{lin.chen@ieee.org}
}
\keyword{rand}
