% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRand.R
\name{getRandN}
\alias{getRandN}
\title{getRandN Function}
\usage{
getRandN(n, s)
}
\arguments{
\item{n}{the number of random numbers}

\item{s}{margin value}
}
\description{
get a sequence of random numbers between 0 and N
}
\examples{
v = getRandN(10, 2)
}
\keyword{rand}
