% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvplots.R
\name{ggsurvplots}
\alias{ggsurvplots}
\title{Drawing Survival Curves Using ggplot2}
\usage{
ggsurvplots(
  data,
  conf.int = FALSE,
  time_col = "OS_MONTHS",
  status_col = "OS_SATUTS",
  group_col = "Risk",
  pvalue_table = F,
  palette = NULL,
  risk.table = T,
  title = "MSk cohort",
  legend.labs = c("High-risk", "Low-risk"),
  xlab = "OS_MONTHS",
  ylab = "Survial probability",
  surv.median.line = "none",
  surv.scale = "default",
  legend = TRUE
)
}
\arguments{
\item{data}{a dataset used to fit survival curves. If not supplied then data will be extracted from 'fit' object.}

\item{time_col}{the colname of time}

\item{status_col}{the colname of status}

\item{group_col}{the colname of sample group}

\item{pvalue_table}{table of Log-rank p test of samples}

\item{palette}{the color palette to be used. Allowed values include "hue" for the default hue color scale;
"grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...;
or custom color palette e.g. c("blue", "red");
and scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".
See details section for more information. Can be also a numeric vector of length(groups);
in this case a basic color palette is created using the function palette.}

\item{risk.table}{Allowed values include:
 TRUE or FALSE specifying whether to show or not the risk table. Default is FALSE.
"absolute" or "percentage". Shows the absolute number and the percentage of subjects at risk by time, respectively.
"abs_pct" to show both absolute number and percentage.
"nrisk_cumcensor" and "nrisk_cumevents". Show the number at risk and, the cumulative number of censoring and events, respectively.}

\item{title}{main title.}

\item{legend.labs}{character vector specifying legend labels. currently max length is 20.
Used to replace the names of the strata from the fit. Should be given in the same order as those strata.}

\item{xlab, ylab}{xlab, ylab: x and y axis labels, respectively.}

\item{surv.median.line}{character vector for drawing a horizontal/vertical line at median survival. Allowed values include one of c("none", "hv", "h", "v"). v: vertical, h:horizontal.}

\item{surv.scale}{surv.scale: scale transformation of survival curves. Allowed values are "default" or "percent".}

\item{legend}{lenged show or not, Default is TRUE.}
}
\value{
A ggplot object.
}
\description{
Drawing Survival Curves Using ggplot2
}
\examples{
data("clin_TCGA")
p <- ggsurvplots(data = clin_TCGA, conf.int = FALSE,time_col = "PFS_MONTHS",
                 status_col = "PFS_STATUS", group_col = "Status", pvalue_table = TRUE,
                 palette = ggsci::pal_ucscgb()(4), risk.table = FALSE, title = NULL,
                 legend.labs = c("no KRAS or TP53", "TP53", "KRAS", "KRAS&TP53"),
                 xlab = "PFS_MONTHS", ylab = "Survival probability",
                 surv.median.line="hv",surv.scale="default",legend=FALSE)
print(p)

}
\author{
Pengmin Yang
}
