% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_of_life.R
\name{go_to}
\alias{go_to}
\title{Change working directory}
\usage{
go_to(folder)
}
\arguments{
\item{folder}{Destination folder}
}
\value{

}
\description{
This function will change the working directory to the specified folder. It
assumes that the destination folder is directly reachable from either the
current working directory or one of its parent directories
}
