% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{GOsorteR}
\alias{GOsorteR}
\title{GOsorteR}
\usage{
GOsorteR(DT, terms)
}
\arguments{
\item{DT}{a data table containing a column labeled "mergecol" housing the characters to query.}

\item{terms}{a vector of character strings indicating the terms to query "mergecol" for.}
}
\value{
The data table that is subsetted according to the search terms listed in the "terms" argument.
}
\description{
Requires a data.table with a column labeled "mergecol" and a vector of query terms
}
\examples{
library(data.table)
gene_symbol <- c("AKT", "PI3K", "SREBP", "FOXO", "ABCA1", "Caspase-1", "PIGPEN", "SNAIL")
chromosome_name <- c("1", "5", "10", "X", "Y", "2", "3", "4")
go_id <- c("GO:0000001", "GO:0000002", "GO:0000003", "GO:0000004", "GO:0000005", "GO:0000006", "GO:0000007", "GO:0000008")
name_1006 <- c("biological_process", "cholesterol", "apoptosis", "differentiation", "Mitochondria Function", "cellular differentiation", "inflammation", "cell cycle")
definition_1006 <- c("Any process specifically pertinent to the functioning of integrated living units",
                     "Any process specifically pertinent to the functioning of cholesterol",
                     "Any process specifically pertinent to the functioning of apoptosis",
                     "Any process specifically pertinent to the functioning of differentiation",
                     "Any process specifically pertinent to the functioning of mitochondrial function",
                     "Any process specifically pertinent to the functioning of cellular differentiation",
                     "Any process specifically pertinent to the functioning of inflammation",
                     "Any process specifically pertinent to the functioning of cell cycle")
data <- data.table(cbind(gene_symbol, chromosome_name, go_id, name_1006, definition_1006))
data[, mergecol := paste(go_id,name_1006,definition_1006,sep="@")]

words <- c("cholesterol", "differentiation", "functioning of mitochondria")

GOsorteR(DT = data, terms = words)
}
\author{
Brendan Gongol
}
