% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPfamily.R
\name{GP}
\alias{GP}
\alias{GPfisher}
\alias{GPquasi}
\alias{dGP}
\alias{pGP}
\alias{qGP}
\alias{rGP}
\title{GP family distribution for fitting a GAMLSS}
\usage{
GPfisher(mu.link = "log", sigma.link = "identity")

GPquasi(mu.link = "log", sigma.link = "identity")

dGP(x, mu = 1, sigma = 0, log = FALSE)

pGP(q, mu = 1, sigma = 0, lower.tail = TRUE, log.p = FALSE)

qGP(p, mu = 1, sigma = 0, lower.tail = TRUE, log.p = FALSE)

rGP(n, mu = 1, sigma = 0)
}
\arguments{
\item{mu.link}{Defines the \code{mu.link}, with \code{"log"} link as the default
for the \code{mu} parameter.}

\item{sigma.link}{Defines the \code{sigma.link}, with \code{"identity"} link as the default
for the \code{sigma} parameter.}

\item{x, q}{Vector of quantiles.}

\item{mu, sigma}{Vectors of scale and shape parameter values.}

\item{log, log.p}{Logical. If \code{TRUE}, probabilities \verb{eqn\{p\}} are given as
\eqn{\log(p)}.}

\item{lower.tail}{Logical. If \code{TRUE} (the default), probabilities are
\eqn{P[X \leq x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken to
be the number required.}
}
\value{
\code{GPfisher()} and \code{GPquasi()} each return a
\code{\link[gamlss.dist:gamlss.family]{gamlss.dist::gamlss.family()}} object
which can be used to fit a regression model with a GP response
distribution using the
\code{\link[gamlss:gamlss]{gamlss::gamlss()}} function. \code{dGP()} gives the density,
\code{pGP()} gives the distribution function, \code{qGP()} gives the quantile
function, and \code{rGP()} generates random deviates.
}
\description{
The functions \code{GPfisher()} and \code{GPquasi()} each define the generalized
Pareto (GP) family distribution, a two parameter distribution, for
a \code{\link[gamlss.dist:gamlss.family]{gamlss.dist::gamlss.family()}} object to
be used in GAMLSS fitting using the function
\code{\link[gamlss:gamlss]{gamlss::gamlss()}}. The only difference
between \code{GPfisher()} and \code{GPquasi()} is the form of scoring method used to
define the weights used in the fitting algorithm. Fisher's scoring,
based on the expected Fisher information is used in \code{GPfisher()}, whereas
a quasi-Newton scoring, based on the cross products of the first derivatives
of the log-likelihood, is used in \code{GPquasi()}. The functions
\code{dGP}, \code{pGP}, \code{qGP} and \code{rGP} define the density, distribution function,
quantile function and random generation for the specific parameterization of
the generalized extreme value distribution given in \strong{Details} below.
}
\details{
For the purposes of using the gamlss package, the two parameters need to
be called \eqn{\mu} and \eqn{\sigma} for scale and shape respectively.
However, in mathematical notation the standard \eqn{\sigma} and \eqn{\xi} notation will be used.

The distribution function of a GP distribution with parameters
\code{scale} = \eqn{\sigma (> 0)} and \code{shape} = \eqn{\xi} (\eqn{= \nu}) is
\deqn{F(x) = P(X \leq x) = 1 - \left( 1 + \frac{\xi x}{\sigma} \right)^{-1/\xi}.}
If \eqn{\xi = 0} the Generalized Pareto distribution is equivalent to an
Exponential distribution with parameter \eqn{1/\sigma}.

The support of the distribution depends on \eqn{\xi}: it is
\eqn{x \geq 0}{x >= 0} for \eqn{\xi \geq 0};
\eqn{0 \leq x \leq - \sigma / \xi}{0 <= x <= - \sigma / \xi} for \eqn{\xi < 0}{\xi < 0}.
See
\url{https://en.wikipedia.org/wiki/Generalized_Pareto_distribution}
and/or Chapter 4 of Coles (2001) for further information.

For each observation in the data, the restriction that \eqn{\xi > -1/2} is
imposed, which is necessary for the usual asymptotic likelihood theory to be
applicable.
}
\section{Examples}{

See the examples in \code{\link{fitGP}}.
}

\references{
Coles, S. G. (2001) \emph{An Introduction to Statistical
Modeling of Extreme Values}, Springer-Verlag, London.
Chapter 4: \doi{10.1007/978-1-4471-3675-0_4}
}
\seealso{
\code{\link{fitGP}},
\code{\link[gamlss.dist:gamlss.family]{gamlss.dist::gamlss.family()}},
\code{\link[gamlss:gamlss]{gamlss::gamlss()}}
}
