% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpExpInfo.R
\name{gpExpInfo}
\alias{gpExpInfo}
\title{GP Distribution Expected Information}
\usage{
gpExpInfo(scale, shape)
}
\arguments{
\item{scale, shape}{Numeric vectors. Respective values of the GP parameters
scale parameter \eqn{\sigma} and shape parameter \eqn{\xi}. For
\code{gpExpInfo}, \code{scale} and \code{shape} must have length 1.}
}
\value{
\code{gpExpInfo} returns a 2 by 2 numeric matrix with row and column
named \verb{scale, shape}.
}
\description{
Calculates the expected information matrix for the GP distribution.
}
\details{
\code{gpExpInfo} calculates, for a single pair of values
\eqn{(\sigma, \xi) = } \verb{(scale, shape)}, the expected information matrix for a
single observation from a GP distribution with distribution function
\deqn{F(x) = P(X \leq x) = 1 - \left[ 1 + \displaystyle\frac{\xi x}{\sigma} \right]_+^{-1/\xi}}
where \eqn{x_+ = \max(x, 0)}.
The GP expected information is defined only for \eqn{\xi > -0.5}.

The matrix is of the form
\deqn{\mathbf{I} = \displaystyle\frac{1}{1 + 2 \xi} \begin{bmatrix}
  \displaystyle\frac{1}{\sigma^2} & \displaystyle\frac{1}{\sigma (1 + \xi)} \\
  \displaystyle\frac{1}{\sigma (1 + \xi)} & \displaystyle\frac{2}{1 + \xi}
  \end{bmatrix}.}
}
\examples{
# Expected information matrices for ...
# ... scale = 1 and shape = 0.1
gpExpInfo(1, 0.1)
}
