% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_Newton.R
\name{gradf_hp}
\alias{gradf_hp}
\title{Gradient for HP-filtering}
\usage{
gradf_hp(y, theta, Dkn, lambda = 0)
}
\arguments{
\item{y}{response}

\item{theta}{regression coefficient vector}

\item{Dkn}{sparse differencing matrix}

\item{lambda}{regularization parameter}
}
\description{
Gradient for HP-filtering
}
