% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_Newton.R
\name{gradient_descent_fixed}
\alias{gradient_descent_fixed}
\title{Gradient Descent (Fixed Step-Size)}
\usage{
gradient_descent_fixed(fx, gradf, x0, t, max_iter = 100, tol = 1e-05)
}
\arguments{
\item{fx}{handle to function that returns objective function values}

\item{gradf}{handle to function that returns gradient of objective function}

\item{x0}{initial parameter estimate}

\item{t}{step-size}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Gradient Descent (Fixed Step-Size)
}
