% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient_Newton.R
\name{gradient_step}
\alias{gradient_step}
\title{Gradient Step}
\usage{
gradient_step(gradf, x, t)
}
\arguments{
\item{gradf}{handle to function that returns gradient of objective function}

\item{x}{current parameter estimate}

\item{t}{step-size}
}
\description{
Gradient Step
}
