% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{greedy}
\alias{greedy}
\title{VI or Binder loss according to Wade Gahramani algorithm}
\usage{
greedy(
  PYclust,
  partition = PYclust[1, ],
  l = 2 * ncol(PYclust),
  maxiter = 1000,
  distance = "VI"
)
}
\description{
Computes the partition with the least expected loss through a greedy algorithm
}
