% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet R}
\usage{
greet(name, loud = FALSE)
}
\arguments{
\item{name}{A character string. The name to greet.}

\item{loud}{Logical. Upper or lower case. Default is FALSE.}
}
\value{
A character string containing a personal greeting
}
\description{
Greet R
}
\examples{
greet("Ada")
greet("Ada", loud = TRUE)
}
