% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAplot2.R
\name{GSEAplot2}
\alias{GSEAplot2}
\title{GSEAplot2}
\usage{
GSEAplot2(
  x,
  geneSetID = c(1:10),
  title = "",
  color = (ggsci::pal_gsea())(10),
  base_size = 11,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  legend.position = "right",
  ES_geom = "line",
  pvalue_table = TRUE,
  Type
)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{title}{plot title}

\item{color}{color of running enrichment score line}

\item{base_size}{base font size}

\item{rel_heights}{relative heights of subplots}

\item{subplots}{which subplots to be displayed}

\item{legend.position}{legend.position=c("top","bottom","right","left","none")}

\item{ES_geom}{geom for plotting running enrichment score,one of 'line' or 'dot'}

\item{pvalue_table}{whether add pvalue table, defalut pvalue_table=TRUE}

\item{Type}{A character vector indicating the type.Type can length from 2 to 5,eg Type=c("Wild", "Mut");(default = Type[[1]] is the control)}
}
\value{
plot
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
\author{
Pengmin Yang
}
