% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.div.R
\name{hap.div}
\alias{hap.div}
\title{Haplotype diversity}
\usage{
hap.div(dnabin, pop)
}
\arguments{
\item{DNAbin}{object.}
}
\value{
Haplotype diversity per population
}
\description{
The function returns the haplotype diversity per population.
}
\details{
The function takes an advantage of other functions from package pegas. Pegas calculates number of haplotypes and haplotype frequency.
This function tries to summarize haplotype diversity by population if specified.
}
