% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{heart}
\alias{heart}
\title{Heart Function}
\usage{
heart(title = "")
}
\arguments{
\item{title}{a string: you can choose a blessing phrase as the title; the default is NULL.}
}
\value{
Draw a heart plot(it may take 20 seconds.)
}
\description{
Draw a heart on the axis.
}
\examples{
heart()
heart("Wish you happy!")
}
