% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{A Simple Hello Function}
\usage{
hello_world(name = "User")
}
\arguments{
\item{name}{A character string representing the name of the person to greet.}
}
\value{
A greeting message.
}
\description{
This function greets the user.
}
