% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how.long.R
\name{how.long}
\alias{how.long}
\title{Print time between first and second function call}
\usage{
how.long(clear = FALSE)
}
\arguments{
\item{clear}{is logical arg to clear current time marker}
}
\value{
time passed between calls
}
\description{
Print time between first and second function call
}
