% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRDscore.R
\name{HRDscore}
\alias{HRDscore}
\title{Scar score}
\usage{
HRDscore(
  dirpath = "E:/R/TCGA/HRD/TCGA/OV-Allele-specific_Copy_Number_Segment/",
  file_id_map = metadata,
  reference,
  ploidy = NULL,
  tumor = "OV"
)
}
\arguments{
\item{dirpath}{THE dir path of Allele specific Copy Number Segment files. sangerbox->数据下载->TCGA->cohort->Copy Number Variation(拷贝数)->Allele-specific Copy Number Segment}

\item{file_id_map}{file name match the sample ID. 一般在下载的文件最后一个,file_id_map_XX.TXT}

\item{reference}{reference genome: either grch38 or grch37 or mouse. Default is grch38.}

\item{ploidy}{Optional parameter, may be used if the ploidy of the sample is known. default is NULL, if have set ploidy=1}

\item{tumor}{Allele specific Copy Number Segment files数据来源,当ploidy非NULL时,需提供; data("tumor_ploidy"),unique(tumor_ploidy$Tumor)}
}
\value{
HRD scores
}
\description{
Determining genomic scar score (telomeric allelic imbalance, loss-off heterozigosity, large-scle transitions), signs of homologous recombination deficiency
}
\examples{
root_dir <-system.file("HRD", package = "Mypackage")
data("metadata")
HRD<-HRDscore(dirpath=root_dir,file_id_map=metadata,ploidy=NA,reference="grch38")
head(HRD)

}
\author{
Pengmin Yang
}
