% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifRep.R
\name{ifRep}
\alias{ifRep}
\title{Give the Glucose Value and Internal Time basing on ID}
\usage{
ifRep(inputId)
}
\arguments{
\item{inputId}{A dataset}
}
\value{
duplicated rows
}
\description{
Give the Glucose Value and Internal Time basing on ID
}
