% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immu_visual.R
\name{immu_visual}
\alias{immu_visual}
\title{Immune score Visualization}
\usage{
immu_visual(
  im,
  exp,
  method,
  exp_cleaning = FALSE,
  im_cleaning = FALSE,
  scale_mrna,
  tumor.TCGA,
  sample_group,
  heatmap = FALSE,
  heatmap_col = NULL,
  tumor,
  Type,
  color,
  geom_text,
  test
)
}
\arguments{
\item{im}{is the data of immune score for samples;if exp=NULL,im must be provided}

\item{exp}{is the data of gene expression for samples;please ensure gene coloum with unique values;if im=NULL,exp must be provided}

\item{method}{IOBR::deconvo_tme 参数或者根据其它方法得到的样本免疫浸润评分数据; method =c('mcpcounter', 'epic',  'cibersort', 'estimate','ips','quantiseq','xCell','timer')}

\item{scale_mrna}{logical. If FALSE, disable correction for mRNA content of different cell types. This is supported by methods that compute an absolute score (EPIC and quanTIseq)}

\item{tumor.TCGA}{method='timer',group_list参数，这里导入的group_list必须是TIMER能够识别的，比如TCGA中33肿瘤类型;Accepted indications are 'kich', 'blca', 'brca', 'cesc', 'gbm', 'hnsc', 'kirp', 'lgg','lihc', 'luad', 'lusc', 'prad', 'sarc', 'pcpg', 'paad', 'tgct','ucec', 'ov', 'skcm', 'dlbc', 'kirc', 'acc', 'meso', 'thca','uvm', 'ucs', 'thym', 'esca', 'stad', 'read', 'coad', 'chol'}

\item{sample_group}{is a dataframe for sample grouping}

\item{heatmap}{defalut heatmap=FALSE}

\item{heatmap_col}{defalut heatmap_col=NULL}

\item{tumor}{data source, "CRC TCGA"}

\item{Type}{can length from 2 to 5,eg Type=c("Wild", "Mut");(default = Type[[1]] is the control)}

\item{color}{for plot of Type, eg color=c("#757575", "#FF4040")}

\item{geom_text}{text annotation for ggplot}

\item{test}{参考ggpubr::compare_means，根据需要选择}

\item{exp_cleaning/im_cleaning}{是否对数据进行清洗，默认FALSE}
}
\value{
A list of immune score for data samples,method and heatmap of immune score for samples/ggplot object visualizing the immune score.
}
\description{
This function visualizes the mutation rate of immune score based on provided immune score or gene expression data.
}
\examples{
#im=NULL
data("exp_raw")
data("Gene_group_CRC1")
exp_CRC<-exp_geneIDtoSYMBOL(exp=exp_raw,genecoltype="ENTREZID")
exp_CRC<-exp_CRC$data
result<-immu_visual(im=NULL,exp=exp_CRC[,-1],
                    method = 'epic',
                    sample_group=Gene_group_CRC1,
                    tumor="CRC TCGA",heatmap=TRUE,
                    Type=c("Wild", "Mut"),
                    color=c("#757575", "#FF4040"),
                    geom_text=TRUE,
                    test = "wilcox.test")
print(result)

#exp=NULL
data("im")
data("sample_group")
result<-immu_visual(im=im,exp=NULL,
                    method = ' ',
                    sample_group=sample_group,
                    tumor="GC TCGA",heatmap=TRUE,
                    Type=c("Wild", "Mut"),
                    color=c("#757575", "#FF4040"),
                    geom_text=FALSE,
                    test = "wilcox.test")
print(result)
}
\author{
Pengmin Yang
}
