% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impuArima.R
\name{impu_arima}
\alias{impu_arima}
\title{Imputation with ARIMA Model}
\usage{
impu_arima(individualDataset)
}
\arguments{
\item{individualDataset}{individual dataset}
}
\value{
The dataset after imputed by ARIMA Model
}
\description{
Imputation with ARIMA Model
}
