% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impu.R
\name{imputing}
\alias{imputing}
\title{imputation}
\usage{
imputing(method, numGap)
}
\arguments{
\item{method}{Method to do imputation}

\item{numGap}{Number of missing value inserted}
}
\value{
imputation
}
\description{
imputation
}
