% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_variables.R
\name{include_variables}
\alias{include_variables}
\title{ダミー変数化された data.frame の特定の文字列を含む列名を抽出}
\usage{
include_variables(DATA, VECTOR)
}
\arguments{
\item{DATA}{data.frame 学習用データ}

\item{VECTOR}{vector 抽出したい文字列}
}
\value{
vector 抽出された文字列
}
\description{
ダミー変数化された data.frame の特定の文字列を含む列名を抽出
}
