% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexToId.R
\name{indexToId}
\alias{indexToId}
\title{Rows counting}
\usage{
indexToId(individualDataset)
}
\arguments{
\item{individualDataset}{An individual dataset}
}
\value{
number of total rows
}
\description{
Rows counting
}
