% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown.R
\name{indiedown_path}
\alias{indiedown_path}
\alias{indiedown_path_tex}
\title{Path helpers}
\usage{
indiedown_path(...)

indiedown_path_tex(...)
}
\arguments{
\item{...}{Path components, passed on to \code{system.file()}.}
}
\value{
Character string, path to this package's indiedown assets
}
\description{
\code{indiedown_path()} creates a path to this package's indiedown assets,
located at \code{inst/indiedown}.

\code{indiedown_path_tex()} creates a path that is usable in LaTeX,
with all special characters escaped.
}
\examples{
indiedown_path()
}
