% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown.R
\name{indiedown_pdf_document_with_asset}
\alias{indiedown_pdf_document_with_asset}
\title{indiedown_pdf_document_with_asset}
\usage{
indiedown_pdf_document_with_asset(includes = NULL, ...)
}
\arguments{
\item{includes}{...}

\item{...}{...}
}
\description{
FIXME
}
