% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{input_classic}
\alias{input_classic}
\title{Classic R approaches to data analysis}
\usage{
input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the pheno data file.}

\item{exprs_file}{character(1) The path to the expression file.}
}
\value{
A data.frame() containing samples as rows and phenotype
     data and expression data as columns.
}
\description{
The functions here illustrate how an 'old-school' R user
might approach data management and analysis.
}
\details{
There are 2 files. They are a microarray experiment.
Ther first file, ALL-phenoData.csv decribes the samples. The
second file, ALL-expression.csv are the normalized expression
values for a 12k probesetes across 128 samples. The files
originally came from sheets in excel, and were exporeted from
Excel using export-to-csv.
}

