% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmp.R
\name{kmp}
\alias{kmp}
\title{match times of appearance of x in y}
\usage{
kmp(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
match the times of appearance of \code{x} in \code{y}
}
\description{
match times of appearance of x in y
}
\examples{
kmp(c(1,2),c(1,2,1,2,3))
kmp(c(1,1),c(1,1,1,1,1))
}
