% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{knn_classification}
\alias{knn_classification}
\title{使用KNN进行分类}
\usage{
knn_classification(train, test, cl, k)
}
\arguments{
\item{train}{训练数据集}

\item{test}{测试数据集}

\item{cl}{训练数据的标签}

\item{k}{KNN算法中的k值}
}
\value{
预测的分类结果和混淆矩阵
}
\description{
这个函数使用KNN算法对数据进行分类。
}
