% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largestGap.R
\name{largestGap}
\alias{largestGap}
\title{Report the largest gap in the dataset for a certain individual}
\usage{
largestGap(individualDataset)
}
\arguments{
\item{inputId}{inputId}
}
\value{
The largest gap
}
\description{
Report the largest gap in the dataset for a certain individual
}
