% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fun.R
\name{lib_summary}
\alias{lib_summary}
\title{Summarise installed R packages by library}
\usage{
lib_summary()
}
\value{
A df with one row per library path, and a column for its package count
}
\description{
This function counts the number of packages installed
in each R library in your current environment.
}
\examples{
# Get a summary of installed packages by library path
lib_summary()

}
