% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.dif.visual.R
\name{limma.dif.visual}
\alias{limma.dif.visual}
\title{Differential Gene Expression Data Visualization}
\usage{
limma.dif.visual(
  exprdata,
  datatype = "counts",
  pdata,
  Type,
  tumor,
  diff_method = c("limma", "DESeq2", "edgeR"),
  contrastfml,
  P.Value,
  logFC,
  ann_colors,
  color,
  Regulate,
  geom_text = FALSE,
  tidyHeatmap = TRUE,
  heatmap_col = NULL,
  top50 = TRUE,
  GO = TRUE,
  GO.plot = "dotplot",
  split = FALSE,
  KEGG = TRUE,
  KEGG.plot = "dotplot",
  rel_heights = c(1.5, 0.5, 1)
)
}
\arguments{
\item{exprdata}{is the data of gene expression for samples}

\item{datatype}{exprdata的类型是"counts"时可以用"DESeq2"/"edgeR"方法;"limma"不挑}

\item{pdata}{A data frame that contains groups of genes.例如data.frame(SMPLE_ID,Group)}

\item{Type}{A character vector indicating the type.eg Type=c("Wild", "Mut")}

\item{tumor}{the source of exprdata}

\item{diff_method}{基因表达差异分析方法c("limma","DESeq2","edgeR")}

\item{contrastfml}{the Type for limma; eg contrastfml="Wild - Mut"}

\item{P.Value}{p阈值for差异分析}

\item{logFC}{logFC阈值for差异分析}

\item{ann_colors}{ComplexHeatmap::pheatmap 参数for sample annotation}

\item{color}{ComplexHeatmap::pheatmap 参数gene expression}

\item{Regulate}{select Regulate=c("Up","Down") for pathway enrichment}

\item{geom_text}{defalut KEGG=FALSE;if TURE,Volcano plot will annotate with}

\item{tidyHeatmap}{defalut tidyHeatmap=TRUE}

\item{heatmap_col}{defalut heatmap_col=NULL,tidyHeatmap::heatmap参数palette_value Regulate}

\item{top50}{defalut top50=TRUE 热图展示所用}

\item{GO}{defalut GO=TRUE}

\item{GO.plot}{GO富集图展现形式，GO.plot=c("dotplot","barplot","heatplot","cnetplot","gseaplot")}

\item{split}{defalut split=FALSE;if TURE,GO with "BP", "MF" and "CC",else GO with "BP"}

\item{KEGG}{defalut KEGG=TRUE}

\item{KEGG.plot}{KEGG富集图展现形式，GO.plot=c("dotplot","barplot","heatplot","cnetplot","gseaplot")}

\item{rel_heights}{relative heights of subplots. defalut rel_heights = c(1.5, 0.5, 1)}
}
\value{
A list of differential analysis results,volcano,heatmap and enrichment plot of GO AND KEGG
}
\description{
This function visualizes the Differential Gene Expression Data.
}
\examples{
data("exp_raw")
data("Gene_group_CRC1")
exp_CRC<-exp_geneIDtoSYMBOL(exp=exp_raw,genecoltype="ENTREZID")
exp_CRC<-exp_CRC$data
result<-limma.dif.visual(exprdata=exp_CRC[,-1],
                           pdata=Gene_group_CRC1,datatype="TPM",
                           Type=c("Wild", "Mut"),diff_method="limma",
                           contrastfml= "Wild - Mut",
                           tumor="CRC TCGA",
                           P.Value=0.05,
                           logFC=0.5,tidyHeatmap=TRUE,
                           color= NULL,
                           ann_colors = list(regulate = c(Down = "#1B9E77", Up = "#D95F02"),
                           PREX2 = c(Wild = "#757575", Mut = "#FF4040")),
                           Regulate=c("Up","Down"),GO=TRUE,GO.plot="dotplot",split=TRUE,
                           KEGG=TRUE,KEGG.plot="dotplot",rel_heights= c(1.5, 0.5, 1))
print(result)
}
\author{
Pengmin Yang
}
