% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression Model}
\usage{
linear_regression(x, y)
}
\arguments{
\item{x}{Numeric vector (independent variable).}

\item{y}{Numeric vector (dependent variable).}
}
\value{
An object of class "lm".
}
\description{
Fits a simple linear regression model.
}
\examples{
x <- 1:10
y <- 1:10 + rnorm(10)
model <- linear_regression(x, y)
summary(model)
}
