% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_anova.R
\name{lm_anova}
\alias{lm_anova}
\title{Linear Regression Anova}
\usage{
lm_anova(x)
}
\arguments{
\item{x}{input list obtained from lm_fit, consisting of sufficient information}
}
\value{
the ANOVA table
}
\description{
Conduct the Analysis of Variance (ANOVA) and output the ANOVA table
}
\examples{
m1 <- lm_fit(mpg ~ cyl + disp, data = mt_cars)
lm_anova(m1)

}
