% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_anova_fit.R
\name{lm_anova_fit}
\alias{lm_anova_fit}
\title{Linear Regression Anova Fit}
\usage{
lm_anova_fit(x)
}
\arguments{
\item{x}{input matrix, consisting of Y and X}
}
\value{
the SSR of the given matrix
}
\description{
Get the sum of squares due to regression (SSR), of a given matrix
}
\examples{
ax <- mt_cars[,1:3]
lm_anova_fit(ax)

}
