% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_detail.R
\name{lm_detail}
\alias{lm_detail}
\title{Linear Regression Detail}
\usage{
lm_detail(x)
}
\arguments{
\item{x}{input list obtained from lm_fit, consisting of sufficient information}
}
\value{
the details of linear regression
}
\description{
Output the details of the linear regression results
}
\examples{
m1 <- lm_fit(mpg ~ cyl + disp, data = mt_cars)
lm_detail(m1)

}
