% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.R
\name{lm_fit}
\alias{lm_fit}
\title{Linear Regression Fit}
\usage{
lm_fit(formula, data)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}
}
\value{
a list containing sufficient information
}
\description{
Conduct the Analysis of Variance (ANOVA) and output the ANOVA table
}
\examples{
m1 <- lm_fit(mpg ~ cyl + disp, data = mt_cars)

}
