% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{MAD}
\alias{MAD}
\title{Calculate mean absolute deviation}
\usage{
MAD(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - mean absolute deviation of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the mean absolute deviation
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MAD(data.frame)
MAD(matrix)
MAD(vector)
MAD(List)
}
