% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneFilter.R
\name{mad_filter}
\alias{mad_filter}
\title{MAD Filter}
\usage{
mad_filter(data, percent = 0.5, rmLower = TRUE)
}
\arguments{
\item{data}{Data matrix}

\item{percent}{The percentage of total gene need to remove Defaults to 0.5}

\item{rmLower}{The \pkg{lower} MAD gene need to remove or remove \pkg{higher} if set to FALSE  Defaults to TRUE}
}
\value{
Filtered data matrix based on MAD
}
\description{
This function perform gene filter based on Median Absolute Deviation (MAD).
}
\examples{
mad_filter(data, percent=0.5, rmLower=TRUE)
}
\author{
WKL
}
\keyword{MAD}
\keyword{filter}
