% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_cor.R
\name{maf_cor}
\alias{maf_cor}
\title{Analyze Mutation and Clinical Data Correlation with MAF Format Data}
\usage{
maf_cor(
  mutation_data,
  clin,
  gene,
  top = 10,
  cin_col,
  corrplot_method = c("circle", "square", "ellipse", "number", "shade", "color", "pie")
)
}
\arguments{
\item{mutation_data}{明确要求MAF格式数据框，且至少包含Tumor_Sample_Barcode和Hugo_Symbol两列}

\item{clin}{临床数据框，至少包含样本ID}

\item{gene}{基因列表，默认为NULL，当top不为NULL时，会自动选择前多少个基因，当top为NULL时，会使用gene列表中的基因}

\item{top}{前多少个基因}

\item{cin_col}{临床变量 clin数据框中的列名，需要第一列为样本ID}

\item{corrplot_method}{相关系数图 corrplot方法 circle, square, ellipse, number, shade, color, pie}
}
\value{
A list containing the following elements: 
\item{cor_data}{A data frame containing the correlation between the selected genes and the clinical variable.}
}
\description{
This function analyzes mutation and clinical data correlation with MAF format data. It calculates the VAF (Variant Allele Frequency) for each gene and selects the top genes based on the specified criteria. Then, it calculates the correlation between the selected genes and a specified clinical variable. The function also generates a correlation plot to visualize the relationships between the genes and the clinical variable.
}
\author{
Pengmin Yang
}
