% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHorizontalBarGraph.R
\name{makeHorizontalBarGraph}
\alias{makeHorizontalBarGraph}
\title{指定した列名の横棒グラフを作成}
\usage{
makeHorizontalBarGraph(DATA, ROWNAME)
}
\arguments{
\item{DATA}{data.frame データ}

\item{ROWNAME}{character 列名}
}
\description{
指定した列名の横棒グラフを作成
}
