% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniqueIdentifiers.R
\name{makeList}
\alias{makeList}
\title{Give the Glucose Value and Internal Time basing on ID}
\usage{
makeList(inputId)
}
\arguments{
\item{inputId}{A string of he unique identifier}

\item{data}{The imported dataset}
}
\value{
A table containing the Glucose Value and Internal Time for that ID
}
\description{
Give the Glucose Value and Internal Time basing on ID
}
