% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFlanders.R
\docType{data}
\name{mapFlanders}
\alias{mapFlanders}
\title{Map of Flanders for data visualizaiotn}
\format{
A \code{\link[=ggmap]{ggmap()}}
}
\usage{
mapFlanders
}
\description{
A map of Flanders from Google,
to use as a background for the geographical visualization of the invasion
of Asian hornets in Flanders.
}
\keyword{datasets}
