% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage.R
\docType{package}
\name{Math4753Alls0003}
\alias{Math4753Alls0003}
\title{My R Package}
\description{
Contains a functions for calculating total sum of squares
}
\author{
Kendall Allsup \email{kendall.allsup@ou.edu}
}
