% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{Max}
\alias{Max}
\title{Calculate maximum}
\usage{
Max(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - maximum of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the maximum
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Max(data.frame)
Max(matrix)
Max(vector)
Max(List)
}
