% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxCol.R
\name{MaxCol}
\alias{MaxCol}
\title{Calculate maximum}
\usage{
MaxCol(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: numeric vector - maximum of every column of input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the maximum of every column
of input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MaxCol(data.frame)
MaxCol(matrix)
MaxCol(vector)
MaxCol(List)
}
