% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{MaxRow}
\alias{MaxRow}
\title{Calculate maximum}
\usage{
MaxRow(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: numeric vector - maximum of every row of input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the maximum of every row
of input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MaxRow(data.frame)
MaxRow(matrix)
MaxRow(vector)
MaxRow(List)
}
