% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{Mean}
\alias{Mean}
\title{Calculate arithmetical mean}
\usage{
Mean(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - arithmetical mean of whole table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the arithmetical mean
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Mean(data.frame)
Mean(matrix)
Mean(vector)
Mean(List)
}
