% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{MeanCol}
\alias{MeanCol}
\title{Calculate arithmetical mean}
\usage{
MeanCol(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: numeric vector - arithmetical means of every
column input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the arithmetical mean
of every column input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MeanCol(data.frame)
MeanCol(matrix)
MeanCol(vector)
MeanCol(List)
}
