% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanGeom.R
\name{MeanGeom}
\alias{MeanGeom}
\title{Calculate geometric mean}
\usage{
MeanGeom(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - geometric mean of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the geometric mean
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MeanGeom(data.frame)
MeanGeom(matrix)
MeanGeom(vector)
MeanGeom(List)
}
