% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{MeanHarm}
\alias{MeanHarm}
\title{Calculate harmonic mean}
\usage{
MeanHarm(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - harmonic mean of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the harmonic mean
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MeanHarm(data.frame)
MeanHarm(matrix)
MeanHarm(vector)
MeanHarm(List)
}
