% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanRow.R
\name{MeanRow}
\alias{MeanRow}
\title{Calculate arithmetical mean}
\usage{
MeanRow(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: numeric vector - arithmetical means of every
row input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the arithmetical mean
of every row input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MeanRow(data.frame)
MeanRow(matrix)
MeanRow(vector)
MeanRow(List)
}
