% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Median.R
\name{Median}
\alias{Median}
\title{Calculate median}
\usage{
Median(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - median of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the median
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Median(data.frame)
Median(matrix)
Median(vector)
Median(List)
}
