% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropolisCpp}
\alias{MetropolisCpp}
\title{Metropolis sampler for random walks using Rcpp.}
\usage{
MetropolisCpp(sigma, x0, N)
}
\arguments{
\item{sigma}{variance of the proposal distribution (double)}

\item{x0}{the initial position (double)}

\item{N}{the length of chain (integer)}
}
\value{
a list consists of a generated chain and accept rate of the sampler
}
\description{
A Metropolis sampler for random walks using Rcpp(in homework 9).
}
\examples{
\dontrun{
N = 2000
sigma = c(.05, .5, 2, 16)
x0 = 25
rw1.c = MetropolisCpp(sigma[1],x0,N)
rw2.c = MetropolisCpp(sigma[2],x0,N)
rw3.c = MetropolisCpp(sigma[3],x0,N)
rw4.c = MetropolisCpp(sigma[4],x0,N)
#number of candidate points rejected
Rej = cbind(rw1.c$k, rw2.c$k, rw3.c$k, rw4.c$k)
Acc = round((N-Rej)/N,4)
rownames(Acc) = "Accept rates"
colnames(Acc) = paste("sigma",sigma)
Acc
}
}
