% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Min.R
\name{Min}
\alias{Min}
\title{Calculate minimum}
\usage{
Min(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - minimum of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the minimum
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Min(data.frame)
Min(matrix)
Min(vector)
Min(List)
}
