% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinCol.R
\name{MinCol}
\alias{MinCol}
\title{Calculate minimum}
\usage{
MinCol(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: numeric vector - minimum of every column of input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the minimum of every column
of input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MinCol(data.frame)
MinCol(matrix)
MinCol(vector)
MinCol(List)
}
