% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinRow.R
\name{MinRow}
\alias{MinRow}
\title{Calculate minimum}
\usage{
MinRow(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: numeric vector - minimum of every row of input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the minimum of every row
of input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
MinRow(data.frame)
MinRow(matrix)
MinRow(vector)
MinRow(List)
}
