% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingNos.R
\name{MissingNos}
\alias{MissingNos}
\title{MissingNos Function}
\usage{
MissingNos(missingnos)
}
\arguments{
\item{missingnos}{can be run directly and 2 randomly picked numbers will be printed out}
}
\description{
This  function finds 2 missing numbers that were randomly picked without looking out of a bag that contains 100 tiles in which numbers of from 1 to 100 written on them
}
\author{
Melih Ozbek
MissingNos()
}
