% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{Fit a multiple linear regression}
\usage{
mlr(X, Y)
}
\arguments{
\item{x}{design matrix}

\item{y}{outcome}
}
\value{
the parameter estimates from a multiple regression model
}
\description{
Fit a multiple linear regression
}
\examples{
import data
mlr(X,Y)
}
