% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moda.R
\name{Moda}
\alias{Moda}
\title{Calculate moda}
\usage{
Moda(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - moda of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the moda
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Moda(data.frame)
Moda(matrix)
Moda(vector)
Moda(List)
}
