% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{- an integer vector}
}
\value{
The most frequently occuring value.
}
\description{
Function returs a value that appears most often in a set of data.
}

