% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode2}
\alias{mode2}
\title{mode2(NumericVector x) function}
\usage{
mode2(x)
}
\value{
An integer value - the mode of the input integer vector x.
}
\description{
Function mode2() determines the most frequently occurring value in an integer
vector (mode). If the mode is ambiguous (e.g. for 1, 2, 2, 2, 3, 3, 1, 3),
function returns any mode.
}

