% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode_Rcpp}
\alias{mode_Rcpp}
\title{mode_Rcpp}
\usage{
mode_Rcpp(x)
}
\description{
This function determines the most frequently occurring value in an integer vector(mode).
}
\details{
ARGUMENTS:
x - an integer vector

RETURN VALUE:
a string value
}
\seealso{
Other fungroup: \code{\link{simplify2array_Rcpp}}
}

