% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_heatmap.R
\name{modified_heatmap}
\alias{modified_heatmap}
\title{Heatmap to modify vcfR::heatmap.bp.}
\usage{
modified_heatmap(x, cbarplot = TRUE, rbarplot = TRUE, legend = TRUE,
  clabels = TRUE, rlabels = TRUE, na.rm = TRUE, scale = c("row",
  "column", "none"), col.ramp = viridisLite::viridis(n = 100, alpha = 1),
  ...)
}
\value{
image
}
\description{
Removes bargraph from axes.
}
\details{
Remove bargraph from axes.
}
