% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostCommonPercentage.R
\name{mostCommonPercentage}
\alias{mostCommonPercentage}
\title{Return the percentage of values which the most common value has.}
\usage{
mostCommonPercentage(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
The percentage presence of the most common value in a data frame. \code{y}.
}
\description{
Return the percentage of values which the most common value has.
}
\examples{
mostCommonPercentage(x)
}
