% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mult}
\alias{mult}
\title{Multiply together two numbers.}
\usage{
mult(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The product of \code{x} and \code{y}.
}
\description{
Multiply together two numbers.
}
\examples{
mult(1, 2)
mult(10, 1/10)
}

