\name{mult_fun}
\alias{mult_fun}
\title{Multiplying two numbers}
\description{Multiplies two numbers}
\usage{
mult_fun(x, y)
}
\arguments{
  \item{x}{The first value to be multiplied}
  \item{y}{The second value to be multiplied}	
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
  An object of class numeric representing the result of multiplying two numbers
}
\author{Rob Foy}
\examples{
val1 <- 11
val2 <- 19
mt <- mult_fun(val1, val2)
mt
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{multiplication}
