% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{multi}
\alias{multi}
\title{As quatro operacoes desagregadas}
\usage{
multi(a, b)
}
\arguments{
\item{a}{primeiro numero}

\item{b}{segundo numero}
}
\value{
Retorna a multiplicacao de a e b
}
\description{
A multiplicacao de a e b
}
\details{
Calculo aritmetico simples
}
\keyword{a,}
\keyword{b}
\keyword{multiplicacao,}
