% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\value{
a plot containing all figures
}
\description{
Plot multiple plots in a single window
}
\examples{
multiplot(p1, p2, p3, cols = 1)
}
\author{
Brendan Gongol
}
