% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{multiply}
\alias{multiply}
\title{Multiply}
\usage{
multiply(x)
}
\arguments{
\item{x:}{numeric list}
}
\value{
Mulltiply the values in the list
}
\description{
Takes in a numeric list and multiply all the values in the list
}
