% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{text string; This is the time of arrival}

\item{ServiceTimes}{text string; This is the time that the service takes}

\item{NumServers}{text string; This is the number of available servers}
}
\value{
A tibble outlining the arrival time, time the service began, the chosen server and the time the service ends
}
\description{
This function is used to simulate how customers will go through a first come, first serve queuing system as long as "fate" has already decided when each customer will arrive and their service times
}
\examples{
Multiserver(Arrivals = 118.9688, ServiceTimes = 193.33, NumServers = 1)
}
