% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_network.R
\name{mut_network}
\alias{mut_network}
\title{Gene Mutation Co_Occurence/Mutually_Exclusive analysis and visualization}
\usage{
mut_network(SNV, top, pValue, customdata = NULL)
}
\arguments{
\item{SNV}{A data frame containing SNV data.即MAF格式的数据}

\item{top}{突变率前x的基因, top=10}

\item{pValue}{pValue}

\item{customdata}{基因注释数据,第一列为基因名，第二列为注释信息参考data("Gene",package = "Mypackage")，默认为NULL}
}
\value{
A list of table of Gene Mutation Co_Occurence/Mutually_Exclusive analysis and the network plot
}
\description{
Gene Mutation Co_Occurence/Mutually_Exclusive analysis and visualization
}
\examples{
data("mutation_CRC")
result<-mut_network(SNV=mutation_CRC,
                    top=20,pValue=0.01,customdata=NULL)

}
\author{
Pengmin Yang
}
