% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutToMAF.R
\name{mutToMAF}
\alias{mutToMAF}
\title{NGS mutation data perform}
\usage{
mutToMAF(
  root_dir,
  clin,
  saveDATA = FALSE,
  mut_filter = TRUE,
  tumor_t = 10,
  site_depth = 100,
  hotspot_vaf = 0.009,
  non_hotspot_vaf = 0.045,
  hotspotloss_vaf = 0.095,
  non_hotspotloss_vaf = 0.195,
  panel = 639,
  loss_genes = c("APC", "ARID1A", "ARID2", "ATM", "ATR", "ATRX", "BAP1", "BARD1", "BLM",
    "BRCA1", "BRCA2", "BRIP1", "CDK12", "CDKN2A", "CHEK1", "CHEK2", "CREBBP", "DNMT3A",
    "FAM175A", "FANCA", "FANCC", "FANCD2", "FANCE", "FANCF", "FANCL", "FANCM", "FBXW7",
    "FLCN", "GEN1", "GNA13", "KDM5C", "KDM6A", "KMT2A", "KMT2C", "KMT2D", "MAP3K1",
    "MLH1", "MRE11", "MSH2", "MSH3", "MSH6", "MYC", "NBN", "NF1", "NF2", "NOTCH1",
    "NSD1", "PALB2", "PBRM1", "PMS2", "POLD1", "PPP2R1A", "PPP2R2A", "PRKDC", "PTCH1",
    "PTEN", "RAD50", 
     "RAD51", "RAD51B", "RAD51C", "RAD51D", "RAD54L", "RB1",
    "RECQL", "RINT1", "RNF43", "SDHA", "SDHB", "SDHC", "SDHD", "SETD2", "SLX4", "SMAD4",
    "SMARCA4", "SMARCB1", "STK11", "TP53", "TSC1", "TSC2", "VHL", "WRN", "XRCC2"),
  tumor = NULL,
  hots_genes = c("ALK", "BRAF", "EGFR", "ERBB2", "KRAS", "MET", "RET", "ROS1"),
  artifact = c("Germline", "background", "base_qual", "normal_artifact", "weak_evidence",
    "contamination", "strand_bias", "haplotype", "multiallelic", "chemotherapy_site",
    "suspected_bg", "low_allele_frac")
)
}
\arguments{
\item{root_dir}{生信拉的数据存储的文件路径.}

\item{clin}{data.frame格式入组患者编号，表头名称需为Tumor_Sample_Barcode}

\item{saveDATA}{if TRUE 将在root_dir路径下新建一个SNV文件夹并生成文件，默认FALSE}

\item{mut_filter}{突变筛选。defalut mut_filter=TRUE}

\item{tumor_t}{该位点肿瘤样本reads数}

\item{site_depth}{该位点测序深度}

\item{hotspot_vaf}{热点突变丰度.Druggable为'0|1|2|-'标签}

\item{non_hotspot_vaf}{非热点突变丰度}

\item{hotspotloss_vaf}{热点失活突变丰度}

\item{non_hotspotloss_vaf}{非热点失活突变丰度}

\item{panel}{测序的panel，例如639}

\item{loss_genes}{抑癌基因/HRR/MMR相关基因，和失活突变筛选有关}

\item{tumor}{对应的癌种tumor：运行data("hotsgenes"),colnames(hotsgenes)}

\item{hots_genes}{癌种对应的用药等级1，2和R1的基因。例如肺癌：hots_genes=c("ALK","BRAF","EGFR","ERBB2","KRAS","MET","RET","ROS1")}

\item{artifact}{假突变标签，默认artifact=c("Germline","background","base_qual","normal_artifact","weak_evidence","contamination","strand_bias","haplotype","multiallelic","chemotherapy_site","suspected_bg","low_allele_frac")}
}
\value{
A MAF type data for NGS mutation data with filtering
}
\description{
This function converts mutation data from NGS sequencing into MAF format files and performs filtering.
}
\examples{
data("clindata")
root_dir <-system.file("example",  package = "Mypackage")
result <-mutToMAF(root_dir=root_dir,clin=clindata,saveDATA=FALSE,mut_filter=TRUE,
                  tumor_t=10,site_depth=100,hotspot_vaf=0.009,
                 non_hotspot_vaf=0.045,hotspotloss_vaf=0.095,non_hotspotloss_vaf=0.195)
print(result)
}
\author{
Pengmin Yang
}
