% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_chol.R
\name{mvnorm_chol}
\alias{mvnorm_chol}
\title{Generate samples from multivariate normal distribution through Cholesky decomposition}
\usage{
mvnorm_chol(n, mu, Sigma)
}
\arguments{
\item{mu}{a nx1 mean vector}

\item{Sigma}{a nxn covariance matrix}

\item{N}{a number, the number of samples}
}
\value{
N realizations from the multivariate normal distribution
}
\description{
Generate samples from multivariate normal distribution through Cholesky decomposition
}
\examples{
N<-4
mu<-rnorm(6)
A<-matrix(rnorm(36),nrow=6)
Sigma<-t(A) \%*\% A
sample<-mvnorm_chol(n,mu,Sigma)
}
