% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{my_first_pkg_fun}
\alias{my_first_pkg_fun}
\title{My First Package Function Help Title}
\usage{
my_first_pkg_fun(x = 5)
}
\arguments{
\item{x}{Numeric input parameter. Default = 5.}
}
\value{
The function returns the squared value of the input.
}
\description{
My first package function takes a number and returns the squared value.
}
\details{
A more detailed explanation of the function.
}
