% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{my_function}
\alias{my_function}
\title{A very brief description of what your function does}
\usage{
my_function(print_statement = TRUE)
}
\arguments{
\item{print_statement}{a logical value indicating
whether to print the statement (default is \code{TRUE})}
}
\value{
\code{my_function} prints a declaration
announcing itself if the parameter is \code{TRUE}
}
\description{
A detailed description of what the function does.
}
\examples{
R code examples

}
