% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Gapminder data on life expectancy, GDP per capita, and population of every country}
\format{
A tibble with 1704 rows and 6 variables:
\describe{
\item{country}{a factor denoting a country}
\item{continent}{a factor denoting each continent}
\item{year}{a number denoting every 5 years from 1952 - 2007}
\item{lifeExp}{a number denoting life expectancy at birth (years)}
\item{pop}{an integer denoting population}
\item{gdpPercap}{an integer denoting GDP per capita ($)}
}
}
\source{
{stuff} \url{http://www.gapminder.org/data/}
}
\usage{
my_gapminder
}
\description{
This data and documentation come from the \code{gapminder} package,
available as \code{\link[gapminder]{gapminder}}.
}
\keyword{datasets}
