% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame that is being used to train}

\item{cl}{True class value of your training data}

\item{k_nn}{Integer representing the number of neighbors}

\item{k_cv}{Integer representing the number of folds}
}
\value{
Numeric Lists of objects that consists: class, which is a vector of
the predicted class Yhat for all observations. And cv_err, which is a
numeric with the cross-validation classifications error.
}
\description{
This function perform k-Nearest Neighbors cross-validation for input data.
}
\examples{
my_penguins <- my_penguins[, c("species", "bill_length_mm", "bill_depth_mm",
 "flipper_length_mm", "body_mass_g")]
my_penguins <- na.omit(my_penguins)
my_knn_cv(train = my_penguins[,2:5], cl = my_penguins$species, k_nn = 1, k_cv = 5)
my_knn_cv(train = my_penguins[,2:5], cl = my_penguins$species, k_nn = 5, k_cv = 5)

}
