% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{my_lm function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object to be used to determine the
symbolic description of the model to be fitted.}

\item{data}{Input data frame that is used to perform the linear model fit.}
}
\value{
Table with rows for each coefficient and columns for the Estimate,
Std.Error, t value, and Pr(>|t|) that are calculated by perform linear model
fit by using the data \code{data} and relation determined by \code{formula}.
}
\description{
This function performs a linear model fit for the input data with a formula.
}
\examples{
my_lm(mpg~hp + wt, mtcars)

}
