% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean.R
\name{my_mean}
\alias{my_mean}
\title{The first line is the title}
\usage{
my_mean(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
The second section is a longer
description of the function.
This can go on for multiple lines.
}
\examples{
my_mean(1:5)
}
\seealso{
\code{\link[base]{mean}}
}
\author{
Daniel D. Sjoberg
}
