% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data to be calculated}

\item{alternative}{Character string specifying the alternative hypothesis for
the test.}

\item{mu}{Number indicating the null hypothesis value of the
mean of the test.}
}
\value{
Numeric Lists of elements that calculated from t-test for the input data
\code{x} that consists the test_stat indicating the t value; df indicating
degrees of freedom; alternative indicating the value of the parameter alternative
; and p_val indicating the numeric p-value.
}
\description{
This function computes t_test statistics for input data.
}
\examples{
data <- rnorm(10, mean = 20, sd = 1)
my_t.test(x = data, 'two.sided', 20)
my_t.test(x = data, 'less', 20)

}
