% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations (default: 100)}

\item{n}{Number of trials (default: 10)}

\item{p}{Probability of success (default: 0.5)}
}
\value{
A bar plot of the binomial distribution and a table of proportions.
}
\description{
This function simulates a binomial experiment with a given number of trials,
probability of success, and number of iterations. It returns a bar plot of the
distribution of successes.
}
\examples{
mybin(iter = 1000, n = 10, p = 0.3)
}
