% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Sampling and resampling}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations}

\item{x}{the original sample}

\item{fun}{the function to pull the statistic from}

\item{alpha}{the confidence interval}

\item{cx}{another statistic}

\item{...}{special argument that can take additional arguments}
}
\value{
A histogram of BootStrap sample statistics
}
\description{
Takes one sample from population and resample from it.
}
\details{
A function from Lab 9
}
\examples{
\dontrun{myboot2(x=sam,fun="mean", alpha=0.05)
}
}
