% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim.R
\name{mychisim}
\alias{mychisim}
\title{mychisim}
\usage{
mychisim(n1 = 10, sigma1 = 3, mean1 = 5, iter = 1000, ymax = 0.1, ...)
}
\arguments{
\item{n1}{sample size}

\item{sigma1}{Sigma 1}

\item{mean1}{Mean 1}

\item{iter}{Number of iterations}

\item{ymax}{maximum y value}

\item{...}{special argument that can take additional arguments}
}
\value{
Summary statistics
}
\description{
A function to retrieve information about chi-squared tests
}
\details{
A function for MATH 4753
}
\examples{
\dontrun{mychisim()}
}
