\name{myci}
\alias{myci}
\title{Confidence Interval Calculation}
\description{
  This function calculates a 95 percent confidence interval for the mean of a given dataset.
}
\usage{
  myci(data)
}
\arguments{
  \item{data}{
    Numeric vector containing the dataset.
  }
}
\value{
  Numeric vector representing the confidence interval for the mean.
}
