% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{sample size as a number}

\item{iter}{the number of iterations}

\item{lambda}{the lambda value for the poisson distribution as a number}
}
\value{
a histogram of the sample mean, a barplot for the sampled y, and a
barplot for the probability
}
\description{
mycltp
}
\examples{
mycltp(n=2, iter=10000,lambda=4)
}
