% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mydata}
\alias{mydata}
\title{Mydata short description.}
\format{A data frame with...:
\describe{
  \item{first_variable}{some details}
  \item{second_variable}{more details}
  ...
}}
\source{
\url{http://wherediditcomefrom.org}
}
\usage{
mydata
}
\description{
This is a longer description.
Can be multiple lines if necessary.
}
\keyword{datasets}

