% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{myddt (Project1)}
\usage{
myddt(df, sp)
}
\arguments{
\item{df}{data frame}

\item{sp}{species}
}
\value{
A sophisticated plot, along with console information
}
\description{
myddt (Project1)
}
\examples{
\dontrun{myddt(ddt, "CCATFISH")}
}
