% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{myEigen_QR}
\alias{myEigen_QR}
\title{Principle Component Analysis}
\usage{
myEigen_QR(A, numIter = 1000)
}
\arguments{
\item{A->}{Square matrix}

\item{numIter->}{Number of iterations}
}
\description{
Performs PCA on matrix A using  QR Decomposition .
Function should output a list with D and V. 
D is a vector of eigenvalues of A. 
V is the matrix of eigenvectors of A (in the 
same order as the eigenvalues in D.)
}
\examples{
X = matrix(runif(n*p), nrow = n)
n = dim(X)[1]
p = dim(X)[2]
output_qr = myEigen_QR(t(X) \%*\% X, 1000)
}
