% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Compute Piecewise Regression Values}
\usage{
myf(x, coef)
}
\arguments{
\item{x}{Numeric vector of predictor values.}

\item{coef}{Numeric vector of coefficients from a linear model.}
}
\value{
Numeric vector of fitted values.
}
\description{
This function calculates values for a piecewise linear regression model with a breakpoint at x = 18.
}
\examples{
coef <- c(5, 1.5, -0.2)
myf(20, coef)

}
