% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation (Univariate)}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A log-likelihood function of the form \code{function(x, param)}.}

\item{x}{A vector of data values.}

\item{param}{A vector of parameter values to evaluate.}

\item{...}{Additional plotting arguments.}
}
\value{
A list containing:
\describe{
  \item{i}{Index of the maximum likelihood estimate.}
  \item{parami}{The MLE parameter value.}
  \item{yi}{The log-likelihood at the MLE.}
  \item{slope}{The approximate slope around the MLE.}
}
}
\description{
Computes and plots the log-likelihood for a univariate parameter
using repeated sampling from the same distribution.
}
\examples{
logbin <- function(x, param) log(dbinom(x, prob = param, size = 10))
mymaxlik(logbin, x = c(9, 9, 1, 9, 9, 9), param = seq(0, 1, length = 1000))

}
