% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{myMean}
\alias{myMean}
\title{myMean}
\usage{
myMean(vect)
}
\arguments{
\item{vect}{numeric vector}
}
\value{
a number indicating the mean of the input vector
}
\description{
computing mean of a vector
}
\examples{
data(subcars)
x <- subcars$speed
print(myMean(x))
}
