% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Creates a negative binomial probability distribution for variable y}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{number of trials until the rth success is observe}

\item{r}{number of successes}

\item{p}{probability of success on a single Bernoulli trial}
}
\value{
probability of y in the negative binomial distribution
}
\description{
Creates a negative binomial probability distribution for variable y
}
\examples{
mynbin(10,3,0.4)
mynbin(21, 2, .1)

}
