% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNormCurve.R
\name{myNormCurve}
\alias{myNormCurve}
\title{Normal distribution plotting function}
\usage{
myNormCurve(mu, sigma, a)
}
\arguments{
\item{mu}{decimal or integer center of distribution}

\item{sigma}{decimal or integer standard deviation}

\item{a}{upper limit for the area under the distribution}
}
\description{
This function will create a normal distribution given then mean and standard deviation,
and will calculate area under the distribution from negative infinity to a, a user-input parameter.
}
\examples{
myNormCurve(2, 3, 1)
myNormCurve(4, 1, 5)

}
