% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatever.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: My First Collection of Functions}
\description{
A (one paragraph) description of what the package does and why it may be useful. Very useful it is.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/romantyk/mypackage/}
  \item Report bugs at \url{https://github.com/romantyk/mypackage/issues}
}

}
\author{
\strong{Maintainer}: Roman Tyk

}
