% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_help.R
\docType{package}
\name{mypackage}
\alias{mypackage}
\title{mypackage: A package for the workshop}
\description{
The mypackage package have two important functions:
\code{\link[mypackage]{colVars}} and \code{\link[mypackage]{rowVars}}
}
\section{mypackage function}{

The \code{colVars} function calculates column variances.
The \code{rowVars} function calculated row variances.
}

