% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{My awesome plot function}
\usage{
myplot(tbhdat, Scientificname)
}
\arguments{
\item{tbhdat}{input dataset}

\item{Scientificname}{chr string of species to plot}
}
\value{
A summary plot and table
}
\description{
My awesome plot function that creates summary plots
}
\examples{
tbhdat <- read.csv('example_len_dat.csv')
myplot(tbhdat, 'Lagodon rhomboides')
}
