% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapp.R
\name{mypvalue}
\alias{mypvalue}
\title{Bootstrapper for p-value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{@description T-value calculated exogenously}

\item{xmax}{@description x-axis calibrator}

\item{n}{@description Number of elements}

\item{alpha}{@description Confidence level you're testing}
}
\value{

}
\description{
Bootstrapper for p-value
}
\examples{
mypvalue()
}
