% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{myQR}
\alias{myQR}
\title{QR Decomposition}
\usage{
myQR(A)
}
\arguments{
\item{A->}{an n x m matrix}
}
\description{
Performs QR decomposition on the matrix A.
Function should output a list with Q.transpose and R
Q is an orthogonal n x n matrix
R is an upper triangular n x m matrix
Q and R satisfy the equation: A = Q %*% R
}
\examples{
myQR(A)
}
