% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRead.R
\name{myRead}
\alias{myRead}
\title{Read in CSV files}
\usage{
myRead(csv)
}
\arguments{
\item{csv}{Alphanumeric file path}
}
\description{
This function will read in CSV files in a way that is preferable to use in
class.
}
\examples{
\dontrun{myRead("SPRUCE.csv")}

}
