% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{myRidge}
\alias{myRidge}
\title{Ridge Regression}
\usage{
myRidge(X, Y, lambda)
}
\arguments{
\item{X->}{a matrix of explanatory variables.}

\item{Y->}{a vector of dependent variables. Y can also be a 
matrix, as long as the function works.}

\item{lambda->}{regularization parameter (lambda >= 0)}
}
\description{
Perform ridge regression of Y on X. Returns beta, the ridge regression solution.
}
\examples{
 my.data <- trees
 X <- as.matrix(trees[,1:2])
 Y <- trees$Volume
 beta_ridge <- myRidge(X, Y, 0)


}
