% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{time}{time intervals}
}
\value{
display curve
}
\description{
generates a sample with a time factor
}
\details{
A function from MATH 4753
}
\examples{
\dontrun{mysample(n=10,iter=100)
}

}
