% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{mySweep}
\alias{mySweep}
\title{Sweep Operator}
\usage{
mySweep(A, m)
}
\arguments{
\item{A->}{Square matrix}

\item{m->}{the pivot element is A[m, m].}
}
\description{
Perform a SWEEP operation on A with the pivot element A[m,m].
Returns a swept matrix B (which is m by m).
}
