% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim.R
\name{myTsim}
\alias{myTsim}
\title{myTsim}
\usage{
myTsim(n1 = 10, sigma1 = 3, mean1 = 5, iter = 1000, ymax = 0.1, ...)
}
\arguments{
\item{n1}{sample size}

\item{sigma1}{Sigma 1}

\item{mean1}{Mean 1}

\item{iter}{Number of iterations}

\item{ymax}{maximum y value}

\item{...}{special argument that can take additional arguments}
}
\value{
Summary statistics
}
\description{
A function to retrieve information about T-sim tests
}
\details{
A function for MATH 4753
}
\examples{
\dontrun{myTsim()}
}
