% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{myttest}
\usage{
myttest(x, y, alpha, paired)
}
\arguments{
\item{x}{#first list of data}

\item{y}{#second list of data}

\item{alpha}{#alpha value for t.test}

\item{paired}{#use paired = TRUE for paired t.test, paired = FALSE for unpaired t.test}
}
\value{
returns a list of length 6: a data frame of x and y,
 a confidence interval of the t.test, the type of test used,
 p.value of the t.test, a rejection statement,
 and the alpha value used. Class is Rttest.
}
\description{
an example function for DSA 5041. Given two data sets,
 performs a ttest, either Student's, Welch, or Paired, depending on the data.
 Outputs a list of Rttest, use print.Rttest or plot.Rttest to display the
 output more cleanly.
}
\examples{
set.seed(21); myttest(x = rnorm(30,5,2), y = rnorm(30,3,2), alpha =.05, paired = FALSE)
}
