% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{n.closest.partition}
\alias{n.closest.partition}
\title{N closest partitions}
\usage{
n.closest.partition(partition, n = 10, distance = "VI")
}
\description{
Finds the n closest partitions covering or being covered wrt a given one under B/VI losses
}
