% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na2zero.R
\name{na2zero}
\alias{na2zero}
\title{mypackage}
\usage{
na2zero(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
dataframe
}
\description{
Turns NAs in a dataframe into zeroes
}
\examples{
dataset <- matrix(sample(c(NA,1:5), 25, replace=TRUE), 5)
df <- as.data.frame(dataset)
mypackage::na2zero(df)

}
