% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naInsert.R
\name{naInsert}
\alias{naInsert}
\title{insert NAs}
\usage{
naInsert(individualDataset)
}
\arguments{
\item{individualDataset}{An individual dataset}
}
\value{
dataset after insertin NAa
}
\description{
insert NAs
}
