% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naNum.R
\name{naNum}
\alias{naNum}
\title{Report the number of NAs in the dataset for a certain individual}
\usage{
naNum(individualDataset)
}
\arguments{
\item{inputId}{inputId}
}
\value{
The the number of NAs
}
\description{
Report the number of NAs in the dataset for a certain individual
}
