% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_event_api.R
\name{naturalevents}
\alias{naturalevents}
\title{EONET}
\usage{
naturalevents(t, event, APIkey)
}
\arguments{
\item{t}{Limit the number of prior days (including today) from which events will be returned.}

\item{event}{The name of event}

\item{APIkey}{Your api_key applied from NASA website.}
}
\value{
Return the number of a specific event happened within the past t days.
}
\description{
The Earth Observatory Natural Event Tracker (EONET) is kindly provided by NASA website which is a curated source of continuously updated natural event metadata around the world.
This function is to 1) get the number of a specific natural event happened in the prior t days; 2) simplify the process of retrieving data from the web site and users just need to grasp their own api key.
}
\examples{
naturalevents(t = 10, event = "Wildfire", APIkey)
}
