% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newModalObservation.R
\name{newModalObservation}
\alias{newModalObservation}
\title{Return a vector made up of the modes of a data frame.}
\usage{
newModalObservation(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
A vector made up of the modes of the cols of a data frame. \code{y}.
}
\description{
Return a vector made up of the modes of a data frame.
}
\examples{
newModalObservation(x)
}
