\name{nogen}
\alias{nogen}
\title{Normal Distribution}
\usage{
nogen(u,s)
}
\description{
This function generates a random number from Normal distribution with mean = u and SD = s.
}
\arguments{
  \item{u}{mean of Normal distribution}
  \item{s}{standard deviation of Normal distribution}
}
\references{
  dr.gharib and mr.heidari
}
\examples{
nogen(4,6)
}
