% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{normalPCA}
\alias{normalPCA}
\title{Principal Component Analysis}
\usage{
normalPCA(data, main = "PCA plot")
}
\arguments{
\item{data}{Data matrix}

\item{main}{The title in plot Defaults to "PVCA plot"}
}
\value{
PCA plot
}
\description{
This function perform PCA with one step code or step-by-step.
}
\examples{
normalPCA(data, main="PCA plot")
}
\author{
WKL
}
\keyword{PCA}
