% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normf_standf.R
\name{normf}
\alias{normf}
\title{Normalizes a numeric vector}
\usage{
normf(x, newMin, newMax, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{newMin}{Numeric value for the new minimum of the vector}

\item{newMax}{Numeric value for the new maximum of the vector}

\item{na.rm}{Logical value determining whether or not to remove NA
values when calculating min and max. Defaults to FALSE.}
}
\value{
Numeric vector that is normalized
}
\description{
Takes a numeric vector and normalizes it with new min and max
}
\examples{
normf(c(1, 2, 3, 4, 5), 10, 20)
normf(c(1, 2, 10, -10), 5, 10, na.rm = TRUE)
}
