% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.me.R
\name{notify.me}
\alias{notify.me}
\title{Print message to notifications when called}
\usage{
notify.me(message = "R completed")
}
\arguments{
\item{message}{is string to display in notif. banner}
}
\value{
notification banner
}
\description{
Print message to notifications when called
}
