% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_diffs.R
\name{nucleotide_diff}
\alias{nucleotide_diff}
\title{Find single character differences in sequences}
\usage{
nucleotide_diff(
  baseline_seq,
  compare_seqs,
  output = c("num", "position"),
  ignore_case = TRUE
)
}
\arguments{
\item{baseline_seq}{Character vector of length one that all other sequences
are being compared to.}

\item{compare_seqs}{Character vector of any length that will be compared to
the baseline seq.}

\item{output}{Either "num" (return the number of differences) or "position"
(return the positions of the differences).}

\item{ignore_case}{Logical value (TRUE / FALSE) designating whether the
function should ignore case. Defaults to TRUE.}
}
\value{
Either a list of numeric vectors describing the positions of
differences or a numeric vector of the number of differences.
}
\description{
Function takes a baseline sequence and any number of sequences to compare to
and determines the number or the positions of single character differences.
Note that this function can be used with amino acid sequences
(1 letter abbreviation) as well.
}
\examples{
nucleotide_diff('AGTGGG', 'CGTGAG', output = 'num')

seq1 = 'ACTGACTGG'
seq2 = c('ACGGGCCGG', 'ACTGACTGG', 'GCTCAGTGG')
nucleotide_diff(seq1, seq2, output = 'num', ignore_case = TRUE)
nucleotide_diff(seq1, seq2, output = 'position', ignore_case = FALSE)

\dontrun{
nucleotide_diff(seq1, seq2)
nucleotide_diff('AGTCGA', 'TTA')
}
}
