% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS_QR.R
\name{OLS_QR}
\alias{OLS_QR}
\title{Perform OLS estimation through QR decomposition}
\usage{
OLS_QR(X, Y)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response}
}
\value{
OLS estimates beta
}
\description{
Perform OLS estimation through QR decomposition
}
