% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS_SVD.R
\name{OLS_SVD}
\alias{OLS_SVD}
\title{Perform OLS estimation through SVD decomposition}
\usage{
OLS_SVD(X, Y)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response}
}
\value{
OLS estimates beta
}
\description{
Perform OLS estimation through SVD decomposition
}
