% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sample_z.R
\name{one_sample_z}
\alias{one_sample_z}
\title{One sample z test}
\usage{
one_sample_z(v, mu)
}
\arguments{
\item{v}{a numeric vector representing a sample}

\item{mu}{the hypothesised mean}
}
\value{
a single element list of two strings. The first gives the
z and p values, the second indicates whether the sample mean is
signifcantly different from the hypothesised values
}
\description{
Tests whether the mean of a sample differs significantly
from a hypothesised value, mu.
}
\examples{
samplex <- rnorm(50, 20, 3)
one_sample_z(samplex, 20)
}
