% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{onestepPCA}
\alias{onestepPCA}
\title{Principal Component Analysis}
\usage{
onestepPCA(data, trait, colour, label = FALSE)
}
\arguments{
\item{data}{Data matrix}

\item{trait}{Sample information for PCA plot}

\item{colour}{One of the trait for colour in PCA plot}

\item{label}{Whether use the label in PCA plot or not Defaults to FALSE}
}
\value{
PCA plot
}
\description{
This function perform PCA with one step code or step-by-step.
}
\examples{
onestepPCA(data, trait, colour="Gender", label=FALSE)
}
\author{
WKL
}
\keyword{PCA}
