% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_outliers.R
\name{outlier_connectivity}
\alias{outlier_connectivity}
\title{Sample filter based on connectivity z-socres}
\usage{
outlier_connectivity(data, onetrait = NULL, sdout = 2, label = FALSE)
}
\arguments{
\item{data}{Data matrix}

\item{onetrait}{One sample information for plot Defaults to NULL}

\item{sdout}{The threshold for outliers Defaults to 2}

\item{label}{Whether use the label in plot or not Defaults to FALSE}
}
\value{
outliers
}
\description{
This function remove outliers based on network connectivity z-scores.
}
\examples{
outlier_connectivity(data, onetrait, sdout=2, label=FALSE)
}
\author{
WKL
}
\keyword{connectivity,}
\keyword{outliers}
