% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_outliers.R
\name{outlier_hcluster}
\alias{outlier_hcluster}
\title{Sample filter based on hierarchical clustering}
\usage{
outlier_hcluster(data, traitNum = NULL, scale = FALSE, cutheight = 50000,
  minSize = 10, plot = TRUE, print = TRUE)
}
\arguments{
\item{data}{Data matrix}

\item{traitNum}{Sample information which must be numeric for cluster plot Defaults to NULL}

\item{scale}{Whether scale is used for the data Defaults to FALSE}

\item{cutheight}{The threshold for outliers Defaults to 50000}

\item{minSize}{The minimum cluster size for outliers Defaults to 10}

\item{plot}{Whether output the plot or not Defaults to TRUE}

\item{print}{Whether print the result or not Defaults to TRUE}
}
\value{
a local webpage for visualization of hcluster
}
\description{
A simple \pkg{shiny} app to show outliers based on hierarchical clustering on local web browser.
}
\examples{
outlier_hcluster(data, trait_num, cutheight=50000)
}
\author{
WKL
}
\keyword{cluster,}
\keyword{shiny}
