% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_outliers.R
\name{outlier_PCA}
\alias{outlier_PCA}
\title{Sample filter based on PCA}
\usage{
outlier_PCA(data, trait = NULL, colour = NULL, label = FALSE,
  pcout = 0.3)
}
\arguments{
\item{data}{Data matrix}

\item{trait}{Sample information for PCA plot Defaults to NULL}

\item{colour}{One of the trait for colour in PCA plot Defaults to NULL}

\item{label}{Whether use the label in PCA plot or not Defaults to FALSE}

\item{pcout}{The threshold for outliers Defaults to 0.3}
}
\value{
outliers
}
\description{
This function remove outliers based on principal component analysis (PCA).
}
\examples{
outlier_PCA(data, trait, colour="Gender", pcout=0.3)
}
\author{
WKL
}
\keyword{cluster}
