% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheatsheet.R
\name{p_normal_greater}
\alias{p_normal_greater}
\title{Probability P(X >= a) for Normal Distribution}
\usage{
p_normal_greater(a, mean, std)
}
\arguments{
\item{a}{Threshold value.}

\item{mean}{Mean of the normal distribution.}

\item{std}{Standard deviation of the normal distribution.}
}
\value{
Probability that X ≥ a.
}
\description{
Computes P(X >= a) for a normal distribution using the tail probability.
}
\examples{
p_normal_greater(5, 6, 1.5)
}
