% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheatsheet.R
\name{p_poisson_greater}
\alias{p_poisson_greater}
\title{Probability P(X >= a) for Poisson Distribution}
\usage{
p_poisson_greater(a, lambda)
}
\arguments{
\item{a}{Threshold integer.}

\item{lambda}{Poisson rate parameter.}
}
\value{
Probability that X ≥ a.
}
\description{
Computes P(X >= a) for a Poisson distribution using the tail probability.
}
\examples{
p_poisson_greater(10, 4)
}
