% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{partitions.counter}
\alias{partitions.counter}
\title{Partition Counter}
\usage{
partitions.counter(clust.unordered)
}
\description{
Counting the partitions visited by an MCMC - or computing the mode if you want
}
