% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_mut_visual.R
\name{path_mut_visual}
\alias{path_mut_visual}
\title{Pathway Mutation Visualization}
\usage{
path_mut_visual(
  result_data = NULL,
  SNV,
  gene,
  Gene_group,
  Type,
  pathway_gene_data,
  tumor,
  heatmap = FALSE,
  heatmap_col = NULL,
  color,
  test = "wilcox.test",
  ns = FALSE,
  p_0.05 = FALSE,
  p_0.01 = FALSE,
  p_0.001 = FALSE,
  p_0.0001 = FALSE
)
}
\arguments{
\item{result_data}{A data frame containing mutation rate of tumor pathways}

\item{SNV}{A data frame containing SNV data.即MAF格式的数据}

\item{gene}{A character vector representing the gene of interest.例如BRAF基因或分组名称Group}

\item{Gene_group}{A data frame that contains groups of genes.例如data.frame(SMPLE_ID,Group)}

\item{Type}{A character vector indicating the type.Type can length from 2 to 5,eg Type=c("Wild", "Mut");(default = Type[[1]] is the control)}

\item{pathway_gene_data}{A data frame that contains pathway gene information.例如data.frame(Genes,Pathway)}

\item{tumor}{A character vector indicating tumor types.例如CRC}

\item{heatmap}{defalut heatmap=FALSE}

\item{heatmap_col}{defalut heatmap_col=NULL}

\item{color}{A character vector for defining colors.例如color=c("#757575", "#FF4040")}

\item{test}{参考ggpubr::compare_means，根据需要选择}

\item{ns}{没有显著性差异的结果展示与否，默认FALSE}

\item{p_0.05}{筛选p值小于0.05，默认FALSE}

\item{p_0.01}{筛选p值小于0.01，默认FALSE}

\item{p_0.001}{筛选p值小于0.001，默认FALSE}

\item{p_0.0001}{筛选p值小于0.0001，默认FALSE}
}
\value{
A list of path mutation data,path mutation frequency in each type and ggplot object visualizing the mutation rates.
}
\description{
This function visualizes the mutation rate of tumor pathways based on SNV and gene data.
}
\examples{
data("mutation_CRC")
data("gene_group_data")
data("pathway_data")
gene_of_interest<-colnames(gene_group_data)[[2]]
tumor_type="CRC TCGA"
color_vector=c("#757575", "#FF4040")
result <- path_mut_visual(result_data=NULL,
                           SNV = mutation_CRC,
                           gene = gene_of_interest,
                           Gene_group = gene_group_data,
                           Type = c("Wild", "Mut"),
                           pathway_gene_data = pathway_data,
                           tumor = tumor_type,
                           heatmap=TRUE,
                           heatmap_col=NULL,
                           color = color_vector,
                           test = "wilcox.test",ns=FALSE,
                           p_0.05=FALSE,p_0.01=FALSE,p_0.001=FALSE,p_0.0001=FALSE)
print(result)
}
\author{
Pengmin Yang
}
