% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percMiss.R
\name{percMiss}
\alias{percMiss}
\title{Calculate the percentage of missing data}
\usage{
percMiss(individualDataset)
}
\arguments{
\item{individualDataset}{A individual dataset}
}
\value{
The percentage of missing data
}
\description{
Calculate the percentage of missing data
}
