% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\title{Creates a piecewise function that is centered around xk}
\usage{
piecewise(x, coef, xk)
}
\arguments{
\item{x}{vector of x-values}

\item{coef}{array of 3 coefficients for the function {B0, B1, z}}

\item{xk}{x value where the the line segments meet}
}
\value{
y value for the function y = B0 + B1x + z(x-xk)(x>xk)
}
\description{
Creates a piecewise function that is centered around xk
}
\examples{
piecewise(2, c(1.23, .234, -.24), 18)
piecewise(3, c(1, 2, 3), 10)

}
