% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{plot.Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{The output from myttest}

\item{...}{Additional parameters to make plot generic happy}
}
\value{
returns boxplots of myttest's output list;
         side by side boxplots for unpaired data, or a single boxplot of the
         difference in x and y for paired data, includes confidence interval.
}
\description{
Defines the plot method for the Rttest class. Use only with output from
 myttest. Plots two plots for unpaired data, and one plot for paired data.
}
\examples{
set.seed(20); plot(myttest(rnorm(30,5,2), rnorm(30,3,2), .05, FALSE))
}
