% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abline.R
\name{plot_model}
\alias{plot_model}
\title{Plotting diamonds dataset}
\usage{
plot_model(model)
}
\arguments{
\item{model}{The model I previously fitted in the other function}
}
\value{
A plot
}
\description{
Plotting diamonds dataset
}
\details{
This function takes as parameter the data and the linear model we fitted using the function fit_model()
of this dataset and it displayes the relationship between carats and the price of the diamonds using as an
intercept the coefficients from the fitted model.
}
\examples{
plot_model(fit_model())
}
