% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_ppt.R
\name{plot_to_ppt}
\alias{plot_to_ppt}
\title{Export a ggplot to PowerPoint}
\usage{
plot_to_ppt(
  plot = ggplot2::last_plot(),
  file_name = "plot.pptx",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{plot}{A `ggplot` object to export. Defaults to the last plot produced (`ggplot2::last_plot()`).}

\item{file_name}{Character. Name of the PowerPoint file. Defaults to `"plot.pptx"`.}

\item{width}{Numeric. Optional width of the plot on the slide in inches. If `NULL`, uses slide width.}

\item{height}{Numeric. Optional height of the plot on the slide in inches. If `NULL`, uses slide height.}
}
\value{
Returns the original `ggplot` object invisibly, allowing for piping or further modifications.
}
\description{
This function exports a `ggplot2` plot to a PowerPoint (`.pptx`) file.
If the file already exists, the plot is added as a new slide; otherwise, a new
presentation is created. The plot can be sized automatically to fill the slide
or manually by specifying width and height.
}
\details{
- Uses the `officer` package to manipulate PowerPoint files and `rvg` to convert
  `ggplot2` objects into editable vector graphics.
- The slide layout is set to `"Title and Content"` and the master to `"Office Theme"`.
- If both `width` and `height` are `NULL`, the plot fills the entire slide.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
plot_to_ppt(p, "myplot.pptx")
}

}
