% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCI.R
\name{plotCI}
\alias{plotCI}
\title{Plot Error Bars and Confidence Intervals}
\usage{
plotCI(..., sfrac = 0, gap = 0, pch = 19, col = "blue")
}
\description{
Given a set of x and y values and interval width or upper and lower bounds,
plot the points with error bars.  This can be a useful tool for visualizing
confidence intervals.
}
\details{
This is a wrapper to \code{plotCI} from the \code{gplots} package, but using
different defaults.
}

