% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGV.R
\name{plotGV}
\alias{plotGV}
\title{Plot the Glucose Value on y-axis and Internal Time on x-axis basing on ID}
\usage{
plotGV(inputId, start_time, end_time)
}
\arguments{
\item{inputId}{A string of he unique identifier}
}
\value{
A GV plot
}
\description{
Plot the Glucose Value on y-axis and Internal Time on x-axis basing on ID
}
