% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_TS_trend.R
\name{plots}
\alias{plots}
\title{Visualize TS trend models}
\usage{
plots(
  ...,
  title = "",
  xlab = "Date",
  ylab = "Value",
  ylim = NULL,
  observed_data = TRUE,
  data
)
}
\arguments{
\item{...}{legend tab names = value}

\item{title}{set the title for the plot}

\item{xlab}{the label of x axis}

\item{ylab}{the label of y axis}

\item{ylim}{set the limit of y(optional); the default value is ylim = NULL.}

\item{data}{data source}
}
\value{
plot several time series trend models together with a legend.
}
\description{
Visualize the initial time series and fitted trend models in one plot automatically with a legend.
}
\examples{
linear_model = lm(value~time, ts_sample_data)
quad_model = lm(value~time + I(time^2), ts_sample_data)
log_model = lm(log(value)~time, ts_sample_data)

1. plots(linear = linear_model$fitted.values,
        quadratic = quad_model$fitted.values,
        log = exp(log_model$fitted.values),
        title = 'Time series Trend graph using linear, quadratic and exponential Models')

2. plots(linear = predict(linear_model),
        quadratic = predict(quad_model),
        log = exp(predict(log_model)),
        title = 'Time series Trend graph using linear, quadratic and exponential Models',
        observed_data = FALSE,
        ylim=c(2000, 10000))
}
