% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{plus10}
\alias{plus10}
\title{Plus 10 to every element}
\usage{
plus10(x, y)
}
\arguments{
\item{x}{A vector in which elements are added 10}

\item{y}{A logical vector}
}
\value{
result of first element added 10 if entered "TRUE" for y
}
\description{
Take in any vector, and add 10 to each element, and test the first element
}
