% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheatsheet.R
\name{poisson_histogram}
\alias{poisson_histogram}
\title{Histogram of Poisson Random Samples}
\usage{
poisson_histogram(n, lambda)
}
\arguments{
\item{n}{Sample size.}

\item{lambda}{Poisson rate parameter.}
}
\value{
A histogram (plot).
}
\description{
Generates a histogram of Poisson-distributed samples.
}
\examples{
poisson_histogram(1000, 4)
}
