\name{pOne}
\alias{pOne}

\title{Calculate One-side P-value}

\usage{
pOne(d.obs, d.base)
}

\arguments{
  \item{d.obs}{the distance for observed data.}
  \item{d.base}{the baseline samples of distances.}
}

\description{
  This function calculates one-side p-value for observed data based on the baseline samples of distances.}

\details{
  ...
} 
\value{

}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{
pOne(d.obs, d.base)     }
}

\seealso{
\code{\link{e2dist}}, \code{\link{baseline.dist}}
}

\keyword{Model Checking}
