% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name, punctuation = "!")
}
\arguments{
\item{name}{Text string which praises someone}

\item{punctuation}{This is an emphasis on the text input}
}
\value{
Text string with praise
}
\description{
Deliver Praise
}
\examples{
praise(name="Steph", punctuation = "!")
}
