% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Properties.R
\name{Properties}
\alias{Properties}
\title{Calculate main statistic values}
\usage{
Properties(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: named numeric vector - main statistic values of whole
input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate main statistic values of whole
input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Properties(data.frame)
Properties(matrix)
Properties(vector)
Properties(List)
}
