% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\alias{propNeeded,ANY-method}
\title{Computes proportion of remaining delegates necessary to secure the nomination}
\usage{
propNeeded(candidate, remaining.delegates)
}
\arguments{
\item{candidate}{An object of the class 'Candidate'}

\item{remaining.delegates}{The number of delegates yet to be allocated in the candidate's party.}
}
\value{
The proportion of the delegates a given candidate still needs to secure his/her party nomination
}
\description{
A function that calculates the proportion of remaining delegates needed by a presidential candidate to secure the nomination
}
\examples{

candidate.Hillary <- createCandidate("Hillary Clinton", "Democratic", 1614)
propNeeded(candidate.Hillary, 2295)

}
\author{
Jacob Hample \email{jacob.hample@wustl.edu}
}
\seealso{
\code{\link{createCandidate}}, \code{\link{Candidate}}
}

