% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVCA.R
\name{pvca}
\alias{pvca}
\title{Principal Variance Component Analysis}
\usage{
pvca(data, trait, pct_threshold = 0.5876, main = "PVCA plot",
  xlab = "Effects")
}
\arguments{
\item{data}{Data matrix}

\item{trait}{Sample information used in the liner model}

\item{pct_threshold}{The threshold for PC Defaults to .5876}

\item{main}{The title in plot Defaults to "PVCA plot"}

\item{xlab}{The xlab in plot Defaults to "Effects"}
}
\value{
PVCA plot
}
\description{
This function perform PVCA with interaction or not.
}
\examples{
pvca(data, trait)
pvca2(data, trait)
}
\keyword{PVCA}
