% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot.R
\name{QC_boxplot}
\alias{QC_boxplot}
\title{Box plot and bar plot}
\usage{
QC_boxplot(data, oneTrait = NULL, ylab = "log2(value+0.001)",
  main = "QC plot", mapping_reads = NULL, height = 5)
}
\arguments{
\item{data}{Data matrix}

\item{oneTrait}{One sample information which must be factor for box plot Defaults to NULL}

\item{ylab}{The title used for the ylab of boxplot Defaults to \pkg{log2(value+0.001)}}

\item{main}{The title for plot Defaults to \pkg{QC plot}}

\item{mapping_reads}{The mapping reads which match the data for bar plot Defaults to NULL}

\item{height}{The threshlod for bar plot Defaults to 5}
}
\value{
a boxplot of the data
}
\description{
Box plot of log transformed data and bar plot of the mapping reads or total counts for quanlity control.
}
\examples{
QC_boxplot(data, trait$Race, mapping_reads=(runif(65,min=1000000,max=10000000)))
}
\author{
WKL
}
\keyword{QC}
\keyword{barplot,}
\keyword{boxplot,}
