% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plot.R
\name{QC_density}
\alias{QC_density}
\title{Density plot}
\usage{
QC_density(data, oneTrait = NULL, main = "Density plot",
  xlab = "log2(value + .001)", ylim = c(0, 0.5))
}
\arguments{
\item{data}{Data matrix}

\item{oneTrait}{One sample information which must be factor for density plot Defaults to NULL}

\item{main}{The title for plot Defaults to \pkg{Density plot}}

\item{xlab}{The title used for the xlab of boxplot Defaults to \pkg{log2(value + .001)}}

\item{ylim}{The y-axis limitaiton for density plot Defaults to \pkg{c(0,0.5)}}
}
\value{
a density plot of the data
}
\description{
Density plot of log transformed data for quanlity control.
}
\examples{
QC_density(data, trait$Race)
}
\author{
WKL
}
\keyword{QC}
\keyword{density,}
