% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartile.R
\name{Quartile}
\alias{Quartile}
\title{Calculate quartiles of input table}
\usage{
Quartile(df)
}
\arguments{
\item{df}{Numeric 2 dimension vector}
}
\value{
Good way: named numeric vector - quartiles of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the quartiles
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Quartile(data.frame)
Quartile(matrix)
Quartile(vector)
Quartile(List)
}
