% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_movie_recommendation.R
\name{random_movie_recommendation}
\alias{random_movie_recommendation}
\title{Random Movie Recommendation}
\usage{
random_movie_recommendation()
}
\value{
A character string representing the title of a randomly selected movie.
}
\description{
This function randomly recommends a movie from the `ggplot2movies` dataset.
}
\examples{
random_movie_recommendation()
}
