% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:random.R
\name{random_number}
\alias{random_number}
\title{Random Number Generator}
\usage{
random_number(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value.
}
\description{
Random Number Generator
}
\examples{
random_number()
}
