% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Letters.R
\name{randomletters}
\alias{randomletters}
\title{Random Letter}
\usage{
randomletters(n = 1)
}
\arguments{
\item{n}{a number between 1 and 26 for the letters in the whole alphabet}
}
\value{
A Random Letter of the Alphabet
}
\description{
Chooses a random letter from the alphabet
}
