% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{Range}
\alias{Range}
\title{Calculate minimum and maximum}
\usage{
Range(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: named numeric vector - minimum and maximum of whole
input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate minimum and maximum of whole
input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
Range(data.frame)
Range(matrix)
Range(vector)
Range(List)
}
