% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remRep.R
\name{removeRep}
\alias{removeRep}
\title{Remove the replicated data entries}
\usage{
removeRep(inputData)
}
\arguments{
\item{inputId}{A dataset}
}
\value{
duplicated rows
}
\description{
Remove the replicated data entries
}
