% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsePredictGraph.R
\name{responsePredictGraph}
\alias{responsePredictGraph}
\title{正解データと予測結果をグラフで表示}
\usage{
responsePredictGraph(DATA, MODEL)
}
\arguments{
\item{DATA}{data.frame データ}

\item{MODEL}{学習済みモデル}
}
\description{
正解データと予測結果をグラフで表示
}
