% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_financial_year.R
\name{return_financial_year}
\alias{return_financial_year}
\title{Get the start year of the given financial year}
\usage{
return_financial_year(input_date)
}
\arguments{
\item{input_date}{A date in "YYYY-MM-DD" format or a Date object.}
}
\value{
An integer representing the start year of the financial year (e.g., 2025 for FY 2025/26).
}
\description{
This function returns the start year of the current financial year based on a given date.
The financial year is assumed to start on April 1st and end on March 31st of the following year.
}
\examples{
return_financial_year(input_date = "2025-10-29")  # Returns 2025
return_financial_year("2026-02-04")              # Returns 2025
}
