% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{RevComp}
\alias{RevComp}
\title{RevComp: Reverse Complement function}
\usage{
RevComp(myseq, type)
}
\arguments{
\item{x}{a DNA string, vector of DNA strings, or an object representing a vector of DNA strings.}

\item{designates}{the type of output.  "comp", "rev", or "revcomp" designate taking the complement, reverse,
or reverse complement of the sequence.}
}
\value{
a list of DNA strings from the desired DNA strings computed according to the selected functionn type as well as the original string.
}
\description{
This function takes the reverse, complement, or reverse complement of 1 or more DNA sequence strings.
}
\examples{
hi <- c("TTGGGGGCGTTAG", "CATTGGTTCTAGTT", "GCATTGGTGTGCATTAG")
RevComp(myseq=hi, type= "rev")
RevComp(myseq=hi, type= "comp")
RevComp(myseq=hi, type= "revcomp")
RevComp(myseq="ATGCATTGGACGTTAG", type="rev")
RevComp(myseq="ATGCATTGGACGTTAG", type="comp")
RevComp(myseq="ATGCATTGGACGTTAG", type="revcomp")
}
\author{
Brendan Gongol
}
