% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightleftorder.R
\name{rightleftorder}
\alias{rightleftorder}
\title{Left-right order
Left-right order}
\usage{
rightleftorder(n, i)
}
\arguments{
\item{n}{A positive integer.}

\item{i}{An integer between 1 and \code{n}.}
}
\description{
Order the numbers 1 through \code{n}, starting at index \code{i} and
proceeding rightward then leftward.
}
