\name{rongelap}
\alias{rongelap}

\title{Data Set of Rongelap Island}

\usage{
data(datRongelap)
}

\arguments{

}

\description{
  This data set contains the Rongelap data.}

\details{
  ...
} 
\value{

}

\author{
  Liang Jing \email{ljing918@gmail.com}
  }

\examples{
\dontrun{

data(datRongelap)
str(rongelap)
plotData(rongelap$data, rongelap$coords)
lines(rongelap$borders)
}
}

\seealso{
\code{\link{plotData}}.
}

\keyword{Data}
