% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCor.R
\name{runCor}
\alias{runCor}
\title{Correlation App}
\usage{
runCor()
}
\description{
This function allows the correlation shiny app to run. The app is a little
game where you are presented with a graph and you guess the correlation
between the two variables. The true correlation will then be shown and the
difference between your guess and the true correlation will be given
}
