% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_createHM_driftsign.r
\name{RW_createHM_driftsign}
\alias{RW_createHM_driftsign}
\title{RW_createHM_driftsign}
\usage{
RW_createHM_driftsign(
  mu,
  nsim = 10000,
  dt = 0.0025,
  intra_sv = 0.1,
  upperRT = 5,
  ev_bound = 0.5,
  ev_window = 0.01,
  HMtype = FALSE
)
}
\arguments{
\item{mu}{: vector of drift rates}

\item{nsim}{: number of paths generated}

\item{dt}{: time resolution}

\item{intra_sv}{: random walk standard deviation}

\item{upperRT}{: maximum time}

\item{ev_bound}{: maximum/minimum evidence}

\item{ev_window}{: evidence resolution}

\item{HMtype}{: Type of heatmap returned, default all: ('sym' -> symmetrical HM, 'upper' -> upper bound hit heatmap, 'lower' -> lower bound hit)}
}
\value{
Matrix (or list of matrices) describing the Time x Evidence heatmap(s)
}
\description{
Generates "empirical" heatmaps (HM) by simulating both positive and negative drift rate trajectories.
}
\author{
Pierre Le Denmat
}
