% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot using ggplot2}
\usage{
scatterplot(data, xvar, yvar, colvar = NULL)
}
\arguments{
\item{data}{A data frame containing the data to be plotted. Example: iris.}

\item{xvar}{A character string specifying the column name for the x-axis. Example: "Sepal.Length".}

\item{yvar}{A character string specifying the column name for the y-axis. Example: "Sepal.Width".}

\item{colvar}{A character string specifying the column name for the color of the points.
If NULL, all points will have the same color. Example: "Species".}
}
\value{
A ggplot2 scatterplot.
}
\description{
This function creates a scatterplot using ggplot2 based on the provided data and variable names.
}
\examples{
scatterplot(data = iris, xvar = "Sepal.Length", yvar = "Sepal.Width", colvar = "Species")

}
