% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScRNA.R
\name{ScRNA}
\alias{ScRNA}
\title{单细胞RNA数据分析处理}
\usage{
ScRNA(
  file.type = "10X",
  file = "GSM3729179/",
  multifile = FALSE,
  file_folder,
  project,
  organism = "Human",
  clusters = FALSE,
  reduction = "tsne",
  interaction.method = c("interaction", "harmony")
)
}
\arguments{
\item{file.type}{输入文件类型，可选"10X"、"h5"或"txt/csv/tsv"，默认"10X"}

\item{file}{文件路径/文件夹路径，10X数据需传文件夹路径}

\item{multifile}{是否处理多个文件，默认FALSE}

\item{file_folder}{多文件模式下的父文件夹路径}

\item{project}{项目名称标识}

\item{organism}{物种类型，可选"Human"或"Mouse"，默认Human}

\item{clusters}{是否进行差异表达分析，默认FALSE}

\item{reduction}{降维可视化方法，可选"tsne"，"umap"，“pca”，默认tsne，If not specified, first searches for umap, then tsne, then pca}

\item{interaction.method}{单细胞数据整合方法，c("interaction","harmony")}
}
\value{
处理后的Seurat对象
}
\description{
该函数用于处理单细胞RNA测序数据，包含数据标准化、特征选择和降维分析。
}
\author{
Pengmin Yang
}
