% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{SD}
\alias{SD}
\title{Calculate standard deviation}
\usage{
SD(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - standard deviation of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the standard deviation
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
SD(data.frame)
SD(matrix)
SD(vector)
SD(List)
}
