% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate standard error}
\usage{
se(s, n)
}
\arguments{
\item{s}{numeric value}

\item{n}{numeric value (integer)}
}
\value{
number
}
\description{
Returns the standard error
from standard deviation and sample size
}
\examples{
se(5,200)
}
