% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.sites2codons.R
\name{seg.sites2codons}
\alias{seg.sites2codons}
\title{Convert segregating sites to codons}
\usage{
seg.sites2codons(y)
}
\arguments{
\item{DNAbin}{object.}
}
\value{
matrix of amino acids
}
\description{
The function returns the codons for segregating sites.
}
\details{
Convert your segregating sites from dnabin to return AA change.
}
