% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM}
\alias{SEM}
\title{Calculate standard error mean}
\usage{
SEM(df)
}
\arguments{
\item{df}{Numeric 2 dimension table}
}
\value{
Good way: one number - standard error mean of whole input table

Bad way: -1 if input data are invalid
}
\description{
This function calculate the standard error mean
of whole input table.
}
\details{
The inputs can be also numeric vector.
}
\examples{
SEM(data.frame)
SEM(matrix)
SEM(vector)
SEM(List)
}
