% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_breaks.R
\name{set_breaks}
\alias{set_breaks}
\title{Generate axis breaks and limits for numeric data}
\usage{
set_breaks(
  values,
  breaks = TRUE,
  top_ratio = 1,
  seq_length = NULL,
  max_n_breaks = 16
)
}
\arguments{
\item{values}{Numeric vector. The data values for which axis breaks or limits are calculated.}

\item{breaks}{Logical. If TRUE, the function returns a sequence of breaks; if FALSE, it returns only the bottom and top limits.}

\item{top_ratio}{Numeric. A multiplier to extend the top of the axis range beyond the maximum data value (default is 1, no extension).}

\item{seq_length}{Numeric or NULL. Optional manual step size for axis breaks. If NULL, the function chooses a reasonable step automatically.}

\item{max_n_breaks}{Numeric. Maximum number of breaks (default is 16).}
}
\value{
If `breaks = TRUE`, a numeric vector of axis breaks. If `breaks = FALSE`, a numeric vector of length 2 with the bottom and top axis limits.
}
\description{
This function calculates nicely rounded axis breaks and limits for a given numeric vector.
It is designed to be used in plots where automatic and visually appealing axis scaling is desired.
}
\examples{
set_breaks(c(2, 7, 15))
set_breaks(c(2, 7, 15), top_ratio = 1.2, breaks = FALSE)

}
