% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff_intersect.R
\name{setdiff_intersect}
\alias{setdiff_intersect}
\title{Print set diff and intersection in up to 3 sets}
\usage{
setdiff_intersect(sets, labels = names(sets), values = FALSE)
}
\arguments{
\item{sets}{a list}

\item{labels}{labels for each set, defaults to set list names}

\item{values}{logical for whether or not to how set values}
}
\value{
print counts (and values) of distinctions and intersections
}
\description{
Print set diff and intersection in up to 3 sets
}
