% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDensity.R
\name{shinyDensity}
\alias{shinyDensity}
\title{Shiny density plot}
\usage{
shinyDensity(data, main = "Density plot", xlab = "log2(value + .001)",
  ylim = c(0, 0.5))
}
\arguments{
\item{data}{Data matrix}

\item{main}{The title for plot Defaults to \pkg{Density plot}}

\item{xlab}{The title used for the xlab of boxplot Defaults to \pkg{log2(value + .001)}}

\item{ylim}{The y-axis limitaiton for density plot Defaults to \pkg{c(0,0.5)}}
}
\value{
a local webpage for visualization of density
}
\description{
A simple \pkg{shiny} app to show density plot of given sample on local web browser.
}
\examples{
shinyDensity(data, ylim=c(0,0.5))
}
\author{
WKL
}
\keyword{density,}
\keyword{shiny}
