% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyHcluster.R
\name{shinyHcluster}
\alias{shinyHcluster}
\title{Shiny hierarchical clustering}
\usage{
shinyHcluster(data, traitNum = NULL, scale = FALSE, cutheight = 50000,
  minSize = 10)
}
\arguments{
\item{data}{Data matrix}

\item{traitNum}{Sample information which must be numeric for cluster plot Defaults to NULL}

\item{scale}{Whether scale is used for the data Defaults to FALSE}

\item{cutheight}{The init threshold for outliers Defaults to 50000}

\item{minSize}{The minimum cluster size for outliers Defaults to 10}
}
\value{
a local webpage for visualization of hcluster
}
\description{
A simple \pkg{shiny} app to show outliers based on hierarchical clustering on local web browser.
}
\examples{
shinyHcluster(data, trait_num, cutheight=1000)
}
\author{
WKL
}
\keyword{cluster,}
\keyword{shiny}
