% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyinit.R
\name{shinyinit}
\alias{shinyinit}
\title{shinyinit
 @description
 Starts an example shiny server in the browser. This is designed to show off
 some of the functionaltiy of the plot and print methods o}
\usage{
shinyinit()
}
\value{
starts the shiny example for mypackage
}
\description{
shinyinit
 @description
 Starts an example shiny server in the browser. This is designed to show off
 some of the functionaltiy of the plot and print methods o
}
\examples{
\dontrun{shinyinit()}

}
