% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPCA.R
\name{shinyPCA}
\alias{shinyPCA}
\title{Shiny PCA}
\usage{
shinyPCA(data, onetrait = NULL)
}
\arguments{
\item{data}{Data matrix}

\item{onetrait}{One sample information for plot Defaults to NULL}

\item{sdout}{The threshold for outliers Defaults to 2}
}
\value{
a local webpage for visualization of PCA
}
\description{
A simple \pkg{shiny} app to show outliers based on PCA on local web browser.
}
\examples{
shinyPCA(data, trait$Gender)
}
\author{
WKL
}
\keyword{PCA,}
\keyword{outliers}
