\docType{methods}
\name{show,DataM-method}
\alias{show,DataM-method}
\title{Show DataM Object}
\usage{
\S4method{show}{DataM}(object)
}
\arguments{
  \item{DataM}{An object of class DataM, or any of its
  subclasses}

  \item{object}{Any R object}
}
\description{
Modifies the "show" function to take objects of class
\code{DataM} (or any of its subclasses) and summarizes each
of the (in)dependent variables in the class
}
\author{
Thomas Carroll: \email{thomasscarroll89
}

