% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_Heatmap.R
\name{sig_Heatmap}
\alias{sig_Heatmap}
\title{Creates a 'InputHeatmap' object from 'tbl_df' on evaluation creates a 'ComplexHeatmap'}
\usage{
sig_Heatmap(
  input,
  ID = "ID",
  features,
  group,
  condiction = NULL,
  id_condiction = "vars",
  col_condiction = "condiction",
  scale = FALSE,
  palette = 2,
  palette_group = "jama",
  heatmap_col,
  show_col = F,
  show_palettes = F,
  cols_group = NULL,
  row_group = NULL,
  legend_show = TRUE,
  show_plot = T,
  width = 8,
  height = NULL,
  size_col = 10,
  size_row = 8,
  angle_col = 90,
  column_title = NULL,
  column_title_size = 13.2,
  row_title = NULL,
  row_title_size = 13.2,
  show_heatmap_col_name = F,
  name
)
}
\arguments{
\item{input}{data frame with 'ID', 'group' and features}

\item{ID}{This parameter specifies the column name in the input data that represents the unique identifier. The default value is "ID".}

\item{features}{This parameter is a vector of column names representing the features to be used in the heatmap.}

\item{group}{This parameter is a column name in the input data that represents the grouping variable.}

\item{condiction}{This parameter is an optional data frame that defines additional conditions to be applied in the heatmap. It should have two columns: one specifying the variables and another specifying the conditions. The default value is NULL.}

\item{id_condiction}{This parameter specifies the column name in the condiction data frame that represents the variables. The default value is "vars".}

\item{col_condiction}{This parameter specifies the column name in the condiction data frame that represents the conditions. The default value is "condiction".}

\item{scale}{This parameter specifies whether to scale the heatmap by column. If set to TRUE, the heatmap will be scaled by column. The default value is FALSE.}

\item{palette}{This parameter specifies the palette to be used in the heatmap. It can be a numeric value representing the palette index or a string specifying the palette name. The default value is 2.}

\item{palette_group}{This parameter specifies the palette group to be used in the heatmap. It should be a string representing the palette group name. The default value is "jama".}

\item{heatmap_col}{defalut heatmap_col=NULL. tidyHeatmap::heatmap参数palette_value}

\item{show_col}{This parameter specifies whether to show the color boxes in the heatmap. If set to TRUE, the color boxes will be shown. The default value is FALSE.}

\item{show_palettes}{This parameter specifies whether to show the available palettes. If set to TRUE, the available palettes will be shown. The default value is FALSE.}

\item{cols_group}{This parameter is a vector of colors to be used for the grouping cols variable. If not specified, the colors will be randomly assigned.}

\item{row_group}{This parameter is a vector of colors to be used for the grouping row variable. If not specified, the colors will be randomly assigned.}

\item{legend_show}{This parameter specifies whether to show the annotaion legend.The default value is TRUE.}

\item{show_plot}{This parameter specifies whether to show the resulting heatmap plot. If set to TRUE, the plot will be shown. The default value is TRUE.}

\item{width}{This parameter specifies the width of the resulting heatmap plot in inches. The default value is 8.}

\item{height}{This parameter specifies the height of the resulting heatmap plot in inches. If not specified, the height will be calculated based on the number of features.}

\item{size_col}{This parameter specifies the font size of the column names in the heatmap plot. The default value is 10.}

\item{size_row}{This parameter specifies the font size of the row names in the heatmap plot. The default value is 8.}

\item{angle_col}{This parameter specifies the rotation angle of the column names in the heatmap plot. The default value is 90.}

\item{column_title}{This parameter specifies the title of the column in the heatmap plot. If not specified, no title will be displayed.}

\item{column_title_size}{This parameter specifies the title size of the column in the heatmap plot.}

\item{row_title}{This parameter specifies the title of the row in the heatmap plot. If not specified, no title will be displayed.}

\item{row_title_size}{This parameter specifies the title size of the row in the heatmap plot.}

\item{show_heatmap_col_name}{This parameter specifies whether to show the column names in the heatmap plot. If set to TRUE, the column names will be shown. The default value is FALSE.}

\item{name}{This parameter is the title of heatmap_col.}
}
\value{
A heatmap plot object.
}
\description{
heatmap() takes a tbl object and easily produces a ComplexHeatmap plot, with integration with tibble and dplyr frameworks.
}
\examples{
data("input1")
data("condiction")

feas1<-colnames(input1)[3:102]
sig_Heatmap(input = input1, features = feas1,ID ="SAMPLE_ID",show_plot=F,
            condiction=condiction,id_condiction=colnames(condiction)[[1]],col_condiction=colnames(condiction)[[2]],
             cols_group=c("#757575","#FF4040"),row_group=c("red","green"),
             legend_show=TRUE,column_title_size=10,row_title_size=8,
             heatmap_col=NULL,
             #heatmap_col=c("#0505FA", "#FFFFFF", "#FA050D"),
             group = "PREX2",row_title="Regulate", scale = TRUE,name="Expression")

}
\author{
Pengmin Yang
}
