% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array}
\alias{simplify2array}
\title{simplify2array}
\usage{
simplify2array(x)
}
\arguments{
\item{x}{- List of numeric vectors}
}
\value{
Simplified matrix
}
\description{
Function mimics behavior of simplify2array from base when higher = T
}

