% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array2}
\alias{simplify2array2}
\title{simplify2array2(List x) function}
\usage{
simplify2array2(x)
}
\value{
A matrix which columns are equivalent with elements of a imput list.
}
\description{
Function simplify2array2() converts a list containing numeric vectors into a matrix.
}

