% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array_Rcpp}
\alias{simplify2array_Rcpp}
\title{simplify2array_Rcpp}
\usage{
simplify2array_Rcpp(x)
}
\description{
This function mimics the behavior of the simplify2array fucntion as close as possible.
}
\details{
ARGUMENTS:
x - a list numeric values

RETURN VALUE:
list or matrix
}
\seealso{
Other fungroup: \code{\link{mode_Rcpp}}
}

