% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_into_days.R
\name{slice_into_days}
\alias{slice_into_days}
\title{Slice the time series data into days before cleaning}
\usage{
slice_into_days(inputId, startTime, endTime)
}
\arguments{
\item{inputId}{A string of he unique identifier}

\item{startTime}{A integer indicates the starting time for plotting}

\item{endTime}{A integer indicates the ending time for plotting}
}
\value{
A table removed repeated entries in time domain
}
\description{
Slice the time series data into days before cleaning
}
