% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{sliceTime}
\alias{sliceTime}
\title{Slice the time series into shorter time period}
\usage{
sliceTime(inputId, specific_date, startTime, endTime)
}
\arguments{
\item{inputId}{A string of he unique identifier}

\item{sliceSize}{The size for slicing}
}
\value{
A table containing the Glucose Value and Internal Time for that ID
}
\description{
Slice the time series into shorter time period
}
