% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_multiples_data.R
\name{small_multiples_data}
\alias{small_multiples_data}
\title{Prepare Data for Small Multiples Plotting}
\usage{
small_multiples_data(data, column)
}
\arguments{
\item{data}{A data.frame or tibble containing the data to be used.}

\item{column}{A column name (unquoted) in `data` to generate small multiples by.}
}
\value{
A tibble where each row of the original dataset is replicated for each
        unique value of the specified column, with a new column `highlighted`
        indicating the current value.
}
\description{
This function takes a dataset and a column, and creates a new dataset
suitable for small multiples plotting. Each unique value of the specified
column is used to create a copy of the dataset, with a new column `highlighted`
indicating the value to be highlighted.
}
\examples{
library(dplyr)
library(purrr)
df <- tibble(group = c("A","B"), value = 1:2)
small_multiples_data(df, group)

}
