% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithPareto.r
\name{smithPareto}
\alias{smithPareto}
\title{Build Pareto graph}
\usage{
smithPareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{vector}
}
\description{
This function will take in a vector and produce a Pareto barplot
}
\examples{
smithPareto(c(1,2,3,4,5,6,7))


}
