% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smrzVector.R
\name{smrzVector}
\alias{smrzVector}
\title{Summarize A Vector}
\usage{
smrzVector(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
a vector of summary values
}
\description{
Produce standard summary measures for a numeric vector.
}
