% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical_functions.R
\name{soc}
\alias{soc}
\title{Sum of cubes computation}
\usage{
soc(list)
}
\arguments{
\item{list}{A list containing elements of integer type}
}
\value{
The sum of cubes of all the elements in the list
}
\description{
Sum of cubes computation
}
\examples{
sumOfCubes <- soc( c(1, 2, 3) );
}
