% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{soma}
\alias{soma}
\title{As quatros operacoes desagregadas}
\usage{
soma(a, b)
}
\arguments{
\item{a}{primeiro numero}

\item{b}{segundo numero}
}
\value{
O "soma", a "subtra", a "multi" e a "divisi" de a e b
}
\description{
O somatorio de a e b
}
\details{
Calculo aritmetico simples
}
\examples{
soma (2, 2)
}
\keyword{a,}
\keyword{b}
\keyword{soma,}
