% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_rightleft.R
\name{sort_rightleft}
\alias{sort_rightleft}
\title{Left-right order}
\usage{
sort_rightleft(x, i)
}
\arguments{
\item{i}{An integer between 1 and \code{n}.}

\item{n}{A positive integer.}
}
\description{
Order the numbers 1 through \code{n}, starting at index \code{i} and
proceeding rightward then leftward.
}
