% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_n_characters.R
\name{split_n_characters}
\alias{split_n_characters}
\title{Split string every n characters}
\usage{
split_n_characters(x, n = 3)
}
\arguments{
\item{x}{Character vector}

\item{n}{Integer - number of characters to split the string(s) on.
Defaults to 3.}
}
\value{
List of character vectors broken into chunks of n characters
}
\description{
Function splits a string every n characters and returns a list.
}
\examples{
split_n_characters('ACTGACGGG')
split_n_characters('This is a sentence.', n = 2)
}
