% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/power.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numneric value to squared}
}
\value{
The square of the input
}
\description{
Takes in any numeric value and squares it.
}

