% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:standarderror.R
\name{standard_error}
\alias{standard_error}
\title{Calculate Standard Error}
\usage{
standard_error(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing standard error.
}
\description{
Calculate Standard Error
}
\examples{
standard_error(c(1,2,3,4,5))
standard_error(c(2,3,5,6,7,8,9))
}
