% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normf_standf.R
\name{standf}
\alias{standf}
\title{Standardize a numeric vector}
\usage{
standf(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Logical value determining whether or not to remove NA values
when calculating standard deviation and mean. Defaults to FALSE.}
}
\value{
Numeric vector that is standardized
}
\description{
Takes a numeric vector and standardizes it based on standard deviation
and mean.
}
\examples{
standf(c(1, 2, 3, 4, 5))
standf(c(1, 2, 3, 4, 5), na.rm = TRUE)
}
