% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statename.R
\name{stateabb}
\alias{stateabb}
\title{State abbreviation transfer(abbr to full name)}
\usage{
stateabb(x)
}
\arguments{
\item{x}{a string, vector: the full name to be transferred to abbreviation.}
}
\value{
the abbreviation of U.S. states
}
\description{
Transfer the full names of 50 U.S. state to corresponding abbreviations.
}
\examples{
1. stateabb("Washington");
   stateabb("WASHINGTON");
2. x <- c("New York", "Virginia")
   stateabb(x)
}
