% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statename.R
\name{statefull}
\alias{statefull}
\title{State abbreviation transfer(full name to abbr)}
\usage{
statefull(x)
}
\arguments{
\item{x}{a string or vector:the values to be transferred.}
}
\value{
full name of U.S. states.
}
\description{
Transfer the abbreviation of 50 U.S. state to full name.
}
\examples{
statefull("CA")
}
