% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{std_err}
\alias{std_err}
\title{Calculate the standard error for a sample}
\usage{
std_err(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
returns the standard error of
a numeric vector
}
\examples{
std_err(rnorm(50))
}
