% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stewartPareto.r
\name{stewartPareto}
\alias{stewartPareto}
\title{Build Pareto graph}
\usage{
stewartPareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{vector}
}
\description{
This function will take in a vector and produce a Pareto barplot
}
\examples{
smithPareto(c("fire","fire","heart","heart","heart"))


}
