% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_duplicated.R
\name{string_duplicated}
\alias{string_duplicated}
\title{String Duplication/Uniqueness Checker}
\usage{
string_duplicated(string)
}
\arguments{
\item{string}{Arguemnt will be the string entered. You can directly run the function.}
}
\description{
This function allows you to enter a string and checks if your string has all unique characters or not.
}
\author{
Melih Ozbek
string_duplicated()
}
