% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{sub}
\alias{sub}
\title{Make the summary table for ID, percentage of missing data, average gap size, number of days and equal spaces.}
\usage{
sub(frequentTimeGap)
}
\arguments{
\item{frequentTimeGap}{A set}

\item{timeGap}{a subset}
}
\value{
A summary table containing summary data
}
\description{
Make the summary table for ID, percentage of missing data, average gap size, number of days and equal spaces.
}
