% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subcars}
\alias{subcars}
\title{subcars}
\format{
a dataframe of 38 observations and 2 variables
\describe{
\item{speed}{djksaljdlsa}
\item{dist}{sdaskjdasl}
}
}
\source{
the cars dataset
}
\usage{
subcars
}
\description{
a subset of cars dataset
}
\keyword{datasets}
