% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublist_vcfbypop.R
\name{sublist_vcfbypop}
\alias{sublist_vcfbypop}
\title{List vcfR by population}
\usage{
sublist_vcfbypop(vcf, pop)
}
\arguments{
\item{vcf}{a vcfR object.}

\item{pop}{population assignment.}
}
\value{
subsetted vcfR.
}
\description{
The function subsets vcfR object by population of interest to a list.
}
\details{
The function subsets vcfR object by population of interest.
}
