% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vcf2vcf.R
\name{subset_vcf2vcf}
\alias{subset_vcf2vcf}
\title{Subset a vcf file}
\usage{
subset_vcf2vcf(large.vcf, small.vcf)
}
\arguments{
\item{Larger}{vcfR object.}

\item{Smaller}{vcfR object.}
}
\value{
subsetted vcf file by small vcfR variants
}
\description{
The function subsets the larger vcf file by variants from other smaller vcf file.
}
\details{
Subset the vcfR object given the variants from smaller vcfR object
}
