% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{subtra}
\alias{subtra}
\title{As quatro operacoes desagregadas}
\usage{
subtra(a, b)
}
\arguments{
\item{a}{primeiro numero}

\item{b}{segundo numero}
}
\value{
Retorna a subtracao de a e b
}
\description{
A subtracao de a e b
}
\details{
Calculo aritmetico simples
}
\keyword{a,}
\keyword{b}
\keyword{subtracao,}
