% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{subtract}
\alias{subtract}
\title{Subtraction Function}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{This is the from which we want to subtract}

\item{y}{This is the value we will subtract}
}
\value{
Returns the difference of x and y
}
\description{
This function generates and returns the difference of x and y.
}
\examples{
### First you can subtract small numbers
subtract(1,1)

### Now you subtract add big numbers
subtract(222232, 2234234233)

}
