% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calc sum of suqares}
\usage{
sum_squares(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
Calc sum of suqares
}
\examples{
sum_squares(rnorm(20))
}
