% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_square_function.R
\name{sum_squares_function}
\alias{sum_squares_function}
\title{Calculate the sums of squares for a sample}
\usage{
sum_squares_function(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
returns the sum of the squared deviations
}
\details{
from mean, sums of squares(x) of a numeric vector
}
\examples{
sum_squares_function(rnorm(50))
}
