% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable.R
\name{summaryTable}
\alias{summaryTable}
\title{Make the summary table for ID, percentage of missing data, average gap size, number of days and equal spaces.}
\usage{
summaryTable(data)
}
\arguments{
\item{data}{An input dataset}
}
\value{
A summary table containing summary data
}
\description{
Make the summary table for ID, percentage of missing data, average gap size, number of days and equal spaces.
}
