% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumOfCubes.R
\name{sumofcubes}
\alias{sumofcubes}
\title{Sum of cubes}
\usage{
sumofcubes(x, y)
}
\arguments{
\item{x}{first value}

\item{y}{second value}
}
\description{
Returns the sum of the cube of twon given arguments
}
\examples{
val <- sumofcubes(3, 4)
}
