% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{svm_classification}
\alias{svm_classification}
\title{使用支持向量机进行分类}
\usage{
svm_classification(train, test)
}
\arguments{
\item{train}{训练数据集}

\item{test}{测试数据集}
}
\value{
预测的分类结果和混淆矩阵
}
\description{
这个函数使用SVM算法对数据进行分类。
}
