\name{test_fun}
\alias{test_fun}
\title{Squaring and adding}
\description{Squares a number and adds it to itself}
\usage{
test_fun(x)
}
\arguments{
  \item{x}{The value to be squared and added}
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
  An object of class numeric representing the result of squaring and adding a number to itself
}
\author{Rob Foy}
\examples{
val <- 11
sqadd <- test_fun(val)
sqadd
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{squaring}
