% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfx1.R
\name{testfx1}
\alias{testfx1}
\title{add two numbers}
\usage{
testfx1(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
a sum of two numbers
}
\description{
add two numbers
}
\examples{
x<-1
y<-2
testfx1(x,y)

}
