% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfx2.R
\name{testfx2}
\alias{testfx2}
\title{Calculate MF}
\usage{
testfx2(dat, mut_col, total_col)
}
\arguments{
\item{dat}{is a dataframe describing the number of mutations in a sample. 'sample' is the sample name. 'mut_depth' is the number of mutations in the sample. 'total_depth' is the total number of bases sequenced for the sample.}

\item{mut_col}{Name of the mutation depth column}

\item{tot_col}{Name of the total depth column}
}
\value{
This function will output the dataframe with an added column MF that contains the calculated mutation frequency for each sample.
}
\description{
Calculate MF
}
