\name{testTPS-package}
\alias{testTPS-package}
\alias{testTPS}
\docType{package}
\title{\packageTitle{testTPS}}
\description{\packageDescription{testTPS}}
\section{Package Content}{\packageIndices{testTPS}}
\author{\packageAuthor{testTPS}}
\section{Maintainer}{\packageMaintainer{testTPS}}
\keyword{package}
