% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Translate}
\alias{Translate}
\title{Translate: Translate function.}
\usage{
Translate(seq = c(s1 = "CTATATAGAGAGA", s2 = "ACACACTGACGAT"))
}
\arguments{
\item{seq}{a vector of DNA strings.}
}
\description{
This function will translate one or many DNA sequences in all three reading frames into proteins.
}
\examples{
hi <- c("TTGGGGGCGTTAG", "CATTGGTTCTAGTT", "GCATTGGTGTGCATTAG")
Translate(seq=hi)
seq <- c(s1="CTATATAGAGAGAGGAGAAATCTCT", s2="ACACACTGACGATTGTGACACAGATC")
Translate(seq=seq)
}
\author{
Brendan Gongol
}
