% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{tree_classification}
\alias{tree_classification}
\title{使用决策树进行分类}
\usage{
tree_classification(train, test)
}
\arguments{
\item{train}{训练数据集}

\item{test}{测试数据集}
}
\value{
预测的分类结果和混淆矩阵
}
\description{
这个函数使用决策树算法对数据进行分类。
}
