% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_sample_data.R
\docType{data}
\name{ts_sample_data}
\alias{ts_sample_data}
\title{ts_sample_data}
\format{
249 x 2 data frame

\describe{
\item{date}{time horizon from 2000 to 2020}
\item{value}{the monthly sales of Beer, Wine and Liquor in the U.S.}
}
}
\usage{
ts_sample_data
}
\description{
Sample data for TS trend model visualization function.
}
\keyword{datasets}
