% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFactorial.R
\name{Vars}
\alias{Vars}
\alias{colVars}
\alias{rowVars}
\title{Calculate column variances

Form row and column variances for numeric arrays}
\usage{
colVars(x)

rowVars(x)
}
\arguments{
\item{x}{a 2 dimensional numerical array}
}
\value{
A numeric or complex vector, of length equal to the number of columns (\code{colVars}) or rows (\code{rowVars}) of x. Each element contains the variance of the respective column or row.
}
\description{
Calculate column variances

Form row and column variances for numeric arrays
}
\examples{
mat <- matrix(rnorm(9), nrow = 3)
colVars(mat)
rowVars(mat)
}
