% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2nexus.R
\name{vcf2nexus}
\alias{vcf2nexus}
\title{Convert vcfR object to nexus format data..}
\usage{
vcf2nexus(vcf, file = "file.nex")
}
\arguments{
\item{vcf}{a vcfR object.}

\item{outfile}{name of output file.}
}
\value{
nexus format file.
}
\description{
The function converts vcfR format data to nexus format.
}
\details{
This function converts the vcfR object to a nexus format file.
}
