% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2QTLseq.R
\name{vcf2QTLseq}
\alias{vcf2QTLseq}
\title{Convert vcfR object to QTLseqR format.}
\usage{
vcf2QTLseq(vcf)
}
\arguments{
\item{vcfR}{object.}
}
\value{
text file
}
\description{
The function converts vcf file to text format file that can be used by QTLseqR package.
}
\details{
Converts vcfR data format to text file to be used by QTLseqR.
}
