% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2SNAPP.R
\name{vcf2SNAPP}
\alias{vcf2SNAPP}
\title{Convert vcfR object to SNAPP format data}
\usage{
vcf2SNAPP(vcf, file = "snapp.nex")
}
\arguments{
\item{vcf}{a vcfR object.}

\item{outfile}{name of output file.}
}
\value{
SNAPP nexus format file.
}
\description{
The function converts vcfR format data to SNAPP nexus format.
}
\details{
This function converts the vcfR object to a SNAPP nexus format file.
}
