% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_popsub.R
\name{vcf_popsub}
\alias{vcf_popsub}
\title{Subset vcfR by population}
\usage{
vcf_popsub(vcf, pop, in_pop)
}
\arguments{
\item{vcf}{a vcfR object.}

\item{pop}{population assignment.}

\item{in_pop}{population of interest.}
}
\value{
subsetted vcfR.
}
\description{
The function subsets vcfR object by population of interest.
}
\details{
The function subsets vcfR object by population of interest.
}
