% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfR2migrate.R
\name{vcfR2migrate}
\alias{vcfR2migrate}
\title{Convert vcfR object to MigrateN format}
\usage{
vcfR2migrate(vcf, pop, in_pop, out_file)
}
\arguments{
\item{vcf}{a vcfR object.}

\item{pop}{factor of population.}

\item{in_pop}{population of interest to report in output file.}

\item{outfile}{name of output file.}
}
\value{
a text file that can be used as an input for MigrateN software (SNP format).
}
\description{
The function converts vcfR format data to text format that can be used as infile for MigrateN.
}
\details{
This function converts the vcfR object to a text file which can be used downstream for MigrateN.
The function will remove loci with missing data, indels, and multiallelic loci. Thus, only SNPs data are analysed
where length of each locus is 1.
}
