% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa_chart.R
\name{vespa_chart}
\alias{vespa_chart}
\title{Bar chart of observations per year}
\usage{
vespa_chart(data)
}
\arguments{
\item{data}{A tibble created by the function \code{\link{vespa_data}}}
}
\value{
A bar chart of observations per year
}
\description{
\code{vespa_data} creates a bar chart of observations per year.
}
\examples{
# First, create a tibble with observations with the function vespa_data
data <- vespa_data(2019,2021)
# Then, create a chart
vespa_chart(data)

# Customize the chart with ggplot2
\dontrun{
vespa_chart(data) +
ggtitle("Number of observations of Vespa velutina in Flanders during 2019-2021")}
}
