% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa_viz.R
\name{vespa_data}
\alias{vespa_data}
\title{Vespa invasion visualization}
\usage{
vespa_data(year1, year2 = NULL)
}
\arguments{
\item{x}{}
}
\value{
The output from \code{\link{print}}
}
\description{
Vespa invasion visualization
}
