% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa_map.R
\name{vespa_map}
\alias{vespa_map}
\title{Map observations}
\usage{
vespa_map(data)
}
\arguments{
\item{data}{A tibble created by the function \code{\link[=vespa_data]{vespa_data()}}}
}
\value{
A map
}
\description{
\code{vespa_map} creates a map of observations in Flanders.
}
\examples{
# First, create a tibble with observations with the function vespa_data
data <- vespa_data(2019,2021)
# Then, create a map
vespa_map(data)

# Customize the map with ggplot2
\dontrun{
vespa_map(data) +
ggtitle("Map of observations of Vespa velutina in Flanders during 2019-2021")}
}
