% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-2019-2021.R
\docType{data}
\name{vespadata_2019_2021}
\alias{vespadata_2019_2021}
\title{Dataset of observations of Asian hornets in Flanders during 2019-2021}
\format{
A dataset
}
\usage{
vespadata_2019_2021
}
\description{
This dataset are occurrences of Vespa velutina in Flanders during 2019 - 2021,
retrieved on 03/02/2024 from the Global Biodiversity Information Facility (GBIF).
}
\keyword{datasets}
