% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VI.ineq.draws}
\alias{VI.ineq.draws}
\title{VI bound using Jensen inequality}
\usage{
VI.ineq.draws(PYclust, parti = PYclust, print.bar = TRUE)
}
\description{
Computes a bound for expected VI loss and the partition minimizing it
}
