% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{waves}
\alias{waves}
\title{Storm Peak Significant Wave Heights in the Northern North Sea}
\format{
This data frame contains the following numeric variables:
\itemize{
\item \code{year}: year in which the storm peak occurred.
\item \code{month}: month in which the storm peak occurred.
\item \code{day}: day in which the storm peak occurred.
\item \code{Hs}: storm peak significant wave height, in metres.
\item \code{direction}: dominant wave direction, a measure of the direction from
which the storm is coming, in degrees from North, increasing clockwise.
\item \code{season}: day of the year, for a standardised year of 360 days.
\item \code{NAO}: the NOAA's daily North Atlantic Oscillation index for the day on
which the storm peak occurred.
}
}
\source{
\href{http://merlot.stat.uconn.edu/~jyan/docs/evrisk_r/evrisk.html}{Computer Code Supplement}
to Chapter 5 of Dey and Yan (2016). The daily NAO data were sourced from the
\href{https://www.cpc.ncep.noaa.gov/products/precip/CWlink/pna/nao.shtml}{NOAA}
(National Oceanic and Atmospheric Administration).
}
\usage{
waves
}
\description{
Data used for extreme value regression modelling in Chapter 5 of Dey and
Yan (2016). The raw data are hindcast time series for significant wave
height \eqn{H_S}, dominant wave direction \eqn{d} and seasonal degree
day \eqn{s} for three-hour sea states, at an undisclosed location in the
Northern North Sea, for the period September 1957 to December 2011.
}
\examples{
# Summaries of data
head(waves)
tail(waves)
summary(waves)

# Plots of Hs vs potential explanatory variables
plot(waves$season, waves$Hs)
plot(waves$NAO, waves$Hs)
plot(waves$direction, waves$Hs)
}
\references{
Dey, D.K. and Yan, J. (Eds.). (2016). Extreme Value Modeling and
Risk Analysis: Methods and Applications (1st ed.). Chapman and Hall/CRC.
\doi{10.1201/b19721}
}
\keyword{datasets}
