% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztest.R
\name{ztest}
\alias{ztest}
\title{Calculates the z score and significance}
\usage{
ztest(value, hypo)
}
\arguments{
\item{value}{numeric vector}

\item{hypo}{numeric value to test against}
}
\value{

}
\description{
Calculates the z score and significance
}
\examples{
ztest(rnorm(50), 0)
}
