% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Addition Function!}
\usage{
add(x, y)
}
\arguments{
\item{x}{This is the first value to add}

\item{y}{This is the second value to add}
}
\value{
Returns the sum of x and y
}
\description{
This function generates and returns the sum of x and y.
}
\examples{
### First you can add small numbers
add(1,1)

### Now you can add big numbers
add(222232, 2234234233)

}
