% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addValue.R
\name{addValue}
\alias{addValue}
\title{Addition}
\usage{
addValue(x, y)
}
\arguments{
\item{x}{One value}

\item{y}{Another value}
}
\value{
numeric value
}
\description{
A function that adds values.
}
\details{
there are no details to this
}
\examples{
val <- addValue (1, 2)

}
