% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Title}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character vector of length 1 indicating the animal}

\item{sound}{A character vector of length 1 indicating the sound the animal makes}
}
\value{
A character string giving a sentence with animal and sound inserted.
}
\description{
Title
}
\details{
For more information on paste0, see: \code{\link[base:paste]{base::paste0()}}
}
\examples{
x <- "lion"
y <- "roar"
animal_sounds(x, y)

}
