% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Create a String Stating the Sound an Animal makes}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A string giving the name of the animal.}

\item{sound}{A string giving the sound of the animal.}
}
\value{
A character string.
}
\description{
Given strings specifying an animal and a corresponding sound, create a
string of the form "The <animal> goes <sound>!".
}
\details{
The \code{\link[base:paste]{base::paste0()}} function is used to create the string from the arguments
\code{animal} and \code{sound}.
}
\examples{
animal_sounds("dog", "woof")
}
