% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Make animal sounds}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string that is the name of an animal}

\item{sound}{A character string that is the sound an animal makes}
}
\value{
A character string that is a sentence telling us what sound the given animal makes
}
\description{
This function creates a sentence for a given animal and sound.
}
\examples{

animal_sounds(animal = "bee", sound = "buzz")

"The bee says buzz, buzz!"

}
