% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainbody}
\alias{brainbody}
\title{Brainbody Data Set}
\format{
A data frame with 2 variables: \code{brainbody} and \code{days}.
}
\usage{
brainbody
}
\description{
This data set contains information on to variables. Where 1 is the brain weight in grams, and the next is body weight in kilograms. Note that the gestation length is in days.
}
\keyword{datasets}
