% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainbody}
\alias{brainbody}
\title{Brainbody Data Set}
\format{
A data frame with 5 variables: \code{species}, \code{brain}, \code{body}, \code{gestation}, \code{litter}.
}
\usage{
brainbody
}
\description{
This data set contains information on different animal's physical attributes and their brain sizes.
}
\keyword{datasets}
