% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainbody}
\alias{brainbody}
\title{Brain Body Data Set}
\format{
A data frame with 5 variables: \code{species}, \code{brain},
\code{body}, \code{gestation}, and \code{litter}.
}
\usage{
brainbody
}
\description{
This data set contains information on the brain weight and body weight of
various species.
}
\keyword{datasets}
