% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainbody}
\alias{brainbody}
\title{Brainbody Data Set}
\format{
A data frame with 2 variables: \code{species} , \code{brain} , \code{body} ,
\code{gestation} , and \code{litter} .
}
\usage{
brainbody
}
\description{
This data set contains information regarding the mass/size of an animal
versus the mass/size of its' brain. It also contains information regarding the
gestation time and litter size of the species.
}
\keyword{datasets}
