% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_format_date.R
\name{cd_format_date}
\alias{cd_format_date}
\title{cd_format_date}
\usage{
cd_format_date(date, lang = default(rmarkdown::metadata$lang, "de_CH"))
}
\arguments{
\item{date}{...}

\item{lang}{...}
}
\description{
FIXME
}
