% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_knit_chunk_opts.R
\name{cd_knit_chunk_opts}
\alias{cd_knit_chunk_opts}
\title{cd_knit_chunk_opts}
\usage{
cd_knit_chunk_opts(
  twocolumn = default(rmarkdown::metadata$twocolumn, FALSE),
  fig.width = NULL,
  fig.height = NULL,
  fig.pos = "h",
  cache = FALSE,
  message = FALSE,
  echo = FALSE,
  tidy = FALSE
)
}
\arguments{
\item{twocolumn}{...}

\item{fig.width}{...}

\item{fig.height}{...}

\item{fig.pos}{...}

\item{cache}{...}

\item{message}{...}

\item{echo}{...}

\item{tidy}{...}
}
\description{
FIXME
}
\details{
FIXME: Rename fig.width to fig_width, etc.?
}
