% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Plots a density plot}
\usage{
density_plot(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
plot
}
\description{
Plots a density plot
}
\examples{
density_plot(rnorm(50))
}
