% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{This is my division function}
\usage{
divide(x, y)
}
\arguments{
\item{x}{this is the first value to divide}

\item{y}{this is the second value to divide}
}
\value{
This function returns x divided by y
}
\description{
This is my division function
}
\examples{
## Start with something simple
divide(4,2)

## Now something more difficult
divide(105,21)

}
