% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{This is my multiplication function}
\usage{
divide(x, y)
}
\arguments{
\item{x}{this is the value for numerator}

\item{y}{this is the value for denominator}
}
\value{
This function returns the product of x and y
}
\description{
This is my multiplication function
}
\examples{
## Start with something simple
divide(1,1)

## Now something more difficult
divide(49,60)

}
