% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Dataset of Farm Animal Sounds}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{animal}{The animal as a character string.}
\item{sound}{The corresponding sound.}
}
}
\usage{
farm_animals
}
\description{
A dataset of farm animals and the sound they make
}
\keyword{datasets}
