% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farm_animals.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm animals and the sounds they make}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{animals}{The animal as a character string.}
\item{sounds}{The sound as a character string.}
}
}
\usage{
farm_animals
}
\description{
A dataset of strings of five farm animals along with strings of their corresponding sounds.
}
\keyword{datasets}
