% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An Integer with the State Code}

\item{year}{A string, or an integer, with the input year}
}
\value{
A plot showing a state map with the accidents location by year

fars_map_state(49, 2015)
}
\description{
A funcion to make a plot with a state map showing the accidents location by year
If the state.num is invalid the function shows an error
}
