% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the yearly accidents of a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric representing the state number}

\item{year}{A numeric or character representing the year included in the filename}
}
\value{
A map of the specified state with accident locations represented as dots.
}
\description{
This function fetches the location data of accident that happened in a state in a certain year.
It returns a map displaying the spatial distribution of the accidents.
}
\note{
If the specified state number or the year is not represented in the data the function raises an error.
}
\examples{
fars_map_state(4, 2013)
fars_map_state(3, 2013)
fars_map_state(4, 2019)

}
