% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Make multiple file names.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric or character vector (atomic or list) representing the years from which data must be read.}
}
\value{
A list of filenames.
}
\description{
This function takes a vector of years and returns a list of tibbles representing the data for the specified year.
}
\note{
This function applies the function 'make_filename' and fetches the according data from a vector of inputs using 'lapply'.
If a year is not found in the input of years, the function will raise a warning: "invalid year: \if{html}{\out{<year>}}". The return list entry for that year will be NULL.
}
\examples{
fars_read_years(c(2013, 2014, 2015))
fars_read_years(c(2013, 2014, 2019))

}
