% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from the fire station in miles}
  \item{DAMAGE}{x 1000 dollars in damage}
}
}
\usage{
fire
}
\description{
A data set describing damage from fire
}
\details{
This is a standard data set from the course MATH 4753 taken from the data sets provided by the text book.
}
\keyword{datasets}
