% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Reads Fire Data in R}
\format{
A data frame with 15 rows and 2 variables:
\describe{
 \item{DISTANCE}{distance from fire station in miles}
 \item{DAMAGE}{X 1000 dollars in damage}
}
}
\source{
\url{https://www.crcpress.com/statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
Reads fire data so that we can see it by accessing the package
}
\keyword{datasets}
