% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fitting the dataset Diamonds to a linear model}
\usage{
fit_model(data = ggplot2::diamonds)
}
\arguments{
\item{data}{simply diamonds free dataset}
}
\description{
Fitting the dataset Diamonds to a linear model
}
\details{
This function takes the diamonds dataset and fits a linear model where price is the dependent
variable and carat is the independent variable. By default it takes the diamonds dataframe because
in the inside of the function I have determined the target and independent feature, but theoretically
it can take any dataframe as an input. It also does not allow the user to put anything else, but a dataframe,
otherwise it will pop an error message.
}
\examples{
fit <-fit_model()
}
