% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{Ftest}
\alias{Ftest}
\title{Fit a multiple linear regression}
\usage{
Ftest(X, Y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the parameter estimates from a multiple regression model
}
\description{
Fit a multiple linear regression
}
\examples{
data(mtcars)
Ftest(mpg,wt)
}
