% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello world!}
\usage{
hello(name = NULL)
}
\arguments{
\item{name}{character string. Insert your name here.}
}
\value{
It prints a hello greeting to the console, said by a randomly selected animal from the cowsay function.
}
\description{
You always need a space between the title and any extra description
this function prints a hello greeting
}
\examples{
hello()
hello("cats")
}
