% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(name = NULL)
}
\arguments{
\item{name}{character string. Eve!}
}
\value{
prints hello greeting to console from me!
}
\description{
Print hello greeting
}
\examples{
hello()
hello("Lottielocks")
}
