% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(name = NULL)
}
\arguments{
\item{name}{character string. Your name.}
}
\value{
prints hello greeting to console from me.
}
\description{
Print personalised hello world greeting from me.
}
\examples{
hello()
hello("Lucy Elen")
}
