% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(name = NULL)
}
\arguments{
\item{name}{A character string. Yoiur name.}
}
\value{
Prints greeting from me to console.
}
\description{
Prints personalised hello greeting from me.
}
\examples{
hello(name="Tony")
}
