% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(name = NULL)
}
\arguments{
\item{name}{character string. Your name}
}
\value{
prints hello greeting to console from me by a crazy animal.
}
\description{
Further description as the line above has to be blank. This code prints hello
}
\examples{
hello()
hello("y'all")
}
