% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(name = NULL)
}
\arguments{
\item{name}{character string. A name to say hello to!}
}
\value{
prints hello greeting to the console from me
}
\description{
Prints hello greeting
}
\examples{
hello()
hello(name = "Rosie")
}
