% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{This is my hello function. There are no parameters.}
\usage{
hello()
}
\value{
This function returns the message "hello world".
}
\description{
This is my hello function. There are no parameters.
}
\examples{
## This is the only thing this function does.
hello()

}
