% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{kmp}
\alias{kmp}
\title{Motivated by HW1, realize KMP algorithm to find times of appearance of x in y, especially for quite long vectors matching}
\usage{
kmp(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
times of appearance of \code{x} in \code{y}
}
\description{
Motivated by HW1, realize KMP algorithm to find times of appearance of x in y, especially for quite long vectors matching
}
\examples{
kmp(c(1,2),c(1,2,1,2,3))
kmp(c(1,1),c(1,1,1,1,1))
}
