% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name
This funciotn makes .csv data file name taking into account the given.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer with the input}
}
\value{
This function returns a string with the data file name for a given
  year, and the file path within the package.
}
\description{
Make file name
This funciotn makes .csv data file name taking into account the given.
}
\examples{
make_filename(2013)

}
