% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or character representing the year included in the filename}
}
\value{
A character string representing the file name of an available file
}
\description{
This function takes a year and returns a character string that can be used as a filename for a .csv file.
The format of the returned filename is "accident_\if{html}{\out{<year>}}.csv".
}
\note{
This function is not exported because it is a helper function for 'fars_read_years'.
}
\examples{
make_filename(2013)
make_filename(2014)
make_filename(2019)

}
