% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{mlr}
\alias{mlr}
\title{Fit a multiple linear regression}
\usage{
mlr(X, Y)
}
\arguments{
\item{X}{design matrix}

\item{Y}{outcome}
}
\value{
the parameter estimates from a multiple regression model
}
\description{
Fit a multiple linear regression
}
\examples{
data(mtcars)
attach(mtcars)
mlr(model.matrix(wt ~ mpg + cyl), as.vector(wt))
detach(mtcars)
}
