% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{a numeric value}

\item{y}{a numeric value}
}
\value{
a numeric value
}
\description{
Multiply
}
\examples{
multiply(3,4)
[1] 12
}
