% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{This is my multiplication function}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{this is the first value to multiply}

\item{y}{this is the second value to multiply}
}
\value{
This function returns the product of x and y
}
\description{
This is my multiplication function
}
\examples{
## Start with something simple
multiply(1,1)

## Now something more difficult
multiply(49,60)

}
