% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{my_function}
\alias{my_function}
\title{my_function}
\usage{
my_function(parameter_one, parameter_two)
}
\arguments{
\item{parameter_one}{Numerical value added to parameter_two}

\item{parameter_two}{Numerical value added to param_one}
}
\value{
Sum of two parameters will be returned
}
\description{
This function will add two parameters together and return the sum
}
