% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbors cross-validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Dataframe to be inputted}

\item{cl}{Class value of training data}

\item{k_nn}{Integer represemmting number of neighbors}

\item{k_cv}{Integer representing the number of folds}
}
\value{
a list with objects class and cv_err. Class is a vector of predicted
class. CV_err is a numeric with cross-validation misclassification error
}
\description{
This function predicts output class using covariates.
}
\examples{
penguins <- mypackage::my_penguins
data <- penguins \%>\% tidyr::drop_na()
train <- data[3:6]
cl <- data$species
my_knn_cv(train, cl, 1, 5)

}
\keyword{prediction}
