% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest Neighbors}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{a matrix representing the prediction variables for the result}

\item{cl}{a numeric vector indicating the true result}

\item{k_nn}{the number of nearest neighbors to use when training the model}

\item{k_cv}{the number of folds to use when performing cross-validation}
}
\value{
a list with the predicted values of cl and the mean squared error
of the cross-validation testing
}
\description{
This function uses k nearest neighbors to train a knn model and test it using
k-fold cross validation.
}
\examples{
subset <- my_gapminder[c("year", "lifeExp", "pop")]
my_knn_cv(subset, my_gapminder["gdpPercap"], 2, 5)

}
\keyword{prediction}
