% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object}

\item{data}{Dataframe to be inputted}
}
\value{
Table with rows for each coefficient (including the (Intercept)!)
and columns for the Estimate, Std. Error, t value, and Pr(>|t|).
}
\description{
This function fits a linear model.
}
\examples{
my_lm(mpg ~ hp, data = mtcars)
}
