% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(data, formula)
}
\arguments{
\item{data}{a data frame with named columns}

\item{formula}{a formula object specifying which columns of the data frame
to use in the model}
}
\value{
table with the coefficients, standard errors, t values, and p values
of each term in the coefficient.
}
\description{
This function fits a linear model for the given data using the given formula.
}
\examples{
data(my_gapminder)
my_lm(my_gapminder, lifeExp ~ pop + gdpPercap)

}
\keyword{regression}
