% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{K-nearest neighbors cross-validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Dataframe to be inputted}
}
\value{
the number of mse errors
}
\description{
This function predicts output class using covariates.
}
