% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{random forest cross validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric indicating number of folds}
}
\value{
Numeric of mean squared error
}
\description{
This function applies random forest cross validation to penguins
data frame given number of folds
}
\examples{
k <- 5
my_rf_cv(k)

}
\keyword{cross}
\keyword{validation}
