% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alt, m)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alt}{Character string specifying the alternative hypothesis.
This only accepts "two.sided", "less", or "greater".}

\item{m}{Numeric number indicating the null hypothesis value of the mean.}
}
\value{
List of elements include test statistic, degree of freedom,
value of \code{alt}, numeric p-value.
}
\description{
This function conducts a t-test
}
\examples{
my_t.test((rnorm(10, mean = 0, sd = 1)), "greater", m = 0)

}
