% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Student's T Test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided", "greater", or "less"}

\item{mu}{the null hypothesis value of the mean}
}
\value{
List with entries for the t-statistic, the degrees of freedom,
the alternative hypothesis, and the p-value of the test
}
\description{
Performs a one-sample t-test on the given vector of data
}
\examples{
my_t.test(c(1, 2, 1.5, 3, 2.1), "two.sided", 2)
my_t.test(c(3, 1, -2, -1, 0, -1), "less", 0)

}
\keyword{inference}
