% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Vector}

\item{alternative}{Character indicating}

\item{mu}{Numeric indicating}
}
\value{
List of test statistic, degrees of freedom, alternative hypothesis,
and p-value
}
\description{
This function applies the t-test
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
my_t.test(x, "two.sided", 1)

}
\keyword{inference}
