% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\alias{ggplot_add.my_theme}
\title{Custom ggplot2 theme with optional automatic axis scaling}
\usage{
my_theme(
  top_ratio = 1,
  autoscale = TRUE,
  zero_y_min = FALSE,
  zero_x_min = NULL,
  cont_x = FALSE,
  date_x = FALSE,
  datetime_x = FALSE,
  n_breaks = 12
)

\method{ggplot_add}{my_theme}(object, plot, object_name)
}
\arguments{
\item{top_ratio}{numeric. Ratio to extend the top of the axis in autoscale mode.
Useful for adding padding above the maximum value.}

\item{autoscale}{logical. If TRUE, automatically adjusts y-axis (and x-axis, if `cont_x = TRUE`)
limits and breaks based on the data and facets.}

\item{zero_y_min}{logical. If TRUE, sets the y-axis minimum to zero.}

\item{zero_x_min}{logical or NULL. If TRUE, sets the x-axis minimum to zero.
If NULL, the default is inferred based on the data type.}

\item{cont_x}{logical. If TRUE, enforces continuous scaling and autoscaling for the x-axis.}

\item{date_x}{logical. If TRUE, enforces date scaling by years for the x-axis.}

\item{datetime_x}{logical. If TRUE, enforces datetime scaling by years for the x-axis.}

\item{n_breaks}{integer. Approximate number of breaks to generate for axes.}
}
\value{
A `my_theme` object (a list with class `"my_theme"`) that can be added to a ggplot
  object using the `+` operator.
}
\description{
This function defines a custom ggplot2 theme based on `ggthemes::theme_calc()`
with additional formatting for text, axes, panels, and legends. When used
with `ggplot_add()`, it can automatically adjust the y-axis (and optionally x-axis)
limits and breaks per panel in faceted plots, allowing for flexible autoscaling
while optionally enforcing a zero baseline.
}
