% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial sampling barplot}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of samples generated}

\item{n}{Size of each sample}

\item{p}{Probability of success in each sample}
}
\value{
A barplot for the percentage of samples vs number of successes
}
\description{
Binomial sampling barplot
}
\examples{
mybin(iter=100, n=10, p=0.7)
}
