% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinFunction.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations that will be simulated}

\item{n}{The number of draws in each iterated sample}

\item{p}{The probabilty of a "success"}
}
\value{

}
\description{
Simulates a binomial distribution
}
\examples{
mybin(10)
}
