% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBin.R
\name{myBin}
\alias{myBin}
\title{Binomial simulation function}
\usage{
myBin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{integer number of iterations}

\item{n}{integer sample size}

\item{p}{decimal probability for a success}
}
\description{
This function will create a binomial simulation. This function was provided by Dr. Stewart
for use within the Applied Statistical Methods course taught at OU.
}
\examples{
myBin(1000,1,0.5)
myBin(1000,10,0.7)

}
