% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bins.R
\name{mybin}
\alias{mybin}
\title{Makes a Normal Through Bins}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{n}{Sample size}
}
\value{
A histogram
}
\description{
Creates a distribution graph of a random sample
}
\examples{
mynorm(n=1000)
}
