% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of interations as an integer}

\item{x}{the sample as a vector}

\item{fun}{the function used for each sample}

\item{alpha}{a number between 0 and 1 for the confidence interval}

\item{cx}{cex variable for the histogram}

\item{...}{additional values for the histogram}
}
\value{
a histogram with the confidence intervals, and a named list with the
confidence interval, function, sample, and the sample after the function is
applied
}
\description{
myboot2
}
\examples{
myboot2(x=rnorm(25, mean=25, sd=10), fun="var", alpha=0.3)
}
