% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{Second bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{@description The number of iteraterations that samples will be taken}

\item{x}{@description The vector that will be sampled from}

\item{fun}{@description The sample statistic that will be estimated}

\item{alpha}{@description The parameter that decides the confidence interval percentage}

\item{cx}{@description Determines the x location of the graph}

\item{...}{@description Adds other text to be placed in.}
}
\value{

}
\description{
A function that estimates a bootstrap and a confidence interval for it.
}
\examples{
myboot2()
}
