% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{myboot2}
\alias{myboot2}
\title{Unique Random Sample Histogram}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Makes a histogram of a given sample of the bootstrap function
}
\examples{


}
