% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Creates a histogram of the distribution of samples in a sample x with replacement}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{amount of re-sampling done}

\item{x}{vector of original sample}

\item{fun}{function to be applied on each sample}

\item{alpha}{1-alpha is the confidence level}

\item{cx}{character size in text}

\item{...}{extra params}
}
\value{
list of confidence interval, function, and sample
}
\description{
Creates a histogram of the distribution of samples in a sample x with replacement
}
\examples{
myboot2(x=rnorm(25,mean=25,sd=10), fun="var", alpha=.20)
myboot2(x=rgamma(30,shape=2,scale=3), fun="mean", alpha=.20)

}
