% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myChiSim.R
\name{myChiSim}
\alias{myChiSim}
\title{Chi-Square distribution function}
\usage{
myChiSim(n1 = 10, sigma1 = 3, mean1 = 5, iter = 1000, ymax = 0.1, ...)
}
\arguments{
\item{n1}{integer representing sample size}

\item{sigma1}{integer representing standard deviation}

\item{mean1}{integer representing mean}

\item{iter}{integer representing number of iterations}

\item{ymax}{decimal or integer representing maximum y value}

\item{...}{additional arguments}
}
\description{
This function will create a histogram with annotation using the Chi-Square distribution. The histogram
will include a density plot and a theoretical curve.
}
