% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence interval.R
\name{myci}
\alias{myci}
\title{Confidence interval program}
\usage{
myci(x, alpha = 0.05)
}
\arguments{
\item{x}{@description A sample goes here}

\item{alpha}{@description The percentage for the confidence interval}
}
\value{

}
\description{
Confidence interval program
}
\examples{
myci(iris)
}
