% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{myci}
\alias{myci}
\title{Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{random single sample}
}
\value{
conf interval
}
\description{
Finds the default confidence interval for a 2 sample population
}
\examples{
x = rnorm(30, mean=10, sd=12)
}
