% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLT.R
\name{myclt}
\alias{myclt}
\title{Uniform Sample Distribution}
\usage{
myclt(n = 5, iter = 100)
}
\arguments{
\item{n}{sample size}
}
\value{
data point
}
\description{
Makes a histogram of the data
}
\examples{



}
