% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Creates a distribution of the sum of samples of a uniform distribution}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{size of each uniform distribution}

\item{iter}{amount of iterations done for each sampling}

\item{a}{lower bound for the uniform distribution}

\item{b}{upper bound for the uniform distribution}
}
\value{
distribution of the sum of uniforms
}
\description{
Creates a distribution of the sum of samples of a uniform distribution
}
\examples{
myclt(n=10,iter=10000)
myclt(n=2,iter=8000)

}
