% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Size of the sample}

\item{iter}{Number of iterations taken}

\item{lambda}{Lambda function that defines it}

\item{...}{Additional test for the graph}
}
\value{

}
\description{
Generates a randomly generated sample of size n from a poisson distribution. It takes "iter" many samples
}
\examples{
mycltp(10, iter=100)
}
