% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{number of different observations.}

\item{iter}{number of iterations.}

\item{a}{minimum values}

\item{b}{maximum value}
}
\description{
Plots a random uniform distribution to show how the distribution changes with sample size.
}
