% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{A function to create a uniform statistic and generate mean plots}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{the number of samples}

\item{iter}{the number of iterations}

\item{a}{the min for Uniform distribution}

\item{b}{the max for Uniform distribution}
}
\value{
w vector and density histogram
}
\description{
A function to create a uniform statistic and generate mean plots
}
\details{
A function from Lab 8
}
\examples{
\dontrun{mycltu(n=1,iter=10000,a=0,b=10))
}
}
