% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{my ddt}
\usage{
myddt(df = read.csv("DDT.csv"), SPECIES = "CCATFISH")
}
\arguments{
\item{df}{a data frame}

\item{SPECIES}{the name of a species}
}
\value{
a Length vs Weight plot for the given Species and a quadratic curve

a named list of the original data frame, subsetted data frame, and
a relative frequency table of River in the original data frame
}
\description{
my ddt
}
\examples{
\dontrun{myddt(ddt, SPECIES="CCATFISH")}
\dontrun{myddt(ddt, SPECIES="SMBUFFALO")}
}
