% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDDT.R
\name{myDDT}
\alias{myDDT}
\title{Project 1 myDDT function}
\usage{
myDDT(ddt, speciesType)
}
\arguments{
\item{ddt}{ddt dataframe}

\item{speciesType}{string value indiciating the fish species type}
}
\description{
This function was created for Project 1, and manipulates the data within the
DDT dataframe to provide a plot and information within the console based on
a user-input value for fish species.
}
