% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project1Func.R
\name{myddt}
\alias{myddt}
\title{Project 1 DDT Plot}
\usage{
myddt(df, species, x = "LENGTH", y = "WEIGHT", col = "RIVER")
}
\arguments{
\item{df}{df variable as the chosen data frame}

\item{species}{speices assigned to the condition statement}

\item{x}{x variable as a parameter in the data frame assigned to the x-axis variable}

\item{y}{y variable as a parameter in the data frame assigned to the y-axis variable}

\item{col}{col variable as an parameter to assign color to the scatter plot}
}
\value{
Returns the data frame, the relative frequency of a chosen parameter, the subset data frame, and a scatter plot of the data with a quadratic model.
}
\description{
Project 1 DDT Plot
}
\examples{
\dontrun{myddt(df=ddt, species="CCATFISH")}
}
