% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Plots length vs weight for a species of fish, colored by river with a quadratic curve}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{dataframe of the ddt file}

\item{SPECIES}{only values from the species will be used}
}
\value{
plot of length vs weight for species
}
\description{
Plots length vs weight for a species of fish, colored by river with a quadratic curve
}
\examples{
\dontrun{
myddt(df = ddt, SPECIES = "CCATFISH")
myddt(df = ddt, SPECIES = "SMBUFFALO")
}

}
