% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{myLasso}
\alias{myLasso}
\title{Lasso Regression}
\usage{
myLasso(X, Y, lambda_all)
}
\arguments{
\item{X->}{n x p matrix of explanatory variables.}

\item{Y:}{n dimensional response vector}

\item{lambda_all:}{Vector of regularization parameters.}
}
\description{
Finds the lasso solution path for various values of 
the regularization parameter lambda. Returns a matrix containing the lasso solution vector 
beta for each regularization parameter.
}
\examples{
 n = 50
 p = 200
 s=10
 X = matrix(rnorm(n*p), nrow=n)
 X = cbind(rep(1,n), X)
 lambda_all = (0.1:100)/10
 beta_true = matrix(rep(0, p+1), nrow = p+1)
 beta_true[1:s] = 1:s
 Y = X \%*\% beta_true + rnorm(n)
 beta_all = myLasso(X,Y,lambda_all)
}
