% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{myLM}
\alias{myLM}
\title{Linear Regression}
\usage{
myLM(X, Y)
}
\arguments{
\item{X->}{X is an n x p matrix of explanatory variables}

\item{Y->}{Y is an n dimensional vector of responses}
}
\description{
Performs the linear regression of Y on X. 
Function returns the 1 x (p + 1) vector beta_ls, the least squares solution vector.
}
\examples{
 X    <- matrix(rnorm(n * p), nrow = n)
 beta <- matrix(1:p, nrow = p)
 Y    <- X \%*\% beta + rnorm(n)
 myLM(X,Y)
}
