% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\name{myLogistic}
\alias{myLogistic}
\title{Logistic Regression}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{X}{-> is an n x p matrix of explanatory variables}

\item{Y}{-> is an n dimensional vector of binary responses}
}
\description{
Performs the logistic regression of Y on X
using iterated reweighted least squares obtained from linear regression based on QR decomposition.
Function returns the logistic regression solution vector.
}
\examples{
X    <- matrix(rnorm(n * p), nrow = n)
beta <- matrix(rep(0, p), nrow = p)
Y    <- 1 * (runif(n) < (1 / (1 + exp(X \%*\% beta))))
logistic_beta <- myLogistic(X, Y)
}
