% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The kind of distribution function to be evaluated}

\item{x}{The vector of quantities}

\item{param}{sequence from point a to point b}

\item{...}{etc}
}
\description{
Plots maximum likelihood function
}
