% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximumlikelihood.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood estimator}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function Function being estimated}

\item{x}{Vector vector of inputs}

\item{param}{Parameter Vector of tiny intervals to estimate}

\item{...}{Additional vector of intervals}
}
\value{

}
\description{
Maximum Likelihood estimator
}
\examples{
mymaxlike(mean, iris)
}
