\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation}
\description{
  This function performs maximum likelihood estimation by optimizing a likelihood function over a range of parameter values and visualizes the likelihood curve.
}
\usage{
  mymaxlik(lfun, x, param, ...)
}
\arguments{
  \item{lfun}{
    Likelihood function to be optimized.
  }
  \item{x}{
    Data used in the likelihood function.
  }
  \item{param}{
    Vector of parameter values to explore.
  }
  \item{...}{
    Additional arguments to be passed to plotting functions.
  }
}
\value{
  List containing the index, parameter value, likelihood value, and slopes.
}
