% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the logarithmic function to apply the x and parameters to}

\item{x}{vector/matrix of samples}

\item{param}{usually a sequence of numbers, with more numbers producing more refinement}

\item{...}{special argument that can take additional arguments}
}
\value{
i, parami, yi, slope, and a graphical solution for maximum likelihood
}
\description{
to use with repeated sampling from same distribution
}
\details{
A function from Lab 10
}
\examples{
\dontrun{y=c(3,3,4,3,4,5,5,4) mymaxlik(lfun=logbin, x=y, param=seq(0,1, length=1000)}
}
