% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxLikelihood.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Makes a plot of the maximum likelihood of a data set
}
\examples{


}
