% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Creates a maximum likelihood distribution}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to apply on x values}

\item{x}{values in the population}

\item{param}{parameters needed for the lfun function}

\item{...}{extra parameters}
}
\value{
list of important parameters for the max likelihood
}
\description{
Creates a maximum likelihood distribution
}
\examples{

y = c(3,3,4,3,4,5,5,4)
logbin = function(x, param) log(dbinom(x,prob=param,size=20))
mymaxlik(lfun=logbin, x=y, param=seq(0,1,length=1000), xlab=expression(pi) ,main="Binomial, n=20", cex.main=2)


}
