% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NegBin.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial Probability}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{y variable as an integer for the number of failures}

\item{r}{r variable as an integer for the number of successes}

\item{p}{p as the probability of a success}
}
\value{
Returns the negative binomial probability that there are y number of failures before the rth success.
}
\description{
Negative Binomial Probability
}
\examples{
mynbin(10,3,0.4)
}
