% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean value as a number}

\item{sigma}{standard deviation as a number}

\item{a}{a variable as a number}
}
\value{
A plot of a normal curve and the probability of X<=a
}
\description{
myncurve
}
\examples{
myncurve(mu=10, sigma=5, a=6)
}
