% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mean, sd, a, color)
}
\arguments{
\item{mean}{Mean of normal distribution function}

\item{sd}{Standard deviation from mean}

\item{a}{End point to shade under curve}

\item{color}{Color of shaded area under curve}
}
\description{
Gives a normal distribution curve and shades the area under the curve
up to a. It also calculates the probability of point a.
}
