% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{MyNCurve}
\alias{MyNCurve}
\title{myncurve}
\arguments{
\item{mu}{mu}

\item{sigma}{sigma}

\item{a}{point a}
}
\description{
gives a normal distribution curve and shades the area under the curve
up to a. It also calculates the probability of point a.
}
