% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinFunction.R
\name{myncurve}
\alias{myncurve}
\title{Fill the area under a curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the function}

\item{sigma}{Standard devaition of the function}

\item{a}{Alpha level}
}
\value{

}
\description{
Graphs a normal distribution and fills in the lower tail probability of some point a
}
\examples{
function(4,8, .05)
}
