\name{myncurve}
\alias{myncurve}
\title{Plot Normal Distribution Curve with Probability Calculation}
\description{
  This function plots a normal distribution curve with shading under the curve up to a specified point 'a'.
  It also calculates and prints the probability that a random variable from this distribution is less than or equal to 'a'.
}
\usage{
  myncurve(mu, sigma, a)
}
\arguments{
  \item{mu}{
    Mean of the normal distribution.
  }
  \item{sigma}{
    Standard deviation of the normal distribution.
  }
  \item{a}{
    Point up to which the curve is plotted and shaded.
  }
}
\value{
  None
}

