% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mu as the mean of the distribution}

\item{sigma}{sigma as the standard deviation of the distribution}

\item{a}{a as the value that the random sample is less than}
}
\value{
Returns a normal distribution curve with the area shaded between the curve
and the x-axis from -10000 to x=a. The area of P(x<=a) is also displayed on the plot.
}
\description{
Normal Curve Distribution
}
\examples{
myncurve(mu=10, sigma=2, a=8)
}
