% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{standard deviation}

\item{a}{xlim}
}
\value{
display curve
}
\description{
display curve, shade area between curve and x axis from  -inf to x=a and calc the area
}
\details{
A function from Lab 9 in MATH 4753
}
\examples{
\dontrun{myncurve(mu=2,sigma=3,a=3)
}

}
