% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Creates a normal distribution curve with x <= alpha}
\usage{
myncurve(mu, sigma, alpha)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{alpha}{point where probability is calculated, x <= alpha}
}
\value{
shaded normal distribution curve with the probability
}
\description{
Creates a normal distribution curve with x <= alpha
}
\examples{
myncurve(0,1,2)
myncurve(5,.5,4)

}
