% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: For training}
\description{
Practice building an R package to help me learn.
}
\author{
\strong{Maintainer}: Annette Dodge \email{annette.dodge@hc-sc.gc.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
