\name{mypackage-package}
\alias{mypackage-package}
\alias{mypackage}
\docType{package}
\title{\packageTitle{mypackage}}
\description{\packageDescription{mypackage}}
\section{Package Content}{\packageIndices{mypackage}}
\author{\packageAuthor{mypackage}}
\section{Maintainer}{\packageMaintainer{mypackage}}
\keyword{package}
