% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: Toy package to fiddle with a usethis workflow}
\description{
Nothing interesting.
}
\author{
\strong{Maintainer}: Felix Cheysson \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
