\name{myPackage}
\alias{myPackage}
\docType{package}
\title{Performs a series of mathematical analyses}
\description{The functions in myPackage perfrom essential mathematical analysis}
\details{
  \tabular{ll}{
  Package: \tab myPackage\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-01-08\cr
  License: \tab GPL-2\cr
  }
  Essential mathematical analysis
}
\author{
  Rob Foy
  Maintainer: Rob Foy <raf51@cam.ac.uk>
}
\keyword{ myPpackage }
