% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage-package}
\title{Homework 6}
\description{
This package is a solution for course R Advanced Programming course the homework 6.
Package contains 3 functions: mode, simplify2array, ass.
}

