\name{mypackage-package}
\alias{mypackage-package}
\alias{mypackage}
\docType{package}
\title{
\packageTitle{mypackage}
}
\description{
\packageDescription{mypackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mypackage}
\packageIndices{mypackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mypackage}

Maintainer: \packageMaintainer{mypackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
