% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage-package}
\title{mypackage}
\description{
The package contains two Rcpp functions - mode2 and simplify2array2.
Function mode2() calculates a mode for a given integer vector.
Function simplify2array2() converts a list containing numeric vectors
into a matrix.
}

