% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage-package.R
\docType{package}
\name{myPackage-package}
\alias{myPackage}
\alias{myPackage-package}
\title{myPackage: A Package of Sorts}
\description{
There are many packages like this, but this package is my package.
}
\author{
\strong{Maintainer}: Samuel Lee \email{samleenz@me.com} (0000-0002-6805-4051)

}
\keyword{internal}
