% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyPackage-package.R
\docType{package}
\name{MyPackage-package}
\alias{MyPackage}
\alias{MyPackage-package}
\title{MyPackage: Vespa Invasion Vizualization}
\description{
MyPackage produces a map or bar chart per year of known observations of Asian hornets (Vespa velutina) in Flanders. They typically want to filter this information on a year range. Create R code that allows: 1. To define a year range (e.g. 2019-2021). 2. To retrieve occurrences from the Global Biodiversity Information Facility (GBIF) for that year range of Vespa velutina in Flanders. 3. To present the result as a map of observations. 4. To present the result as a bar chart of observations per year. Your code should be documented and tested. Your code should be maintained on GitHub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sangovae/MyPackage}
  \item Report bugs at \url{https://github.com/sangovae/MyPackage/issues}
}

}
\keyword{internal}
