% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyPackage-package.R
\docType{package}
\name{MyPackage-package}
\alias{MyPackage}
\alias{MyPackage-package}
\title{MyPackage: Vespa Invasion Visualization}
\description{
The goal of this package is to produce a map or bar chart per year of known observations of Asian hornets *Vespa velutina* in Flanders, and you are able to filter this information on a year range.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sangovae/MyPackage}
  \item Report bugs at \url{https://github.com/sangovae/MyPackage/issues}
}

}
\keyword{internal}
