% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: Explore Your R Libraries}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sungoose/mypackage}
  \item Report bugs at \url{https://github.com/sungoose/mypackage/issues}
}

}
\author{
\strong{Maintainer}: Osung Seo \email{seoosung@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
