% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageCode.R
\docType{package}
\name{myPackage-package}
\alias{myPackage}
\alias{myPackage-package}
\title{Package to do Statistical Computing}
\description{
Functions available: Linear Regression, Logistic Regression, Sweep Operation,
QR Decomposition, Lasso Rigression and Ridge Regression.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
