\name{euclidean(x,y)}
\alias{lab_3}
\docType{package}
\title{
Euclidean(x, y)
}
\description{
Euclidean algortihm with two numbers
}
\arguments{
x as numeric \cr
y as numeric \cr
}
\details{
The function should give us the greatest common divisor between x and y
 }
\tabular{ll}{
Package: \tab mypackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-07\cr
License: \tab free \cr
}
~~ Euclidean function ~~
}
\author{
Violeta & Yixuan

Maintainers: c(Vileta <vioga273@student.liu.se>,Yixuan <yixxu916@student.liu.se>)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
~~ documentation directory ~~
\keyword{ package }
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(10, 100)
euclidean(44, 2)
}
