\name{mypackage-LAb_3}
\alias{lab_3}
\docType{package}
\title{
Euclidean function
}
\description{
Euclidean algortihm with two numbers
}
\details{
\tabular{ll}{
Package: \tab mypackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-07\cr
License: \tab free \cr
}
~~ Euclidean function ~~
}
\author{
Violeta & Yixuan

Maintainer: Vileta <violeta26_11_92@hotmail.com>
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
~~ documentation directory ~~
\keyword{ package }
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(10, 100)
}
