\name{MyPackage-package}
\alias{MyPackage-package}
\alias{MyPackage}
\docType{package}
\title{
\packageTitle{MyPackage}
}
\description{
\packageDescription{MyPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyPackage}
\packageIndices{MyPackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MyPackage}

Maintainer: \packageMaintainer{MyPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
