\name{MyPackage-package}
\alias{MyPackage-package}
\alias{MyPackage}
\docType{package}
\title{
\packageTitle{MyPackage}
}
\description{
This package is built by copying the sample code for package.skeleton() and then editing the help files
}
\details{

\tabular{ll}{
Package: \tab MyPackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-28\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}
}
\author{
Liew Kit Loong{MyPackage}

Maintainer: Liew Kit Loong(xia0t99@gmail.com){MyPackage}
}
\references{
 Package source is in the help file for package.skeleton}
\keyword{ package }

\examples{
test <- square(x^2)
test2 <- cube(y^2)
}
