% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage-package.R
\docType{package}
\name{mypackage-package}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage: Explore Your R Libraries}
\description{
Provides functions for learning about your R libraries, and the packages you have installed.
}
\author{
\strong{Maintainer}: Tianyu Zheng \email{zhengtianyunc@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
