% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage-package.R
\name{myPackage}
\alias{myPackage}
\title{myPackage: Summarise Installed R Packages by Library}
\description{
This package provides a single function, \code{lib_summary()}, that returns the number of installed packages in each R library path used by your environment.
}
\details{
"_PACKAGE"
}
