% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown_pdf_document.R
\name{mypackage}
\alias{mypackage}
\title{Customized RMarkdown Document}
\usage{
mypackage(...)
}
\arguments{
\item{...}{Passed on to \code{\link[=indiedown_pdf_document_with_asset]{indiedown_pdf_document_with_asset()}}.}
}
\description{
\code{mypackage()} is the main function of the mypackage package. Use as
\code{output: mypackage::mypackage} in the RMarkdown YAML header.
}
