% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyPackage.R
\docType{package}
\name{MyPackage}
\alias{MyPackage}
\title{For Applied Statistical Methods}
\description{
This package was created for use within the Applied Statistical
Methods course that I took at OU.
}
\author{
Joseph Grimes \email{joseph.grimes@ou.edu}
}
