% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackage.R
\docType{package}
\name{myPackage}
\alias{myPackage}
\title{Custom Package}
\description{
Contains various functions from previous labs for newer labs where using this package is necesssary.
}
\author{
Thuan Tran
}
