% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{myttest}
\alias{myttest}
\title{Constructor Function}
\usage{
myttest(x, y, alpha)
}
\arguments{
\item{x}{The x dataset}

\item{y}{The y dataset}
}
\value{
A list
}
\description{
Takes in Data and produces a list
}
\examples{
myttest(x = x, y = y, alpha = 0.05 )
}
