% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_outliers.R
\name{outlier_hcluster}
\alias{outlier_hcluster}
\title{Sample filter based on hierarchical clustering}
\usage{
outlier_hcluster(data, traitNum = NULL, scale = FALSE, cutheight = 50000,
  minSize = 10)
}
\arguments{
\item{data}{Data matrix}

\item{traitNum}{Sample information which must be numeric for cluster plot Defaults to NULL}

\item{scale}{Whether scale is used for the data Defaults to FALSE}

\item{cutheight}{The threshold for outliers Defaults to 50000}

\item{minSize}{The minimum cluster size for outliers Defaults to 10}
}
\value{
outliers
}
\description{
This function remove outliers based on hierarchical clustering.
}
\examples{
outlier_hcluster(data, traitNum, cutheight=50000)
}
\author{
WKL
}
\keyword{cluster}
