% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Function.R
\name{plot.rttest}
\alias{plot.rttest}
\title{Plot Function}
\usage{
\method{plot}{rttest}(z)
}
\arguments{
\item{x}{A variable}

\item{...}{A table variable}
}
\value{
A boxplot
}
\description{
Plots Results of Data Using ggplot2
}
\examples{
plot(obj)
}
