% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{plot.Rttest}
\usage{
\method{plot}{Rttest}(w, ...)
}
\arguments{
\item{w}{output of myttest function}

\item{...}{additional ggplot parameters if desired}
}
\value{
ggplot2 boxplot of the dataframe, differs depending on t.test used
}
\description{
plot.Rttest
}
\examples{
set.seed(20); plot(myttest(rnorm(30,5,2), rnorm(30,3,2,), .05, FALSE))
}
