% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abline.R
\name{plot_model}
\alias{plot_model}
\title{Plotting diamonds dataset}
\usage{
plot_model(model, data)
}
\value{
A plot
}
\description{
Plotting diamonds dataset
}
\details{
This function takes as parameter the data and the linear model we fitted the data at the other function
}
\examples{
plot_model <- function(model, data) {
 ggplot(data) +
  geom_point(aes(x = carat, y = price)) +
   geom_abline(intercept = model[1], slope = model[2])+
   ggtitle("The relationship of the price and the carats of the diamond")
}
}
