% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_stats}
\alias{plot_stats}
\title{Plotting diamonds dataset}
\usage{
plot_stats(data)
}
\arguments{
\item{data}{simply diamonds free dataset}
}
\value{
A plot
}
\description{
Plotting diamonds dataset
}
\details{
This function takes some data and uses them to fit the model
}
\examples{
# basic function
fit_model <- function(data) {
lm(price ~ carat, data) \%>\%
coefficients()
}
}
