% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{print.Rttest
defines the print method for the Rttest class}
\usage{
\method{print}{Rttest}(list)
}
\arguments{
\item{list}{#the output of the myttest function}
}
\value{
returns a list of length 6: kable table of x and y
a confidence interval of the t.test, the type of test used,
p.value of the t.test, a rejection statement,
and the alpha value used.
}
\description{
print.Rttest
defines the print method for the Rttest class
}
\examples{
set.seed(20); print(myttest(rnorm(30,5,2), rnorm(30,3,2), .05, FALSE))
}
