% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resampling}
\usage{
resample(x, ...)
}
\description{
\code{resample} takes a random sample from a given vector.
}
\details{
This function is an alternative to \code{sample}.  Unlike the latter, it
always interprets the first argument as the set of elements to sample from,
even when it is of length 1.  This consistency makes it more suitable for
programmatic use.
}

