% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightleftorder.R
\name{rightleftorder}
\alias{rightleftorder}
\title{rightleftorder}
\usage{
rightleftorder(n, i)
}
\arguments{
\item{i}{An integer between 1 and \code{length(x)}.}

\item{x}{A vector}
}
\description{
Reorder a vector \code{x}, starting at a given index \code{i} and proceeding
rightward then leftward.
}
