% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate standard error}
\usage{
se(value)
}
\arguments{
\item{value}{numeric vector}
}
\value{
number
}
\description{
Returns the standard error
from standard deviation and sample size
}
\examples{
se(rnorm(20))
}
