% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{soma}
\alias{soma}
\title{As quatro operacoes desagregadas}
\usage{
soma(a, b)
}
\arguments{
\item{a}{primeiro numero}

\item{b}{segundo numero}
}
\value{
Retorna o somatorio de a e b
}
\description{
O somatorio de a e b
}
\details{
Calculo aritmetico simples
}
\keyword{a,}
\keyword{b}
\keyword{soma,}
