% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square.R
\name{square}
\alias{square}
\title{Square of a number}
\usage{
square(x)
}
\arguments{
\item{A}{numeric value to be squared}
}
\value{
The square of the input
}
\description{
Takes in any numeric value and squares it
}
