\name{square}
\alias{square}

\title{ Square Function}

\description{ The function adds two numbers}

\usage{
square(x^2)

}

\arguments{
  \item{x}{ The argument for square}


}
\details{ No more details required. It adds two numbers}

\value{ Returns a number that is the sum }

\references{ package.Skeleton code}

\author{ Liew Kit Loong} 

\note{ There are no further notes}

\examples{
  x <- 2
  test <- square(x^2)
  print(test)
}

\keyword{ arith }
