\name{square}
\alias{square}

\title{ Square and Cube Function}

\description{ The function square and cube a number}

\usage{
square(x^2)
cube(y^2)

}

\arguments{
  \item{x}{ The argument for square}
   \item{y}{ The argument for cube}



}
\details{ Square a number and cube a number}

\value{ It return a value multiple by 2 or a value multiple by 3 }

\references{ package.Skeleton code}

\author{ Liew Kit Loong} 

\note{ There are no further notes}

\examples{
  x <- 2
  y <- 2
  test <- square(x^2)
  print(test)
  test2 <- cube(y^2)
  print(test2)
}

\keyword{ arith }
