% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error from a distribution}
\usage{
standard_error(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
Returns the standard error of the mean (SEM) from a numeric vector
}
