% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate standard error of a sample}
\usage{
standard_error(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
Calculate standard error of a sample
}
\examples{
standard_error(rnorm(50))
}
