% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error for a sample.
Returns the standard error of sample, SE(x) of a numeric vector}
\usage{
standard_error(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
Calculate the standard error for a sample.
Returns the standard error of sample, SE(x) of a numeric vector
}
\examples{
standard_error(rnorm(50))
}
