% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error}
\usage{
standard_error(sd, n)
}
\arguments{
\item{sd, n}{numeric vector, numeric vector}
}
\value{
number
}
\description{
returns the standard error
}
\details{
from standard deviation, sample size of a numeric vector
}
\examples{
sum_squares_function(rnorm(50))
}
