% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculate the sums if squares of a sample}
\usage{
sum_squares(value)
}
\arguments{
\item{value}{numeric vector}
}
\value{
number
}
\description{
Returns the sum of the squared deviations
from mean, SS(x) of a numeric vector
}
\examples{
sum_squares(rnorm(50))
}
