% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculate the sums of squares for a sample}
\usage{
sum_squares(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
number
}
\description{
returns the sum of the squared deviations
from mean, SS(x) of a numeric vector
}
\examples{
sum_squares(rnorm(50))
}
