% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_plot.R
\name{temp_plot}
\alias{temp_plot}
\title{Save plot to quick directory}
\usage{
temp_plot(plot, path = "~/temp-plot.png", width = 6, height = 4, dpi = 300)
}
\arguments{
\item{plot}{is a ggplot object}

\item{path}{is where to save the plot}

\item{width}{width of plot}

\item{height}{height of plot}

\item{dpi}{pixel density}
}
\value{
save plot and print path where plot saved
}
\description{
Save plot to quick directory
}
