% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vespa_data.R
\name{vespa_data}
\alias{vespa_data}
\title{Retrieve occurrences from GBIF of Vespa velutina}
\usage{
vespa_data(year1, year2 = NULL)
}
\arguments{
\item{year1}{The 4 digit year. A single year or lower number of the year range.}

\item{year2}{Optional. 4 digit year, higher number of the year range.}
}
\value{
A \code{\link[=tibble]{tibble()}} with occurrences from GBIF for that year range of Vespa velutina in Flanders
}
\description{
code{vespa_data} retrieves occurrences from the Global Biodiversity Information Facility (GBIF) for
a single year or year range of Vespa velutina in Flanders.
}
\examples{
vespa_data(2019,2021)
}
